Add a Row with a Name

add row to data frame and give a row name at the same time

addRowWithName(x, y, row.name)

Arguments

x
data frame to which row is to be appended

y
data frame containing the row to be appended (exacly one row expected)

row.name
name of row to be given in result data frame

Value

x with row of y (named row.name) appended to it