Copy Directory Structure

Copy the full directory structure from a source directory to a target directory

copyDirectoryStructure(sourcedir, targetdir, excludePattern = "^$",
  recursive = TRUE, dbg = TRUE)

Arguments

sourcedir
path to the source directory

targetdir
path to the target directory

excludePattern
pattern matching directory names to be excluded from the copying process

recursive
if TRUE (default) the full tree of directories and subdirectories is copied, otherwise only the top-level directories

dbg
if TRUE (default) debug messages are shown

Value

This function invisibly returns a vector of character containing the full paths of the directories that were created.