List Elements Recursively up to Depth

List Elements Recursively up to Depth

Usage

listToDepth(
  path,
  max_depth = 0L,
  full_info = FALSE,
  FUN = listFiles,
  ...,
  depth = 0,
  prob_mutate = 0,
  template = NULL
)

Arguments

path

path to the element at which to start listing

max_depth

maximal depth level of which to list elements. A value of 0 means non-recursive listing, a value of NA represents fully recursive listing.

full_info

return only path and isdir information or the full information provided by FUN(full_info = TRUE)?

FUN

function called to get the listing of the element given in path. The function must accept a path as its first argument and it must define the argument full_info second. It may accept further arguments. It must always return a data frame. For full_info = FALSE the data frame must have columns file and isdir (is the "file" a directory?). For full_info = TRUE the function may return further columns. The function must provide an empty data frame with the expected columns when being called with character() as the first argument. The function is expected to set the attribute "failed" to the given path in case that the path could not be accessed (e.g. because of a broken internet connection if the listing is done remotely). See kwb.utils:::listFiles for an example implementation that somehow simulates the behaviour of the dir function. See kwb.dwd::list_url() for a more advanced usage of this function in order to recursively list the files on an FTP server (FTP = file transfer protocol).

...

further arguments passed to FUN

depth

start depth of recursion if max_depth > 0. This argument is for internal use and not intended to be set by the user!

prob_mutate

probability to alter the path so that it becomes useless. This is zero by default. Set the value only if you want to test how the function behaves if the listing of a path fails.

template

empty data frame (zero rows) and with columns that are identical to the columns in the data frame returned by FUN. If not provided the function will call FUN once with a zero-length path argument expecting to retrieve the template that is expected here.

Value

data frame containing at least the columns file and

isdir. If full_info = TRUE the result data frame may contain further columns, as provided by the function given in FUN for

full_info = TRUE.

Examples

# Example list function provided in this package (file listing)
FUN <- kwb.utils:::listFiles

# The list function must provide empty records when no path is given. The 
# returned data frame must provide the columns "file" and "isdir" ...
FUN()
#> listing .
#>                                        file isdir
#> 1                               DIN.A4.html FALSE
#> 2                              Rplot001.png FALSE
#> 3                              Rplot002.png FALSE
#> 4                             addClass.html FALSE
#> 5                       addRowWithName.html FALSE
#> 6                   addSuffixToColumns.html FALSE
#> 7                          allAreEqual.html FALSE
#> 8                      allAreIdentical.html FALSE
#> 9                          almostEqual.html FALSE
#> 10                        appendSuffix.html FALSE
#> 11                            arglist-1.png FALSE
#> 12                            arglist-2.png FALSE
#> 13                             arglist.html FALSE
#> 14                             argsCsv.html FALSE
#> 15                    arrayToDataFrame.html FALSE
#> 16                        asColumnList.html FALSE
#> 17                 asNoFactorDataFrame.html FALSE
#> 18                           asRowList.html FALSE
#> 19                    assertFinalSlash.html FALSE
#> 20                assertRowsAndColumns.html FALSE
#> 21                           assignAll.html FALSE
#> 22              assignArgumentDefaults.html FALSE
#> 23                      assignGlobally.html FALSE
#> 24                       assignObjects.html FALSE
#> 25                assignPackageObjects.html FALSE
#> 26                     atLeastOneRowIn.html FALSE
#> 27                           backspace.html FALSE
#> 28                     breakInSequence.html FALSE
#> 29                           callWith-1.png FALSE
#> 30                           callWith-2.png FALSE
#> 31                            callWith.html FALSE
#> 32            callWithStringsAsFactors.html FALSE
#> 33                           catAndRun.html FALSE
#> 34                          catChanges.html FALSE
#> 35                        catChangesIf.html FALSE
#> 36                               catIf.html FALSE
#> 37                            catLines.html FALSE
#> 38                        catNewLineIf.html FALSE
#> 39                        catchWarning.html FALSE
#> 40              checkForMissingColumns.html FALSE
#> 41                        clearConsole.html FALSE
#> 42                          clipMatrix.html FALSE
#> 43                         cmdLinePath.html FALSE
#> 44                           colMaxima.html FALSE
#> 45                           colMinima.html FALSE
#> 46                        colNaNumbers.html FALSE
#> 47             colStatisticOneFunction.html FALSE
#> 48                       colStatistics.html FALSE
#> 49                           collapsed.html FALSE
#> 50                    columnDescriptor.html FALSE
#> 51                        columnToDate.html FALSE
#> 52                columnwisePercentage.html FALSE
#> 53                combineAlternatingly.html FALSE
#> 54                      commaCollapsed.html FALSE
#> 55                         commonNames.html FALSE
#> 56                   compareDataFrames.html FALSE
#> 57                         compareSets.html FALSE
#> 58                   containsNulString.html FALSE
#> 59                      convertCsvFile.html FALSE
#> 60                      copyAttributes.html FALSE
#> 61              copyDirectoryStructure.html FALSE
#> 62                    copyListElements.html FALSE
#> 63                     countNaInColumn.html FALSE
#> 64                          countOrSum.html FALSE
#> 65                      createAccessor.html FALSE
#> 66              createDirAndReturnPath.html FALSE
#> 67                   createDirectories.html FALSE
#> 68                     createDirectory.html FALSE
#> 69         createFunctionAssignObjects.html FALSE
#> 70           createFunctionExtdataFile.html FALSE
#> 71                       createIdAlong.html FALSE
#> 72                        createMatrix.html FALSE
#> 73                  createPasswordFile.html FALSE
#> 74                       createStorage.html FALSE
#> 75                  csvTextToDataFrame.html FALSE
#> 76                              decode.html FALSE
#> 77                           defaultIf.html FALSE
#> 78                         defaultIfNA.html FALSE
#> 79                       defaultIfNULL.html FALSE
#> 80                       defaultIfZero.html FALSE
#> 81                       defaultLevels.html FALSE
#> 82        defaultWindowsProgramFolders.html FALSE
#> 83                             desktop.html FALSE
#> 84                            diffrows.html FALSE
#> 85                       directoryName.html FALSE
#> 86                          dot-OStype.html FALSE
#> 87                             dot-log.html FALSE
#> 88                         dot-logline.html FALSE
#> 89                           dot-logok.html FALSE
#> 90                        dot-logstart.html FALSE
#> 91                             dropDim.html FALSE
#> 92              dropUnusedFactorLevels.html FALSE
#> 93                              encode.html FALSE
#> 94                       enlargeVector.html FALSE
#> 95                         excludeNULL.html FALSE
#> 96                          expandGrid.html FALSE
#> 97                         extdataFile.html FALSE
#> 98                        extendLimits.html FALSE
#> 99                    extractRowRanges.html FALSE
#> 100                   extractSubstring.html FALSE
#> 101                      fileExtension.html FALSE
#> 102                        findChanges.html FALSE
#> 103              findPartialDuplicates.html FALSE
#> 104                   finishAndShowPdf.html FALSE
#> 105                 finishAndShowPdfIf.html FALSE
#> 106                       firstElement.html FALSE
#> 107                   firstPosixColumn.html FALSE
#> 108                      frenchToAscii.html FALSE
#> 109                     frequencyTable.html FALSE
#> 110                        fullWinPath.html FALSE
#> 111                        fullySorted.html FALSE
#> 112                    generateKeyFile.html FALSE
#> 113                       getAttribute.html FALSE
#> 114       getByPositiveOrNegativeIndex.html FALSE
#> 115                  getElementLengths.html FALSE
#> 116                     getEvenNumbers.html FALSE
#> 117          getFunctionValueOrDefault.html FALSE
#> 118                        getGlobally.html FALSE
#> 119                getKeywordPositions.html FALSE
#> 120                        getListNode.html FALSE
#> 121      getNamesOfObjectsInRDataFiles.html FALSE
#> 122             getObjectFromRDataFile.html FALSE
#> 123                      getOddNumbers.html FALSE
#> 124                        getPassword.html FALSE
#> 125 getPathsAndValuesFromRecursiveList.html FALSE
#> 126                        getTagNames.html FALSE
#> 127                        get_homedir.html FALSE
#> 128                   guessSeparator.1.html FALSE
#> 129                   guessSeparator.2.html FALSE
#> 130                     guessSeparator.html FALSE
#> 131                      hasFinalSlash.html FALSE
#> 132                      hasZeroLength.html FALSE
#> 133                           headtail.html FALSE
#> 134                   hsAddMissingCols.html FALSE
#> 135                         hsChrToNum.html FALSE
#> 136                       hsCountInStr.html FALSE
#> 137                     hsDelEmptyCols.html FALSE
#> 138                 hsMatrixToListForm.html FALSE
#> 139                      hsMovingMean-1.png FALSE
#> 140                       hsMovingMean.html FALSE
#> 141              hsOpenWindowsExplorer.html FALSE
#> 142                          hsPrepPdf.html FALSE
#> 143                         hsQuoteChr.html FALSE
#> 144                    hsRenameColumns.html FALSE
#> 145                          hsResolve.html FALSE
#> 146                hsRestoreAttributes.html FALSE
#> 147                         hsSafeName.html FALSE
#> 148                            hsShell.html FALSE
#> 149                          hsShowPdf.html FALSE
#> 150                     hsStringToDate.html FALSE
#> 151                   hsStringToDouble.html FALSE
#> 152                   hsSubstSpecChars.html FALSE
#> 153                           hsSystem.html FALSE
#> 154                             hsTrim.html FALSE
#> 155                       hsValidValue.html FALSE
#> 156                            inRange.html FALSE
#> 157                             indent.html FALSE
#> 158                              index.html FALSE
#> 159                      insertColumns.html FALSE
#> 160                 intToNumeralSystem.html FALSE
#> 161                         is.unnamed.html FALSE
#> 162                            isASCII.html FALSE
#> 163                   isDotOrDoubleDot.html FALSE
#> 164                       isEvenNumber.html FALSE
#> 165                           isLoaded.html FALSE
#> 166                   isNaInAllColumns.html FALSE
#> 167                      isNaInAllRows.html FALSE
#> 168                        isNaOrEmpty.html FALSE
#> 169                      isNetworkPath.html FALSE
#> 170                      isNullOrEmpty.html FALSE
#> 171                        isOddNumber.html FALSE
#> 172                         isTryError.html FALSE
#> 173                        lastElement.html FALSE
#> 174                               left.html FALSE
#> 175                leftSubstringEquals.html FALSE
#> 176                       limitToRange.html FALSE
#> 177                  linearCombination.html FALSE
#> 178                        listObjects.html FALSE
FUN(full_info = TRUE)
#> listing .
#>                                        file isdir   size mode
#> 1                               DIN.A4.html FALSE   8232  644
#> 2                              Rplot001.png FALSE   1011  644
#> 3                              Rplot002.png FALSE   5908  644
#> 4                             addClass.html FALSE   9409  644
#> 5                       addRowWithName.html FALSE   9093  644
#> 6                   addSuffixToColumns.html FALSE  11854  644
#> 7                          allAreEqual.html FALSE   9048  644
#> 8                      allAreIdentical.html FALSE   8893  644
#> 9                          almostEqual.html FALSE   9064  644
#> 10                        appendSuffix.html FALSE  10746  644
#> 11                            arglist-1.png FALSE  18498  644
#> 12                            arglist-2.png FALSE  18423  644
#> 13                             arglist.html FALSE  12977  644
#> 14                             argsCsv.html FALSE   8929  644
#> 15                    arrayToDataFrame.html FALSE  13569  644
#> 16                        asColumnList.html FALSE  10960  644
#> 17                 asNoFactorDataFrame.html FALSE  11294  644
#> 18                           asRowList.html FALSE  10844  644
#> 19                    assertFinalSlash.html FALSE  11890  644
#> 20                assertRowsAndColumns.html FALSE  12432  644
#> 21                           assignAll.html FALSE   8884  644
#> 22              assignArgumentDefaults.html FALSE  10313  644
#> 23                      assignGlobally.html FALSE   8572  644
#> 24                       assignObjects.html FALSE   9643  644
#> 25                assignPackageObjects.html FALSE   9141  644
#> 26                     atLeastOneRowIn.html FALSE   8646  644
#> 27                           backspace.html FALSE  10414  644
#> 28                     breakInSequence.html FALSE   9025  644
#> 29                           callWith-1.png FALSE  18498  644
#> 30                           callWith-2.png FALSE  18423  644
#> 31                            callWith.html FALSE  12077  644
#> 32            callWithStringsAsFactors.html FALSE  14551  644
#> 33                           catAndRun.html FALSE  12821  644
#> 34                          catChanges.html FALSE   9499  644
#> 35                        catChangesIf.html FALSE   8946  644
#> 36                               catIf.html FALSE   8605  644
#> 37                            catLines.html FALSE   8708  644
#> 38                        catNewLineIf.html FALSE   8957  644
#> 39                        catchWarning.html FALSE  11160  644
#> 40              checkForMissingColumns.html FALSE   9851  644
#> 41                        clearConsole.html FALSE   8164  644
#> 42                          clipMatrix.html FALSE   8898  644
#> 43                         cmdLinePath.html FALSE   8681  644
#> 44                           colMaxima.html FALSE   8695  644
#> 45                           colMinima.html FALSE   8695  644
#> 46                        colNaNumbers.html FALSE   8631  644
#> 47             colStatisticOneFunction.html FALSE   9122  644
#> 48                       colStatistics.html FALSE   9738  644
#> 49                           collapsed.html FALSE   8888  644
#> 50                    columnDescriptor.html FALSE   8799  644
#> 51                        columnToDate.html FALSE  11334  644
#> 52                columnwisePercentage.html FALSE  14742  644
#> 53                combineAlternatingly.html FALSE  12009  644
#> 54                      commaCollapsed.html FALSE   8511  644
#> 55                         commonNames.html FALSE  10056  644
#> 56                   compareDataFrames.html FALSE  13471  644
#> 57                         compareSets.html FALSE  11780  644
#> 58                   containsNulString.html FALSE   8834  644
#> 59                      convertCsvFile.html FALSE  22511  644
#> 60                      copyAttributes.html FALSE  10537  644
#> 61              copyDirectoryStructure.html FALSE   9732  644
#> 62                    copyListElements.html FALSE  11873  644
#> 63                     countNaInColumn.html FALSE   8886  644
#> 64                          countOrSum.html FALSE  13815  644
#> 65                      createAccessor.html FALSE  10691  644
#> 66              createDirAndReturnPath.html FALSE   9366  644
#> 67                   createDirectories.html FALSE   8857  644
#> 68                     createDirectory.html FALSE   9259  644
#> 69         createFunctionAssignObjects.html FALSE  10050  644
#> 70           createFunctionExtdataFile.html FALSE  10070  644
#> 71                       createIdAlong.html FALSE  11871  644
#> 72                        createMatrix.html FALSE  13883  644
#> 73                  createPasswordFile.html FALSE   9698  644
#> 74                       createStorage.html FALSE  14833  644
#> 75                  csvTextToDataFrame.html FALSE   8744  644
#> 76                              decode.html FALSE   8808  644
#> 77                           defaultIf.html FALSE   9235  644
#> 78                         defaultIfNA.html FALSE  10413  644
#> 79                       defaultIfNULL.html FALSE  10436  644
#> 80                       defaultIfZero.html FALSE  11728  644
#> 81                       defaultLevels.html FALSE  10527  644
#> 82        defaultWindowsProgramFolders.html FALSE   8308  644
#> 83                             desktop.html FALSE   8873  644
#> 84                            diffrows.html FALSE  10810  644
#> 85                       directoryName.html FALSE  10130  644
#> 86                          dot-OStype.html FALSE   8557  644
#> 87                             dot-log.html FALSE   8606  644
#> 88                         dot-logline.html FALSE   8760  644
#> 89                           dot-logok.html FALSE   8696  644
#> 90                        dot-logstart.html FALSE   8891  644
#> 91                             dropDim.html FALSE  13927  644
#> 92              dropUnusedFactorLevels.html FALSE  13875  644
#> 93                              encode.html FALSE  11130  644
#> 94                       enlargeVector.html FALSE  11472  644
#> 95                         excludeNULL.html FALSE  10974  644
#> 96                          expandGrid.html FALSE  14651  644
#> 97                         extdataFile.html FALSE  10236  644
#> 98                        extendLimits.html FALSE   9216  644
#> 99                    extractRowRanges.html FALSE  19593  644
#> 100                   extractSubstring.html FALSE  15271  644
#> 101                      fileExtension.html FALSE  10011  644
#> 102                        findChanges.html FALSE  13199  644
#> 103              findPartialDuplicates.html FALSE  13280  644
#> 104                   finishAndShowPdf.html FALSE   9360  644
#> 105                 finishAndShowPdfIf.html FALSE   8969  644
#> 106                       firstElement.html FALSE   8696  644
#> 107                   firstPosixColumn.html FALSE   8591  644
#> 108                      frenchToAscii.html FALSE   8613  644
#> 109                     frequencyTable.html FALSE  15794  644
#> 110                        fullWinPath.html FALSE  10567  644
#> 111                        fullySorted.html FALSE  14484  644
#> 112                    generateKeyFile.html FALSE   8913  644
#> 113                       getAttribute.html FALSE  13266  644
#> 114       getByPositiveOrNegativeIndex.html FALSE   9130  644
#> 115                  getElementLengths.html FALSE  10027  644
#> 116                     getEvenNumbers.html FALSE   8626  644
#> 117          getFunctionValueOrDefault.html FALSE   9160  644
#> 118                        getGlobally.html FALSE   8965  644
#> 119                getKeywordPositions.html FALSE   9383  644
#> 120                        getListNode.html FALSE  12703  644
#> 121      getNamesOfObjectsInRDataFiles.html FALSE   8761  644
#> 122             getObjectFromRDataFile.html FALSE   9316  644
#> 123                      getOddNumbers.html FALSE   8616  644
#> 124                        getPassword.html FALSE   9922  644
#> 125 getPathsAndValuesFromRecursiveList.html FALSE  12516  644
#> 126                        getTagNames.html FALSE   9615  644
#> 127                        get_homedir.html FALSE   9029  644
#> 128                   guessSeparator.1.html FALSE   8910  644
#> 129                   guessSeparator.2.html FALSE   8784  644
#> 130                     guessSeparator.html FALSE   9143  644
#> 131                      hasFinalSlash.html FALSE   9641  644
#> 132                      hasZeroLength.html FALSE  11225  644
#> 133                           headtail.html FALSE  15070  644
#> 134                   hsAddMissingCols.html FALSE   9573  644
#> 135                         hsChrToNum.html FALSE   9541  644
#> 136                       hsCountInStr.html FALSE   8936  644
#> 137                     hsDelEmptyCols.html FALSE  10100  644
#> 138                 hsMatrixToListForm.html FALSE  10556  644
#> 139                      hsMovingMean-1.png FALSE 113977  644
#> 140                       hsMovingMean.html FALSE  12468  644
#> 141              hsOpenWindowsExplorer.html FALSE   9143  644
#> 142                          hsPrepPdf.html FALSE   9904  644
#> 143                         hsQuoteChr.html FALSE   9211  644
#> 144                    hsRenameColumns.html FALSE   8891  644
#> 145                          hsResolve.html FALSE   9445  644
#> 146                hsRestoreAttributes.html FALSE   8786  644
#> 147                         hsSafeName.html FALSE  11062  644
#> 148                            hsShell.html FALSE   8623  644
#> 149                          hsShowPdf.html FALSE  12136  644
#> 150                     hsStringToDate.html FALSE   9239  644
#> 151                   hsStringToDouble.html FALSE   9037  644
#> 152                   hsSubstSpecChars.html FALSE   9120  644
#> 153                           hsSystem.html FALSE   8635  644
#> 154                             hsTrim.html FALSE   9313  644
#> 155                       hsValidValue.html FALSE   9436  644
#> 156                            inRange.html FALSE   9201  644
#> 157                             indent.html FALSE   8863  644
#> 158                              index.html FALSE  65439  644
#> 159                      insertColumns.html FALSE  16372  644
#> 160                 intToNumeralSystem.html FALSE  13978  644
#> 161                         is.unnamed.html FALSE  11299  644
#> 162                            isASCII.html FALSE  10179  644
#> 163                   isDotOrDoubleDot.html FALSE   9617  644
#> 164                       isEvenNumber.html FALSE   8492  644
#> 165                           isLoaded.html FALSE  10834  644
#> 166                   isNaInAllColumns.html FALSE   8836  644
#> 167                      isNaInAllRows.html FALSE   8815  644
#> 168                        isNaOrEmpty.html FALSE   8935  644
#> 169                      isNetworkPath.html FALSE   9132  644
#> 170                      isNullOrEmpty.html FALSE   8751  644
#> 171                        isOddNumber.html FALSE   8482  644
#> 172                         isTryError.html FALSE   9787  644
#> 173                        lastElement.html FALSE   8844  644
#> 174                               left.html FALSE   9292  644
#> 175                leftSubstringEquals.html FALSE   9890  644
#> 176                       limitToRange.html FALSE   9715  644
#> 177                  linearCombination.html FALSE  12186  644
#> 178                        listObjects.html FALSE  10212  644
#>                   mtime               ctime               atime uid gid  uname
#> 1   2022-09-09 15:57:25 2022-09-09 15:57:25 2022-09-09 15:57:25 501  20 runner
#> 2   2022-09-09 15:59:00 2022-09-09 15:59:00 2022-09-09 15:57:26 501  20 runner
#> 3   2022-09-09 15:57:38 2022-09-09 15:57:38 2022-09-09 15:57:29 501  20 runner
#> 4   2022-09-09 15:57:25 2022-09-09 15:57:25 2022-09-09 15:57:25 501  20 runner
#> 5   2022-09-09 15:57:26 2022-09-09 15:57:26 2022-09-09 15:57:26 501  20 runner
#> 6   2022-09-09 15:57:27 2022-09-09 15:57:27 2022-09-09 15:57:27 501  20 runner
#> 7   2022-09-09 15:57:27 2022-09-09 15:57:27 2022-09-09 15:57:27 501  20 runner
#> 8   2022-09-09 15:57:28 2022-09-09 15:57:28 2022-09-09 15:57:28 501  20 runner
#> 9   2022-09-09 15:57:28 2022-09-09 15:57:28 2022-09-09 15:57:28 501  20 runner
#> 10  2022-09-09 15:57:29 2022-09-09 15:57:29 2022-09-09 15:57:29 501  20 runner
#> 11  2022-09-09 15:57:29 2022-09-09 15:57:29 2022-09-09 15:57:29 501  20 runner
#> 12  2022-09-09 15:57:29 2022-09-09 15:57:29 2022-09-09 15:57:29 501  20 runner
#> 13  2022-09-09 15:57:30 2022-09-09 15:57:30 2022-09-09 15:57:30 501  20 runner
#> 14  2022-09-09 15:57:30 2022-09-09 15:57:30 2022-09-09 15:57:30 501  20 runner
#> 15  2022-09-09 15:57:31 2022-09-09 15:57:31 2022-09-09 15:57:31 501  20 runner
#> 16  2022-09-09 15:57:31 2022-09-09 15:57:31 2022-09-09 15:57:31 501  20 runner
#> 17  2022-09-09 15:57:32 2022-09-09 15:57:32 2022-09-09 15:57:32 501  20 runner
#> 18  2022-09-09 15:57:33 2022-09-09 15:57:33 2022-09-09 15:57:33 501  20 runner
#> 19  2022-09-09 15:57:33 2022-09-09 15:57:33 2022-09-09 15:57:33 501  20 runner
#> 20  2022-09-09 15:57:34 2022-09-09 15:57:34 2022-09-09 15:57:34 501  20 runner
#> 21  2022-09-09 15:57:34 2022-09-09 15:57:34 2022-09-09 15:57:34 501  20 runner
#> 22  2022-09-09 15:57:35 2022-09-09 15:57:35 2022-09-09 15:57:35 501  20 runner
#> 23  2022-09-09 15:57:35 2022-09-09 15:57:35 2022-09-09 15:57:35 501  20 runner
#> 24  2022-09-09 15:57:36 2022-09-09 15:57:36 2022-09-09 15:57:36 501  20 runner
#> 25  2022-09-09 15:57:36 2022-09-09 15:57:36 2022-09-09 15:57:36 501  20 runner
#> 26  2022-09-09 15:57:37 2022-09-09 15:57:37 2022-09-09 15:57:37 501  20 runner
#> 27  2022-09-09 15:57:37 2022-09-09 15:57:37 2022-09-09 15:57:37 501  20 runner
#> 28  2022-09-09 15:57:38 2022-09-09 15:57:38 2022-09-09 15:57:38 501  20 runner
#> 29  2022-09-09 15:57:38 2022-09-09 15:57:38 2022-09-09 15:57:38 501  20 runner
#> 30  2022-09-09 15:57:38 2022-09-09 15:57:38 2022-09-09 15:57:38 501  20 runner
#> 31  2022-09-09 15:57:39 2022-09-09 15:57:39 2022-09-09 15:57:39 501  20 runner
#> 32  2022-09-09 15:57:39 2022-09-09 15:57:39 2022-09-09 15:57:39 501  20 runner
#> 33  2022-09-09 15:57:40 2022-09-09 15:57:40 2022-09-09 15:57:40 501  20 runner
#> 34  2022-09-09 15:57:40 2022-09-09 15:57:40 2022-09-09 15:57:40 501  20 runner
#> 35  2022-09-09 15:57:41 2022-09-09 15:57:41 2022-09-09 15:57:41 501  20 runner
#> 36  2022-09-09 15:57:41 2022-09-09 15:57:41 2022-09-09 15:57:41 501  20 runner
#> 37  2022-09-09 15:57:42 2022-09-09 15:57:42 2022-09-09 15:57:42 501  20 runner
#> 38  2022-09-09 15:57:42 2022-09-09 15:57:42 2022-09-09 15:57:42 501  20 runner
#> 39  2022-09-09 15:57:43 2022-09-09 15:57:43 2022-09-09 15:57:43 501  20 runner
#> 40  2022-09-09 15:57:43 2022-09-09 15:57:43 2022-09-09 15:57:43 501  20 runner
#> 41  2022-09-09 15:57:44 2022-09-09 15:57:44 2022-09-09 15:57:44 501  20 runner
#> 42  2022-09-09 15:57:44 2022-09-09 15:57:44 2022-09-09 15:57:44 501  20 runner
#> 43  2022-09-09 15:57:45 2022-09-09 15:57:45 2022-09-09 15:57:45 501  20 runner
#> 44  2022-09-09 15:57:45 2022-09-09 15:57:45 2022-09-09 15:57:45 501  20 runner
#> 45  2022-09-09 15:57:46 2022-09-09 15:57:46 2022-09-09 15:57:46 501  20 runner
#> 46  2022-09-09 15:57:46 2022-09-09 15:57:46 2022-09-09 15:57:46 501  20 runner
#> 47  2022-09-09 15:57:47 2022-09-09 15:57:47 2022-09-09 15:57:47 501  20 runner
#> 48  2022-09-09 15:57:47 2022-09-09 15:57:47 2022-09-09 15:57:47 501  20 runner
#> 49  2022-09-09 15:57:48 2022-09-09 15:57:48 2022-09-09 15:57:48 501  20 runner
#> 50  2022-09-09 15:57:48 2022-09-09 15:57:48 2022-09-09 15:57:48 501  20 runner
#> 51  2022-09-09 15:57:49 2022-09-09 15:57:49 2022-09-09 15:57:49 501  20 runner
#> 52  2022-09-09 15:57:50 2022-09-09 15:57:50 2022-09-09 15:57:50 501  20 runner
#> 53  2022-09-09 15:57:50 2022-09-09 15:57:50 2022-09-09 15:57:50 501  20 runner
#> 54  2022-09-09 15:57:51 2022-09-09 15:57:51 2022-09-09 15:57:51 501  20 runner
#> 55  2022-09-09 15:57:51 2022-09-09 15:57:51 2022-09-09 15:57:51 501  20 runner
#> 56  2022-09-09 15:57:52 2022-09-09 15:57:52 2022-09-09 15:57:52 501  20 runner
#> 57  2022-09-09 15:57:53 2022-09-09 15:57:53 2022-09-09 15:57:53 501  20 runner
#> 58  2022-09-09 15:57:53 2022-09-09 15:57:53 2022-09-09 15:57:53 501  20 runner
#> 59  2022-09-09 15:57:54 2022-09-09 15:57:54 2022-09-09 15:57:54 501  20 runner
#> 60  2022-09-09 15:57:55 2022-09-09 15:57:55 2022-09-09 15:57:55 501  20 runner
#> 61  2022-09-09 15:57:55 2022-09-09 15:57:55 2022-09-09 15:57:55 501  20 runner
#> 62  2022-09-09 15:57:56 2022-09-09 15:57:56 2022-09-09 15:57:56 501  20 runner
#> 63  2022-09-09 15:57:56 2022-09-09 15:57:56 2022-09-09 15:57:56 501  20 runner
#> 64  2022-09-09 15:57:57 2022-09-09 15:57:57 2022-09-09 15:57:57 501  20 runner
#> 65  2022-09-09 15:57:57 2022-09-09 15:57:57 2022-09-09 15:57:57 501  20 runner
#> 66  2022-09-09 15:57:58 2022-09-09 15:57:58 2022-09-09 15:57:58 501  20 runner
#> 67  2022-09-09 15:57:58 2022-09-09 15:57:58 2022-09-09 15:57:58 501  20 runner
#> 68  2022-09-09 15:57:59 2022-09-09 15:57:59 2022-09-09 15:57:59 501  20 runner
#> 69  2022-09-09 15:57:59 2022-09-09 15:57:59 2022-09-09 15:57:59 501  20 runner
#> 70  2022-09-09 15:58:00 2022-09-09 15:58:00 2022-09-09 15:58:00 501  20 runner
#> 71  2022-09-09 15:58:00 2022-09-09 15:58:00 2022-09-09 15:58:00 501  20 runner
#> 72  2022-09-09 15:58:01 2022-09-09 15:58:01 2022-09-09 15:58:01 501  20 runner
#> 73  2022-09-09 15:58:01 2022-09-09 15:58:01 2022-09-09 15:58:01 501  20 runner
#> 74  2022-09-09 15:58:02 2022-09-09 15:58:02 2022-09-09 15:58:02 501  20 runner
#> 75  2022-09-09 15:58:03 2022-09-09 15:58:03 2022-09-09 15:58:03 501  20 runner
#> 76  2022-09-09 15:58:03 2022-09-09 15:58:03 2022-09-09 15:58:03 501  20 runner
#> 77  2022-09-09 15:58:04 2022-09-09 15:58:04 2022-09-09 15:58:04 501  20 runner
#> 78  2022-09-09 15:58:04 2022-09-09 15:58:04 2022-09-09 15:58:04 501  20 runner
#> 79  2022-09-09 15:58:05 2022-09-09 15:58:05 2022-09-09 15:58:05 501  20 runner
#> 80  2022-09-09 15:58:05 2022-09-09 15:58:05 2022-09-09 15:58:05 501  20 runner
#> 81  2022-09-09 15:58:06 2022-09-09 15:58:06 2022-09-09 15:58:06 501  20 runner
#> 82  2022-09-09 15:58:06 2022-09-09 15:58:06 2022-09-09 15:58:06 501  20 runner
#> 83  2022-09-09 15:58:07 2022-09-09 15:58:07 2022-09-09 15:58:07 501  20 runner
#> 84  2022-09-09 15:58:07 2022-09-09 15:58:07 2022-09-09 15:58:07 501  20 runner
#> 85  2022-09-09 15:58:08 2022-09-09 15:58:08 2022-09-09 15:58:08 501  20 runner
#> 86  2022-09-09 15:58:08 2022-09-09 15:58:08 2022-09-09 15:58:08 501  20 runner
#> 87  2022-09-09 15:58:09 2022-09-09 15:58:09 2022-09-09 15:58:09 501  20 runner
#> 88  2022-09-09 15:58:09 2022-09-09 15:58:09 2022-09-09 15:58:09 501  20 runner
#> 89  2022-09-09 15:58:10 2022-09-09 15:58:10 2022-09-09 15:58:10 501  20 runner
#> 90  2022-09-09 15:58:10 2022-09-09 15:58:10 2022-09-09 15:58:10 501  20 runner
#> 91  2022-09-09 15:58:11 2022-09-09 15:58:11 2022-09-09 15:58:11 501  20 runner
#> 92  2022-09-09 15:58:12 2022-09-09 15:58:12 2022-09-09 15:58:12 501  20 runner
#> 93  2022-09-09 15:58:13 2022-09-09 15:58:13 2022-09-09 15:58:13 501  20 runner
#> 94  2022-09-09 15:58:14 2022-09-09 15:58:14 2022-09-09 15:58:14 501  20 runner
#> 95  2022-09-09 15:58:16 2022-09-09 15:58:16 2022-09-09 15:58:16 501  20 runner
#> 96  2022-09-09 15:58:17 2022-09-09 15:58:17 2022-09-09 15:58:17 501  20 runner
#> 97  2022-09-09 15:58:18 2022-09-09 15:58:18 2022-09-09 15:58:18 501  20 runner
#> 98  2022-09-09 15:58:19 2022-09-09 15:58:19 2022-09-09 15:58:19 501  20 runner
#> 99  2022-09-09 15:58:20 2022-09-09 15:58:20 2022-09-09 15:58:20 501  20 runner
#> 100 2022-09-09 15:58:20 2022-09-09 15:58:20 2022-09-09 15:58:20 501  20 runner
#> 101 2022-09-09 15:58:21 2022-09-09 15:58:21 2022-09-09 15:58:21 501  20 runner
#> 102 2022-09-09 15:58:21 2022-09-09 15:58:21 2022-09-09 15:58:21 501  20 runner
#> 103 2022-09-09 15:58:22 2022-09-09 15:58:22 2022-09-09 15:58:22 501  20 runner
#> 104 2022-09-09 15:58:23 2022-09-09 15:58:23 2022-09-09 15:58:23 501  20 runner
#> 105 2022-09-09 15:58:23 2022-09-09 15:58:23 2022-09-09 15:58:23 501  20 runner
#> 106 2022-09-09 15:58:24 2022-09-09 15:58:24 2022-09-09 15:58:24 501  20 runner
#> 107 2022-09-09 15:58:24 2022-09-09 15:58:24 2022-09-09 15:58:24 501  20 runner
#> 108 2022-09-09 15:58:25 2022-09-09 15:58:25 2022-09-09 15:58:25 501  20 runner
#> 109 2022-09-09 15:58:25 2022-09-09 15:58:25 2022-09-09 15:58:25 501  20 runner
#> 110 2022-09-09 15:58:26 2022-09-09 15:58:26 2022-09-09 15:58:26 501  20 runner
#> 111 2022-09-09 15:58:26 2022-09-09 15:58:26 2022-09-09 15:58:26 501  20 runner
#> 112 2022-09-09 15:58:27 2022-09-09 15:58:27 2022-09-09 15:58:27 501  20 runner
#> 113 2022-09-09 15:58:27 2022-09-09 15:58:27 2022-09-09 15:58:27 501  20 runner
#> 114 2022-09-09 15:58:28 2022-09-09 15:58:28 2022-09-09 15:58:28 501  20 runner
#> 115 2022-09-09 15:58:28 2022-09-09 15:58:28 2022-09-09 15:58:28 501  20 runner
#> 116 2022-09-09 15:58:29 2022-09-09 15:58:29 2022-09-09 15:58:29 501  20 runner
#> 117 2022-09-09 15:58:30 2022-09-09 15:58:30 2022-09-09 15:58:30 501  20 runner
#> 118 2022-09-09 15:58:30 2022-09-09 15:58:30 2022-09-09 15:58:30 501  20 runner
#> 119 2022-09-09 15:58:31 2022-09-09 15:58:31 2022-09-09 15:58:31 501  20 runner
#> 120 2022-09-09 15:58:31 2022-09-09 15:58:31 2022-09-09 15:58:31 501  20 runner
#> 121 2022-09-09 15:58:32 2022-09-09 15:58:32 2022-09-09 15:58:32 501  20 runner
#> 122 2022-09-09 15:58:32 2022-09-09 15:58:32 2022-09-09 15:58:32 501  20 runner
#> 123 2022-09-09 15:58:33 2022-09-09 15:58:33 2022-09-09 15:58:33 501  20 runner
#> 124 2022-09-09 15:58:33 2022-09-09 15:58:33 2022-09-09 15:58:33 501  20 runner
#> 125 2022-09-09 15:58:34 2022-09-09 15:58:34 2022-09-09 15:58:34 501  20 runner
#> 126 2022-09-09 15:58:34 2022-09-09 15:58:34 2022-09-09 15:58:34 501  20 runner
#> 127 2022-09-09 15:58:34 2022-09-09 15:58:34 2022-09-09 15:58:34 501  20 runner
#> 128 2022-09-09 15:58:35 2022-09-09 15:58:35 2022-09-09 15:58:35 501  20 runner
#> 129 2022-09-09 15:58:35 2022-09-09 15:58:35 2022-09-09 15:58:35 501  20 runner
#> 130 2022-09-09 15:58:36 2022-09-09 15:58:36 2022-09-09 15:58:36 501  20 runner
#> 131 2022-09-09 15:58:36 2022-09-09 15:58:36 2022-09-09 15:58:36 501  20 runner
#> 132 2022-09-09 15:58:37 2022-09-09 15:58:37 2022-09-09 15:58:37 501  20 runner
#> 133 2022-09-09 15:58:37 2022-09-09 15:58:37 2022-09-09 15:58:37 501  20 runner
#> 134 2022-09-09 15:58:38 2022-09-09 15:58:38 2022-09-09 15:58:38 501  20 runner
#> 135 2022-09-09 15:58:38 2022-09-09 15:58:38 2022-09-09 15:58:38 501  20 runner
#> 136 2022-09-09 15:58:39 2022-09-09 15:58:39 2022-09-09 15:58:39 501  20 runner
#> 137 2022-09-09 15:58:39 2022-09-09 15:58:39 2022-09-09 15:58:39 501  20 runner
#> 138 2022-09-09 15:58:40 2022-09-09 15:58:40 2022-09-09 15:58:40 501  20 runner
#> 139 2022-09-09 15:58:40 2022-09-09 15:58:40 2022-09-09 15:58:40 501  20 runner
#> 140 2022-09-09 15:58:41 2022-09-09 15:58:41 2022-09-09 15:58:41 501  20 runner
#> 141 2022-09-09 15:58:41 2022-09-09 15:58:41 2022-09-09 15:58:41 501  20 runner
#> 142 2022-09-09 15:58:42 2022-09-09 15:58:42 2022-09-09 15:58:42 501  20 runner
#> 143 2022-09-09 15:58:42 2022-09-09 15:58:42 2022-09-09 15:58:42 501  20 runner
#> 144 2022-09-09 15:58:42 2022-09-09 15:58:42 2022-09-09 15:58:42 501  20 runner
#> 145 2022-09-09 15:58:43 2022-09-09 15:58:43 2022-09-09 15:58:43 501  20 runner
#> 146 2022-09-09 15:58:44 2022-09-09 15:58:44 2022-09-09 15:58:44 501  20 runner
#> 147 2022-09-09 15:58:44 2022-09-09 15:58:44 2022-09-09 15:58:44 501  20 runner
#> 148 2022-09-09 15:58:45 2022-09-09 15:58:45 2022-09-09 15:58:45 501  20 runner
#> 149 2022-09-09 15:58:45 2022-09-09 15:58:45 2022-09-09 15:58:45 501  20 runner
#> 150 2022-09-09 15:58:46 2022-09-09 15:58:46 2022-09-09 15:58:46 501  20 runner
#> 151 2022-09-09 15:58:46 2022-09-09 15:58:46 2022-09-09 15:58:46 501  20 runner
#> 152 2022-09-09 15:58:47 2022-09-09 15:58:47 2022-09-09 15:58:47 501  20 runner
#> 153 2022-09-09 15:58:47 2022-09-09 15:58:47 2022-09-09 15:58:47 501  20 runner
#> 154 2022-09-09 15:58:48 2022-09-09 15:58:48 2022-09-09 15:58:48 501  20 runner
#> 155 2022-09-09 15:58:48 2022-09-09 15:58:48 2022-09-09 15:58:48 501  20 runner
#> 156 2022-09-09 15:58:49 2022-09-09 15:58:49 2022-09-09 15:58:49 501  20 runner
#> 157 2022-09-09 15:58:49 2022-09-09 15:58:49 2022-09-09 15:58:49 501  20 runner
#> 158 2022-09-09 15:57:24 2022-09-09 15:57:24 2022-09-09 15:57:24 501  20 runner
#> 159 2022-09-09 15:58:50 2022-09-09 15:58:50 2022-09-09 15:58:50 501  20 runner
#> 160 2022-09-09 15:58:50 2022-09-09 15:58:50 2022-09-09 15:58:50 501  20 runner
#> 161 2022-09-09 15:58:51 2022-09-09 15:58:51 2022-09-09 15:58:51 501  20 runner
#> 162 2022-09-09 15:58:51 2022-09-09 15:58:51 2022-09-09 15:58:51 501  20 runner
#> 163 2022-09-09 15:58:52 2022-09-09 15:58:52 2022-09-09 15:58:52 501  20 runner
#> 164 2022-09-09 15:58:53 2022-09-09 15:58:53 2022-09-09 15:58:53 501  20 runner
#> 165 2022-09-09 15:58:53 2022-09-09 15:58:53 2022-09-09 15:58:53 501  20 runner
#> 166 2022-09-09 15:58:54 2022-09-09 15:58:54 2022-09-09 15:58:54 501  20 runner
#> 167 2022-09-09 15:58:54 2022-09-09 15:58:54 2022-09-09 15:58:54 501  20 runner
#> 168 2022-09-09 15:58:55 2022-09-09 15:58:55 2022-09-09 15:58:55 501  20 runner
#> 169 2022-09-09 15:58:55 2022-09-09 15:58:55 2022-09-09 15:58:55 501  20 runner
#> 170 2022-09-09 15:58:56 2022-09-09 15:58:56 2022-09-09 15:58:56 501  20 runner
#> 171 2022-09-09 15:58:56 2022-09-09 15:58:56 2022-09-09 15:58:56 501  20 runner
#> 172 2022-09-09 15:58:57 2022-09-09 15:58:57 2022-09-09 15:58:57 501  20 runner
#> 173 2022-09-09 15:58:57 2022-09-09 15:58:57 2022-09-09 15:58:57 501  20 runner
#> 174 2022-09-09 15:58:58 2022-09-09 15:58:58 2022-09-09 15:58:58 501  20 runner
#> 175 2022-09-09 15:58:58 2022-09-09 15:58:58 2022-09-09 15:58:58 501  20 runner
#> 176 2022-09-09 15:58:59 2022-09-09 15:58:59 2022-09-09 15:58:59 501  20 runner
#> 177 2022-09-09 15:59:00 2022-09-09 15:59:00 2022-09-09 15:59:00 501  20 runner
#> 178 2022-09-09 15:59:00 2022-09-09 15:59:00 2022-09-09 15:59:00 501  20 runner
#>     grname
#> 1    staff
#> 2    staff
#> 3    staff
#> 4    staff
#> 5    staff
#> 6    staff
#> 7    staff
#> 8    staff
#> 9    staff
#> 10   staff
#> 11   staff
#> 12   staff
#> 13   staff
#> 14   staff
#> 15   staff
#> 16   staff
#> 17   staff
#> 18   staff
#> 19   staff
#> 20   staff
#> 21   staff
#> 22   staff
#> 23   staff
#> 24   staff
#> 25   staff
#> 26   staff
#> 27   staff
#> 28   staff
#> 29   staff
#> 30   staff
#> 31   staff
#> 32   staff
#> 33   staff
#> 34   staff
#> 35   staff
#> 36   staff
#> 37   staff
#> 38   staff
#> 39   staff
#> 40   staff
#> 41   staff
#> 42   staff
#> 43   staff
#> 44   staff
#> 45   staff
#> 46   staff
#> 47   staff
#> 48   staff
#> 49   staff
#> 50   staff
#> 51   staff
#> 52   staff
#> 53   staff
#> 54   staff
#> 55   staff
#> 56   staff
#> 57   staff
#> 58   staff
#> 59   staff
#> 60   staff
#> 61   staff
#> 62   staff
#> 63   staff
#> 64   staff
#> 65   staff
#> 66   staff
#> 67   staff
#> 68   staff
#> 69   staff
#> 70   staff
#> 71   staff
#> 72   staff
#> 73   staff
#> 74   staff
#> 75   staff
#> 76   staff
#> 77   staff
#> 78   staff
#> 79   staff
#> 80   staff
#> 81   staff
#> 82   staff
#> 83   staff
#> 84   staff
#> 85   staff
#> 86   staff
#> 87   staff
#> 88   staff
#> 89   staff
#> 90   staff
#> 91   staff
#> 92   staff
#> 93   staff
#> 94   staff
#> 95   staff
#> 96   staff
#> 97   staff
#> 98   staff
#> 99   staff
#> 100  staff
#> 101  staff
#> 102  staff
#> 103  staff
#> 104  staff
#> 105  staff
#> 106  staff
#> 107  staff
#> 108  staff
#> 109  staff
#> 110  staff
#> 111  staff
#> 112  staff
#> 113  staff
#> 114  staff
#> 115  staff
#> 116  staff
#> 117  staff
#> 118  staff
#> 119  staff
#> 120  staff
#> 121  staff
#> 122  staff
#> 123  staff
#> 124  staff
#> 125  staff
#> 126  staff
#> 127  staff
#> 128  staff
#> 129  staff
#> 130  staff
#> 131  staff
#> 132  staff
#> 133  staff
#> 134  staff
#> 135  staff
#> 136  staff
#> 137  staff
#> 138  staff
#> 139  staff
#> 140  staff
#> 141  staff
#> 142  staff
#> 143  staff
#> 144  staff
#> 145  staff
#> 146  staff
#> 147  staff
#> 148  staff
#> 149  staff
#> 150  staff
#> 151  staff
#> 152  staff
#> 153  staff
#> 154  staff
#> 155  staff
#> 156  staff
#> 157  staff
#> 158  staff
#> 159  staff
#> 160  staff
#> 161  staff
#> 162  staff
#> 163  staff
#> 164  staff
#> 165  staff
#> 166  staff
#> 167  staff
#> 168  staff
#> 169  staff
#> 170  staff
#> 171  staff
#> 172  staff
#> 173  staff
#> 174  staff
#> 175  staff
#> 176  staff
#> 177  staff
#> 178  staff

# ... even when being called with an empty character vector
FUN(character())
#> listing 
#> listing .
#> [1] file  isdir
#> <0 rows> (or 0-length row.names)
FUN(character(), full_info = TRUE)
#> listing 
#> listing .
#>  [1] file   isdir  size   mode   mtime  ctime  atime  uid    gid    uname 
#> [11] grname
#> <0 rows> (or 0-length row.names)

# Call the function recursively up to the given depth level
kwb.utils:::listToDepth(".", max_depth = 1, FUN = FUN)
#> listing .
#> listing .
#>                                        file isdir
#> 1                               DIN.A4.html FALSE
#> 2                              Rplot001.png FALSE
#> 3                              Rplot002.png FALSE
#> 4                             addClass.html FALSE
#> 5                       addRowWithName.html FALSE
#> 6                   addSuffixToColumns.html FALSE
#> 7                          allAreEqual.html FALSE
#> 8                      allAreIdentical.html FALSE
#> 9                          almostEqual.html FALSE
#> 10                        appendSuffix.html FALSE
#> 11                            arglist-1.png FALSE
#> 12                            arglist-2.png FALSE
#> 13                             arglist.html FALSE
#> 14                             argsCsv.html FALSE
#> 15                    arrayToDataFrame.html FALSE
#> 16                        asColumnList.html FALSE
#> 17                 asNoFactorDataFrame.html FALSE
#> 18                           asRowList.html FALSE
#> 19                    assertFinalSlash.html FALSE
#> 20                assertRowsAndColumns.html FALSE
#> 21                           assignAll.html FALSE
#> 22              assignArgumentDefaults.html FALSE
#> 23                      assignGlobally.html FALSE
#> 24                       assignObjects.html FALSE
#> 25                assignPackageObjects.html FALSE
#> 26                     atLeastOneRowIn.html FALSE
#> 27                           backspace.html FALSE
#> 28                     breakInSequence.html FALSE
#> 29                           callWith-1.png FALSE
#> 30                           callWith-2.png FALSE
#> 31                            callWith.html FALSE
#> 32            callWithStringsAsFactors.html FALSE
#> 33                           catAndRun.html FALSE
#> 34                          catChanges.html FALSE
#> 35                        catChangesIf.html FALSE
#> 36                               catIf.html FALSE
#> 37                            catLines.html FALSE
#> 38                        catNewLineIf.html FALSE
#> 39                        catchWarning.html FALSE
#> 40              checkForMissingColumns.html FALSE
#> 41                        clearConsole.html FALSE
#> 42                          clipMatrix.html FALSE
#> 43                         cmdLinePath.html FALSE
#> 44                           colMaxima.html FALSE
#> 45                           colMinima.html FALSE
#> 46                        colNaNumbers.html FALSE
#> 47             colStatisticOneFunction.html FALSE
#> 48                       colStatistics.html FALSE
#> 49                           collapsed.html FALSE
#> 50                    columnDescriptor.html FALSE
#> 51                        columnToDate.html FALSE
#> 52                columnwisePercentage.html FALSE
#> 53                combineAlternatingly.html FALSE
#> 54                      commaCollapsed.html FALSE
#> 55                         commonNames.html FALSE
#> 56                   compareDataFrames.html FALSE
#> 57                         compareSets.html FALSE
#> 58                   containsNulString.html FALSE
#> 59                      convertCsvFile.html FALSE
#> 60                      copyAttributes.html FALSE
#> 61              copyDirectoryStructure.html FALSE
#> 62                    copyListElements.html FALSE
#> 63                     countNaInColumn.html FALSE
#> 64                          countOrSum.html FALSE
#> 65                      createAccessor.html FALSE
#> 66              createDirAndReturnPath.html FALSE
#> 67                   createDirectories.html FALSE
#> 68                     createDirectory.html FALSE
#> 69         createFunctionAssignObjects.html FALSE
#> 70           createFunctionExtdataFile.html FALSE
#> 71                       createIdAlong.html FALSE
#> 72                        createMatrix.html FALSE
#> 73                  createPasswordFile.html FALSE
#> 74                       createStorage.html FALSE
#> 75                  csvTextToDataFrame.html FALSE
#> 76                              decode.html FALSE
#> 77                           defaultIf.html FALSE
#> 78                         defaultIfNA.html FALSE
#> 79                       defaultIfNULL.html FALSE
#> 80                       defaultIfZero.html FALSE
#> 81                       defaultLevels.html FALSE
#> 82        defaultWindowsProgramFolders.html FALSE
#> 83                             desktop.html FALSE
#> 84                            diffrows.html FALSE
#> 85                       directoryName.html FALSE
#> 86                          dot-OStype.html FALSE
#> 87                             dot-log.html FALSE
#> 88                         dot-logline.html FALSE
#> 89                           dot-logok.html FALSE
#> 90                        dot-logstart.html FALSE
#> 91                             dropDim.html FALSE
#> 92              dropUnusedFactorLevels.html FALSE
#> 93                              encode.html FALSE
#> 94                       enlargeVector.html FALSE
#> 95                         excludeNULL.html FALSE
#> 96                          expandGrid.html FALSE
#> 97                         extdataFile.html FALSE
#> 98                        extendLimits.html FALSE
#> 99                    extractRowRanges.html FALSE
#> 100                   extractSubstring.html FALSE
#> 101                      fileExtension.html FALSE
#> 102                        findChanges.html FALSE
#> 103              findPartialDuplicates.html FALSE
#> 104                   finishAndShowPdf.html FALSE
#> 105                 finishAndShowPdfIf.html FALSE
#> 106                       firstElement.html FALSE
#> 107                   firstPosixColumn.html FALSE
#> 108                      frenchToAscii.html FALSE
#> 109                     frequencyTable.html FALSE
#> 110                        fullWinPath.html FALSE
#> 111                        fullySorted.html FALSE
#> 112                    generateKeyFile.html FALSE
#> 113                       getAttribute.html FALSE
#> 114       getByPositiveOrNegativeIndex.html FALSE
#> 115                  getElementLengths.html FALSE
#> 116                     getEvenNumbers.html FALSE
#> 117          getFunctionValueOrDefault.html FALSE
#> 118                        getGlobally.html FALSE
#> 119                getKeywordPositions.html FALSE
#> 120                        getListNode.html FALSE
#> 121      getNamesOfObjectsInRDataFiles.html FALSE
#> 122             getObjectFromRDataFile.html FALSE
#> 123                      getOddNumbers.html FALSE
#> 124                        getPassword.html FALSE
#> 125 getPathsAndValuesFromRecursiveList.html FALSE
#> 126                        getTagNames.html FALSE
#> 127                        get_homedir.html FALSE
#> 128                   guessSeparator.1.html FALSE
#> 129                   guessSeparator.2.html FALSE
#> 130                     guessSeparator.html FALSE
#> 131                      hasFinalSlash.html FALSE
#> 132                      hasZeroLength.html FALSE
#> 133                           headtail.html FALSE
#> 134                   hsAddMissingCols.html FALSE
#> 135                         hsChrToNum.html FALSE
#> 136                       hsCountInStr.html FALSE
#> 137                     hsDelEmptyCols.html FALSE
#> 138                 hsMatrixToListForm.html FALSE
#> 139                      hsMovingMean-1.png FALSE
#> 140                       hsMovingMean.html FALSE
#> 141              hsOpenWindowsExplorer.html FALSE
#> 142                          hsPrepPdf.html FALSE
#> 143                         hsQuoteChr.html FALSE
#> 144                    hsRenameColumns.html FALSE
#> 145                          hsResolve.html FALSE
#> 146                hsRestoreAttributes.html FALSE
#> 147                         hsSafeName.html FALSE
#> 148                            hsShell.html FALSE
#> 149                          hsShowPdf.html FALSE
#> 150                     hsStringToDate.html FALSE
#> 151                   hsStringToDouble.html FALSE
#> 152                   hsSubstSpecChars.html FALSE
#> 153                           hsSystem.html FALSE
#> 154                             hsTrim.html FALSE
#> 155                       hsValidValue.html FALSE
#> 156                            inRange.html FALSE
#> 157                             indent.html FALSE
#> 158                              index.html FALSE
#> 159                      insertColumns.html FALSE
#> 160                 intToNumeralSystem.html FALSE
#> 161                         is.unnamed.html FALSE
#> 162                            isASCII.html FALSE
#> 163                   isDotOrDoubleDot.html FALSE
#> 164                       isEvenNumber.html FALSE
#> 165                           isLoaded.html FALSE
#> 166                   isNaInAllColumns.html FALSE
#> 167                      isNaInAllRows.html FALSE
#> 168                        isNaOrEmpty.html FALSE
#> 169                      isNetworkPath.html FALSE
#> 170                      isNullOrEmpty.html FALSE
#> 171                        isOddNumber.html FALSE
#> 172                         isTryError.html FALSE
#> 173                        lastElement.html FALSE
#> 174                               left.html FALSE
#> 175                leftSubstringEquals.html FALSE
#> 176                       limitToRange.html FALSE
#> 177                  linearCombination.html FALSE
#> 178                        listObjects.html FALSE