Create a Matrix with Random Integer Values

Create a matrix of given dimension and fill it with random integer values

randomMatrix(dim = c(sample(10, 1), sample(10, 1)), values = seq_len(100))

Arguments

dim

integer vector of length two containing the number of rows and columns, respectively, that the output matrix shall contain

values

set of values to be used within the matrix

Examples

# By default, the matrix has a random number of rows between 1 and 10 and
# a random number of columns between 1 and 10 and random values of 1:100
randomMatrix()
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7]
#> [1,]   58    9    2   83   69   74   65
#> [2,]   46   52   69   44   44   63   12
#> [3,]   14   35   72   10   11   34   32
#> [4,]   74   93    8   60    1   82   75
#> [5,]   90   44   10   63   50   33   48
#> [6,]   77   24   45   22   34   54   50

# You may specify the dimensions (here: 5 rows, 3 columns)...
randomMatrix(dim = c(5, 3))
#>      [,1] [,2] [,3]
#> [1,]   97    7   39
#> [2,]   68    7   86
#> [3,]   53   47   40
#> [4,]    1   62   75
#> [5,]   84   16   64

# ... and the set of values to be used within the matrix
randomMatrix(dim = c(5, 3), values = c(0, 0.5, 1, NA))
#>      [,1] [,2] [,3]
#> [1,]   NA    1  1.0
#> [2,]  0.0   NA  1.0
#> [3,]  1.0    1  0.5
#> [4,]  0.5    1  0.0
#> [5,]  1.0    0   NA