Select (and Rename) Elements from List

select (and rename, if required) elements from list. Stop with message if elements do not exist

selectElements(x, elements = NULL, do.stop = TRUE, do.warn = TRUE)

Arguments

x
list

elements
vector of element names. The names of named elements will be the names in the output list

do.stop
this flag controls whether the function stops (do.stop = TRUE) or not (do.stop = FALSE) if there are non-existing elements to be selected. If do.stop = FALSE only those elements are selected that actually exist

do.warn
if TRUE (default) and do.stop = FALSE a warning is given if elements do not exist. Set to FALSE to suppress warnings

Value

list containing the elements of x that are specified in elements or x[[elements]] if length of elements is 1 or list() if elements is empty. If the elements in vector elements are named, these names are used in the output list.

Examples

L <- list(a = 1, b = 2, c = 3, d = 4)

# Select elements
selectElements(L, c("a", "c"))

#> $a
#> [1] 1
#> 
#> $c
#> [1] 3
#> 

# Select and rename at the same time
selectElements(L, elements = c(a.new = "a", c.new = "c", "b"))

#> $a.new
#> [1] 1
#> 
#> $c.new
#> [1] 3
#> 
#> $b
#> [1] 2
#>