Columnwise Percentage

Calculate the percentage (value divided by sum of values in the column) for each column

Usage

columnwisePercentage(x, default = 0, digits = 1)

Arguments

x

two dimensional numeric data structure

default

default value to be used if the calculated percentage is NA.

digits

number of digits (default: 1) to which the resulting percentages are to be rounded. Set to NA to suppress rounding

Examples

# Create a random matrix of integer values
M1 <- matrix(sample(100, 12), nrow = 4, dimnames = list(LETTERS[1:4], 1:3))

# Introduce some NA
values <- as.numeric(M1)
values[sample(length(values), 3)] <- NA
M2 <- matrix(values, nrow = nrow(M1), dimnames = dimnames(M1))

M1
#>    1  2  3
#> A  1  3 83
#> B 99 78 26
#> C 76 84 72
#> D 28 47 60
columnwisePercentage(M1)
#>      1    2    3
#> A  0.5  1.4 34.4
#> B 48.5 36.8 10.8
#> C 37.3 39.6 29.9
#> D 13.7 22.2 24.9

M2
#>    1  2  3
#> A  1  3 83
#> B 99 78 NA
#> C 76 84 72
#> D NA NA 60
columnwisePercentage(M2)
#>      1    2    3
#> A  0.6  1.8 38.6
#> B 56.2 47.3  0.0
#> C 43.2 50.9 33.5
#> D  0.0  0.0 27.9
columnwisePercentage(M2, default = 0)
#>      1    2    3
#> A  0.6  1.8 38.6
#> B 56.2 47.3  0.0
#> C 43.2 50.9 33.5
#> D  0.0  0.0 27.9