List Elements Recursively up to Depth

List Elements Recursively up to Depth

Usage

listToDepth(
  path,
  max_depth = 0L,
  full_info = FALSE,
  FUN = listFiles,
  ...,
  depth = 0,
  prob_mutate = 0,
  template = NULL
)

Arguments

path

path to the element at which to start listing

max_depth

maximal depth level of which to list elements. A value of 0 means non-recursive listing, a value of NA represents fully recursive listing.

full_info

return only path and isdir information or the full information provided by FUN(full_info = TRUE)?

FUN

function called to get the listing of the element given in path. The function must accept a path as its first argument and it must define the argument full_info second. It may accept further arguments. It must always return a data frame. For full_info = FALSE the data frame must have columns file and isdir (is the "file" a directory?). For full_info = TRUE the function may return further columns. The function must provide an empty data frame with the expected columns when being called with character() as the first argument. The function is expected to set the attribute "failed" to the given path in case that the path could not be accessed (e.g. because of a broken internet connection if the listing is done remotely). See kwb.utils:::listFiles for an example implementation that somehow simulates the behaviour of the dir function. See kwb.dwd::list_url() for a more advanced usage of this function in order to recursively list the files on an FTP server (FTP = file transfer protocol).

...

further arguments passed to FUN

depth

start depth of recursion if max_depth > 0. This argument is for internal use and not intended to be set by the user!

prob_mutate

probability to alter the path so that it becomes useless. This is zero by default. Set the value only if you want to test how the function behaves if the listing of a path fails.

template

empty data frame (zero rows) and with columns that are identical to the columns in the data frame returned by FUN. If not provided the function will call FUN once with a zero-length path argument expecting to retrieve the template that is expected here.

Value

data frame containing at least the columns file and isdir. If full_info = TRUE the result data frame may contain further columns, as provided by the function given in FUN for full_info = TRUE.

Examples

# Example list function provided in this package (file listing)
FUN <- kwb.utils:::listFiles

# The list function must provide empty records when no path is given. The 
# returned data frame must provide the columns "file" and "isdir" ...
FUN()
#> listing .
#>                                        file isdir
#> 1                               DIN.A4.html FALSE
#> 2                              Rplot001.png FALSE
#> 3                              Rplot002.png FALSE
#> 4                             addClass.html FALSE
#> 5                       addRowWithName.html FALSE
#> 6                   addSuffixToColumns.html FALSE
#> 7                          allAreEqual.html FALSE
#> 8                      allAreIdentical.html FALSE
#> 9                          almostEqual.html FALSE
#> 10                        appendSuffix.html FALSE
#> 11                            arglist-1.png FALSE
#> 12                            arglist-2.png FALSE
#> 13                             arglist.html FALSE
#> 14                             argsCsv.html FALSE
#> 15                    arrayToDataFrame.html FALSE
#> 16                        asColumnList.html FALSE
#> 17                 asNoFactorDataFrame.html FALSE
#> 18                           asRowList.html FALSE
#> 19                    assertFinalSlash.html FALSE
#> 20                assertRowsAndColumns.html FALSE
#> 21                           assignAll.html FALSE
#> 22              assignArgumentDefaults.html FALSE
#> 23                      assignGlobally.html FALSE
#> 24                assignPackageObjects.html FALSE
#> 25                     atLeastOneRowIn.html FALSE
#> 26                           backspace.html FALSE
#> 27                     breakInSequence.html FALSE
#> 28                           callWith-1.png FALSE
#> 29                           callWith-2.png FALSE
#> 30                            callWith.html FALSE
#> 31            callWithStringsAsFactors.html FALSE
#> 32                           catAndRun.html FALSE
#> 33                          catChanges.html FALSE
#> 34                        catChangesIf.html FALSE
#> 35                               catIf.html FALSE
#> 36                            catLines.html FALSE
#> 37                        catNewLineIf.html FALSE
#> 38              checkForMissingColumns.html FALSE
#> 39                        clearConsole.html FALSE
#> 40                          clipMatrix.html FALSE
#> 41                         cmdLinePath.html FALSE
#> 42                           colMaxima.html FALSE
#> 43                           colMinima.html FALSE
#> 44                        colNaNumbers.html FALSE
#> 45             colStatisticOneFunction.html FALSE
#> 46                       colStatistics.html FALSE
#> 47                           collapsed.html FALSE
#> 48                    columnDescriptor.html FALSE
#> 49                        columnToDate.html FALSE
#> 50                columnwisePercentage.html FALSE
#> 51                combineAlternatingly.html FALSE
#> 52                      commaCollapsed.html FALSE
#> 53                         commonNames.html FALSE
#> 54                   compareDataFrames.html FALSE
#> 55                         compareSets.html FALSE
#> 56                   containsNulString.html FALSE
#> 57                      convertCsvFile.html FALSE
#> 58                      copyAttributes.html FALSE
#> 59              copyDirectoryStructure.html FALSE
#> 60                    copyListElements.html FALSE
#> 61                     countNaInColumn.html FALSE
#> 62                          countOrSum.html FALSE
#> 63                      createAccessor.html FALSE
#> 64              createDirAndReturnPath.html FALSE
#> 65                   createDirectories.html FALSE
#> 66                     createDirectory.html FALSE
#> 67           createFunctionExtdataFile.html FALSE
#> 68                       createIdAlong.html FALSE
#> 69                        createMatrix.html FALSE
#> 70                  createPasswordFile.html FALSE
#> 71                       createStorage.html FALSE
#> 72                  csvTextToDataFrame.html FALSE
#> 73                              decode.html FALSE
#> 74                           defaultIf.html FALSE
#> 75                         defaultIfNA.html FALSE
#> 76                       defaultIfNULL.html FALSE
#> 77                       defaultIfZero.html FALSE
#> 78                       defaultLevels.html FALSE
#> 79        defaultWindowsProgramFolders.html FALSE
#> 80                             desktop.html FALSE
#> 81                            diffrows.html FALSE
#> 82                       directoryName.html FALSE
#> 83                          dot-OStype.html FALSE
#> 84                             dot-log.html FALSE
#> 85                         dot-logline.html FALSE
#> 86                           dot-logok.html FALSE
#> 87                        dot-logstart.html FALSE
#> 88                             dropDim.html FALSE
#> 89              dropUnusedFactorLevels.html FALSE
#> 90                              encode.html FALSE
#> 91                       enlargeVector.html FALSE
#> 92                         excludeNULL.html FALSE
#> 93                          expandGrid.html FALSE
#> 94                         extdataFile.html FALSE
#> 95                        extendLimits.html FALSE
#> 96                    extractRowRanges.html FALSE
#> 97                    extractSubstring.html FALSE
#> 98                       fileExtension.html FALSE
#> 99                         findChanges.html FALSE
#> 100              findPartialDuplicates.html FALSE
#> 101                   finishAndShowPdf.html FALSE
#> 102                 finishAndShowPdfIf.html FALSE
#> 103                       firstElement.html FALSE
#> 104                   firstPosixColumn.html FALSE
#> 105                      frenchToAscii.html FALSE
#> 106                     frequencyTable.html FALSE
#> 107                        fullWinPath.html FALSE
#> 108                        fullySorted.html FALSE
#> 109                    generateKeyFile.html FALSE
#> 110                       getAttribute.html FALSE
#> 111       getByPositiveOrNegativeIndex.html FALSE
#> 112                  getElementLengths.html FALSE
#> 113                     getEvenNumbers.html FALSE
#> 114          getFunctionValueOrDefault.html FALSE
#> 115                        getGlobally.html FALSE
#> 116                getKeywordPositions.html FALSE
#> 117                        getListNode.html FALSE
#> 118      getNamesOfObjectsInRDataFiles.html FALSE
#> 119             getObjectFromRDataFile.html FALSE
#> 120                      getOddNumbers.html FALSE
#> 121                        getPassword.html FALSE
#> 122 getPathsAndValuesFromRecursiveList.html FALSE
#> 123                        getTagNames.html FALSE
#> 124                        get_homedir.html FALSE
#> 125                   guessSeparator.1.html FALSE
#> 126                   guessSeparator.2.html FALSE
#> 127                     guessSeparator.html FALSE
#> 128                      hasFinalSlash.html FALSE
#> 129                      hasZeroLength.html FALSE
#> 130                           headtail.html FALSE
#> 131                   hsAddMissingCols.html FALSE
#> 132                         hsChrToNum.html FALSE
#> 133                       hsCountInStr.html FALSE
#> 134                     hsDelEmptyCols.html FALSE
#> 135                 hsMatrixToListForm.html FALSE
#> 136                      hsMovingMean-1.png FALSE
#> 137                       hsMovingMean.html FALSE
#> 138              hsOpenWindowsExplorer.html FALSE
#> 139                          hsPrepPdf.html FALSE
#> 140                         hsQuoteChr.html FALSE
#> 141                    hsRenameColumns.html FALSE
#> 142                          hsResolve.html FALSE
#> 143                hsRestoreAttributes.html FALSE
#> 144                         hsSafeName.html FALSE
#> 145                            hsShell.html FALSE
#> 146                          hsShowPdf.html FALSE
#> 147                     hsStringToDate.html FALSE
#> 148                   hsStringToDouble.html FALSE
#> 149                   hsSubstSpecChars.html FALSE
#> 150                           hsSystem.html FALSE
#> 151                             hsTrim.html FALSE
#> 152                       hsValidValue.html FALSE
#> 153                            inRange.html FALSE
#> 154                             indent.html FALSE
#> 155                              index.html FALSE
#> 156                      insertColumns.html FALSE
#> 157                 intToNumeralSystem.html FALSE
#> 158                         is.unnamed.html FALSE
#> 159                            isASCII.html FALSE
#> 160                   isDotOrDoubleDot.html FALSE
#> 161                       isEvenNumber.html FALSE
#> 162                           isLoaded.html FALSE
#> 163                   isNaInAllColumns.html FALSE
#> 164                      isNaInAllRows.html FALSE
#> 165                        isNaOrEmpty.html FALSE
#> 166                      isNetworkPath.html FALSE
#> 167                      isNullOrEmpty.html FALSE
#> 168                        isOddNumber.html FALSE
#> 169                         isTryError.html FALSE
#> 170                        lastElement.html FALSE
#> 171                               left.html FALSE
#> 172                leftSubstringEquals.html FALSE
#> 173                       limitToRange.html FALSE
#> 174                  linearCombination.html FALSE
#> 175                        listObjects.html FALSE
FUN(full_info = TRUE)
#> listing .
#>                                        file isdir   size mode
#> 1                               DIN.A4.html FALSE   8219  644
#> 2                              Rplot001.png FALSE   1011  644
#> 3                              Rplot002.png FALSE   5908  644
#> 4                             addClass.html FALSE   9357  644
#> 5                       addRowWithName.html FALSE   9073  644
#> 6                   addSuffixToColumns.html FALSE  11665  644
#> 7                          allAreEqual.html FALSE   9028  644
#> 8                      allAreIdentical.html FALSE   8861  644
#> 9                          almostEqual.html FALSE   9044  644
#> 10                        appendSuffix.html FALSE  10620  644
#> 11                            arglist-1.png FALSE  18498  644
#> 12                            arglist-2.png FALSE  18423  644
#> 13                             arglist.html FALSE  12842  644
#> 14                             argsCsv.html FALSE   8904  644
#> 15                    arrayToDataFrame.html FALSE  13424  644
#> 16                        asColumnList.html FALSE  10866  644
#> 17                 asNoFactorDataFrame.html FALSE  11176  644
#> 18                           asRowList.html FALSE  10750  644
#> 19                    assertFinalSlash.html FALSE  11783  644
#> 20                assertRowsAndColumns.html FALSE  12269  644
#> 21                           assignAll.html FALSE   8866  644
#> 22              assignArgumentDefaults.html FALSE  10182  644
#> 23                      assignGlobally.html FALSE   8556  644
#> 24                assignPackageObjects.html FALSE   9101  644
#> 25                     atLeastOneRowIn.html FALSE   8632  644
#> 26                           backspace.html FALSE  10333  644
#> 27                     breakInSequence.html FALSE   9007  644
#> 28                           callWith-1.png FALSE  18498  644
#> 29                           callWith-2.png FALSE  18423  644
#> 30                            callWith.html FALSE  11942  644
#> 31            callWithStringsAsFactors.html FALSE  14217  644
#> 32                           catAndRun.html FALSE  12602  644
#> 33                          catChanges.html FALSE   9470  644
#> 34                        catChangesIf.html FALSE   8928  644
#> 35                               catIf.html FALSE   8589  644
#> 36                            catLines.html FALSE   8694  644
#> 37                        catNewLineIf.html FALSE   8941  644
#> 38              checkForMissingColumns.html FALSE   9764  644
#> 39                        clearConsole.html FALSE   8151  644
#> 40                          clipMatrix.html FALSE   8882  644
#> 41                         cmdLinePath.html FALSE   8667  644
#> 42                           colMaxima.html FALSE   8679  644
#> 43                           colMinima.html FALSE   8679  644
#> 44                        colNaNumbers.html FALSE   8617  644
#> 45             colStatisticOneFunction.html FALSE   9104  644
#> 46                       colStatistics.html FALSE   9653  644
#> 47                           collapsed.html FALSE   8872  644
#> 48                    columnDescriptor.html FALSE   8783  644
#> 49                        columnToDate.html FALSE  11219  644
#> 50                columnwisePercentage.html FALSE  14519  644
#> 51                combineAlternatingly.html FALSE  11870  644
#> 52                      commaCollapsed.html FALSE   8497  644
#> 53                         commonNames.html FALSE  10010  644
#> 54                   compareDataFrames.html FALSE  13226  644
#> 55                         compareSets.html FALSE  11641  644
#> 56                   containsNulString.html FALSE   8802  644
#> 57                      convertCsvFile.html FALSE  21939  644
#> 58                      copyAttributes.html FALSE  10480  644
#> 59              copyDirectoryStructure.html FALSE   9630  644
#> 60                    copyListElements.html FALSE  11791  644
#> 61                     countNaInColumn.html FALSE   8866  644
#> 62                          countOrSum.html FALSE  13598  644
#> 63                      createAccessor.html FALSE  10599  644
#> 64              createDirAndReturnPath.html FALSE   9346  644
#> 65                   createDirectories.html FALSE   8839  644
#> 66                     createDirectory.html FALSE   9239  644
#> 67           createFunctionExtdataFile.html FALSE  10011  644
#> 68                       createIdAlong.html FALSE  11786  644
#> 69                        createMatrix.html FALSE  13625  644
#> 70                  createPasswordFile.html FALSE   9678  644
#> 71                       createStorage.html FALSE  14435  644
#> 72                  csvTextToDataFrame.html FALSE   8728  644
#> 73                              decode.html FALSE   8792  644
#> 74                           defaultIf.html FALSE   9215  644
#> 75                         defaultIfNA.html FALSE  10354  644
#> 76                       defaultIfNULL.html FALSE  10377  644
#> 77                       defaultIfZero.html FALSE  11630  644
#> 78                       defaultLevels.html FALSE  10470  644
#> 79        defaultWindowsProgramFolders.html FALSE   8295  644
#> 80                             desktop.html FALSE   8857  644
#> 81                            diffrows.html FALSE  10672  644
#> 82                       directoryName.html FALSE  10567  644
#> 83                          dot-OStype.html FALSE   8544  644
#> 84                             dot-log.html FALSE   8592  644
#> 85                         dot-logline.html FALSE   8746  644
#> 86                           dot-logok.html FALSE   8682  644
#> 87                        dot-logstart.html FALSE   8875  644
#> 88                             dropDim.html FALSE  13621  644
#> 89              dropUnusedFactorLevels.html FALSE  13662  644
#> 90                              encode.html FALSE  11045  644
#> 91                       enlargeVector.html FALSE  11370  644
#> 92                         excludeNULL.html FALSE  10891  644
#> 93                          expandGrid.html FALSE  14427  644
#> 94                         extdataFile.html FALSE   9763  644
#> 95                        extendLimits.html FALSE   9196  644
#> 96                    extractRowRanges.html FALSE  18729  644
#> 97                    extractSubstring.html FALSE  14890  644
#> 98                       fileExtension.html FALSE   9893  644
#> 99                         findChanges.html FALSE  13116  644
#> 100              findPartialDuplicates.html FALSE  13128  644
#> 101                   finishAndShowPdf.html FALSE   9342  644
#> 102                 finishAndShowPdfIf.html FALSE   8951  644
#> 103                       firstElement.html FALSE   8680  644
#> 104                   firstPosixColumn.html FALSE   8577  644
#> 105                      frenchToAscii.html FALSE   8598  644
#> 106                     frequencyTable.html FALSE  15532  644
#> 107                        fullWinPath.html FALSE  10508  644
#> 108                        fullySorted.html FALSE  14399  644
#> 109                    generateKeyFile.html FALSE   8899  644
#> 110                       getAttribute.html FALSE  13053  644
#> 111       getByPositiveOrNegativeIndex.html FALSE   9110  644
#> 112                  getElementLengths.html FALSE   9972  644
#> 113                     getEvenNumbers.html FALSE   8612  644
#> 114          getFunctionValueOrDefault.html FALSE   9140  644
#> 115                        getGlobally.html FALSE   8947  644
#> 116                getKeywordPositions.html FALSE   9363  644
#> 117                        getListNode.html FALSE  12531  644
#> 118      getNamesOfObjectsInRDataFiles.html FALSE   8747  644
#> 119             getObjectFromRDataFile.html FALSE   9296  644
#> 120                      getOddNumbers.html FALSE   8602  644
#> 121                        getPassword.html FALSE   9904  644
#> 122 getPathsAndValuesFromRecursiveList.html FALSE  12368  644
#> 123                        getTagNames.html FALSE   9595  644
#> 124                        get_homedir.html FALSE   9013  644
#> 125                   guessSeparator.1.html FALSE   8892  644
#> 126                   guessSeparator.2.html FALSE   8768  644
#> 127                     guessSeparator.html FALSE   9125  644
#> 128                      hasFinalSlash.html FALSE   9597  644
#> 129                      hasZeroLength.html FALSE  11064  644
#> 130                           headtail.html FALSE  14972  644
#> 131                   hsAddMissingCols.html FALSE   9551  644
#> 132                         hsChrToNum.html FALSE   9507  644
#> 133                       hsCountInStr.html FALSE   8916  644
#> 134                     hsDelEmptyCols.html FALSE  10080  644
#> 135                 hsMatrixToListForm.html FALSE  10439  644
#> 136                      hsMovingMean-1.png FALSE 112525  644
#> 137                       hsMovingMean.html FALSE  12263  644
#> 138              hsOpenWindowsExplorer.html FALSE   9088  644
#> 139                          hsPrepPdf.html FALSE   9789  644
#> 140                         hsQuoteChr.html FALSE   9193  644
#> 141                    hsRenameColumns.html FALSE   8875  644
#> 142                          hsResolve.html FALSE   9423  644
#> 143                hsRestoreAttributes.html FALSE   8770  644
#> 144                         hsSafeName.html FALSE  10953  644
#> 145                            hsShell.html FALSE   8607  644
#> 146                          hsShowPdf.html FALSE  11910  644
#> 147                     hsStringToDate.html FALSE   9221  644
#> 148                   hsStringToDouble.html FALSE   9017  644
#> 149                   hsSubstSpecChars.html FALSE   9104  644
#> 150                           hsSystem.html FALSE   8619  644
#> 151                             hsTrim.html FALSE   9293  644
#> 152                       hsValidValue.html FALSE   9416  644
#> 153                            inRange.html FALSE   9179  644
#> 154                             indent.html FALSE   8845  644
#> 155                              index.html FALSE  64664  644
#> 156                      insertColumns.html FALSE  16072  644
#> 157                 intToNumeralSystem.html FALSE  13893  644
#> 158                         is.unnamed.html FALSE  11218  644
#> 159                            isASCII.html FALSE  10098  644
#> 160                   isDotOrDoubleDot.html FALSE   9575  644
#> 161                       isEvenNumber.html FALSE   8478  644
#> 162                           isLoaded.html FALSE  10716  644
#> 163                   isNaInAllColumns.html FALSE   8806  644
#> 164                      isNaInAllRows.html FALSE   8785  644
#> 165                        isNaOrEmpty.html FALSE   8919  644
#> 166                      isNetworkPath.html FALSE   9092  644
#> 167                      isNullOrEmpty.html FALSE   8735  644
#> 168                        isOddNumber.html FALSE   8468  644
#> 169                         isTryError.html FALSE   9745  644
#> 170                        lastElement.html FALSE   8819  644
#> 171                               left.html FALSE   9248  644
#> 172                leftSubstringEquals.html FALSE   9835  644
#> 173                       limitToRange.html FALSE   9671  644
#> 174                  linearCombination.html FALSE  12012  644
#> 175                        listObjects.html FALSE  10003  644
#>                   mtime               ctime               atime uid gid  uname
#> 1   2022-06-08 05:58:33 2022-06-08 05:58:33 2022-06-08 05:58:33 501  20 runner
#> 2   2022-06-08 05:59:55 2022-06-08 05:59:55 2022-06-08 05:58:34 501  20 runner
#> 3   2022-06-08 05:58:45 2022-06-08 05:58:45 2022-06-08 05:58:37 501  20 runner
#> 4   2022-06-08 05:58:33 2022-06-08 05:58:33 2022-06-08 05:58:33 501  20 runner
#> 5   2022-06-08 05:58:34 2022-06-08 05:58:34 2022-06-08 05:58:34 501  20 runner
#> 6   2022-06-08 05:58:35 2022-06-08 05:58:35 2022-06-08 05:58:35 501  20 runner
#> 7   2022-06-08 05:58:35 2022-06-08 05:58:35 2022-06-08 05:58:35 501  20 runner
#> 8   2022-06-08 05:58:36 2022-06-08 05:58:36 2022-06-08 05:58:36 501  20 runner
#> 9   2022-06-08 05:58:36 2022-06-08 05:58:36 2022-06-08 05:58:36 501  20 runner
#> 10  2022-06-08 05:58:37 2022-06-08 05:58:37 2022-06-08 05:58:37 501  20 runner
#> 11  2022-06-08 05:58:37 2022-06-08 05:58:37 2022-06-08 05:58:37 501  20 runner
#> 12  2022-06-08 05:58:37 2022-06-08 05:58:37 2022-06-08 05:58:37 501  20 runner
#> 13  2022-06-08 05:58:37 2022-06-08 05:58:37 2022-06-08 05:58:37 501  20 runner
#> 14  2022-06-08 05:58:38 2022-06-08 05:58:38 2022-06-08 05:58:38 501  20 runner
#> 15  2022-06-08 05:58:38 2022-06-08 05:58:38 2022-06-08 05:58:38 501  20 runner
#> 16  2022-06-08 05:58:39 2022-06-08 05:58:39 2022-06-08 05:58:39 501  20 runner
#> 17  2022-06-08 05:58:40 2022-06-08 05:58:40 2022-06-08 05:58:40 501  20 runner
#> 18  2022-06-08 05:58:40 2022-06-08 05:58:40 2022-06-08 05:58:40 501  20 runner
#> 19  2022-06-08 05:58:41 2022-06-08 05:58:41 2022-06-08 05:58:41 501  20 runner
#> 20  2022-06-08 05:58:41 2022-06-08 05:58:41 2022-06-08 05:58:41 501  20 runner
#> 21  2022-06-08 05:58:42 2022-06-08 05:58:42 2022-06-08 05:58:42 501  20 runner
#> 22  2022-06-08 05:58:42 2022-06-08 05:58:42 2022-06-08 05:58:42 501  20 runner
#> 23  2022-06-08 05:58:43 2022-06-08 05:58:43 2022-06-08 05:58:43 501  20 runner
#> 24  2022-06-08 05:58:43 2022-06-08 05:58:43 2022-06-08 05:58:43 501  20 runner
#> 25  2022-06-08 05:58:44 2022-06-08 05:58:44 2022-06-08 05:58:44 501  20 runner
#> 26  2022-06-08 05:58:44 2022-06-08 05:58:44 2022-06-08 05:58:44 501  20 runner
#> 27  2022-06-08 05:58:45 2022-06-08 05:58:45 2022-06-08 05:58:45 501  20 runner
#> 28  2022-06-08 05:58:45 2022-06-08 05:58:45 2022-06-08 05:58:45 501  20 runner
#> 29  2022-06-08 05:58:45 2022-06-08 05:58:45 2022-06-08 05:58:45 501  20 runner
#> 30  2022-06-08 05:58:45 2022-06-08 05:58:45 2022-06-08 05:58:45 501  20 runner
#> 31  2022-06-08 05:58:46 2022-06-08 05:58:46 2022-06-08 05:58:46 501  20 runner
#> 32  2022-06-08 05:58:46 2022-06-08 05:58:46 2022-06-08 05:58:46 501  20 runner
#> 33  2022-06-08 05:58:47 2022-06-08 05:58:47 2022-06-08 05:58:47 501  20 runner
#> 34  2022-06-08 05:58:47 2022-06-08 05:58:47 2022-06-08 05:58:47 501  20 runner
#> 35  2022-06-08 05:58:48 2022-06-08 05:58:48 2022-06-08 05:58:48 501  20 runner
#> 36  2022-06-08 05:58:48 2022-06-08 05:58:48 2022-06-08 05:58:48 501  20 runner
#> 37  2022-06-08 05:58:49 2022-06-08 05:58:49 2022-06-08 05:58:49 501  20 runner
#> 38  2022-06-08 05:58:49 2022-06-08 05:58:49 2022-06-08 05:58:49 501  20 runner
#> 39  2022-06-08 05:58:50 2022-06-08 05:58:50 2022-06-08 05:58:50 501  20 runner
#> 40  2022-06-08 05:58:50 2022-06-08 05:58:50 2022-06-08 05:58:50 501  20 runner
#> 41  2022-06-08 05:58:51 2022-06-08 05:58:51 2022-06-08 05:58:51 501  20 runner
#> 42  2022-06-08 05:58:51 2022-06-08 05:58:51 2022-06-08 05:58:51 501  20 runner
#> 43  2022-06-08 05:58:51 2022-06-08 05:58:51 2022-06-08 05:58:51 501  20 runner
#> 44  2022-06-08 05:58:52 2022-06-08 05:58:52 2022-06-08 05:58:52 501  20 runner
#> 45  2022-06-08 05:58:52 2022-06-08 05:58:52 2022-06-08 05:58:52 501  20 runner
#> 46  2022-06-08 05:58:53 2022-06-08 05:58:53 2022-06-08 05:58:53 501  20 runner
#> 47  2022-06-08 05:58:53 2022-06-08 05:58:53 2022-06-08 05:58:53 501  20 runner
#> 48  2022-06-08 05:58:54 2022-06-08 05:58:54 2022-06-08 05:58:54 501  20 runner
#> 49  2022-06-08 05:58:54 2022-06-08 05:58:54 2022-06-08 05:58:54 501  20 runner
#> 50  2022-06-08 05:58:55 2022-06-08 05:58:55 2022-06-08 05:58:55 501  20 runner
#> 51  2022-06-08 05:58:55 2022-06-08 05:58:55 2022-06-08 05:58:55 501  20 runner
#> 52  2022-06-08 05:58:56 2022-06-08 05:58:56 2022-06-08 05:58:56 501  20 runner
#> 53  2022-06-08 05:58:56 2022-06-08 05:58:56 2022-06-08 05:58:56 501  20 runner
#> 54  2022-06-08 05:58:57 2022-06-08 05:58:57 2022-06-08 05:58:57 501  20 runner
#> 55  2022-06-08 05:58:57 2022-06-08 05:58:57 2022-06-08 05:58:57 501  20 runner
#> 56  2022-06-08 05:58:58 2022-06-08 05:58:58 2022-06-08 05:58:58 501  20 runner
#> 57  2022-06-08 05:58:59 2022-06-08 05:58:59 2022-06-08 05:58:59 501  20 runner
#> 58  2022-06-08 05:58:59 2022-06-08 05:58:59 2022-06-08 05:58:59 501  20 runner
#> 59  2022-06-08 05:59:00 2022-06-08 05:59:00 2022-06-08 05:59:00 501  20 runner
#> 60  2022-06-08 05:59:00 2022-06-08 05:59:00 2022-06-08 05:59:00 501  20 runner
#> 61  2022-06-08 05:59:01 2022-06-08 05:59:01 2022-06-08 05:59:01 501  20 runner
#> 62  2022-06-08 05:59:01 2022-06-08 05:59:01 2022-06-08 05:59:01 501  20 runner
#> 63  2022-06-08 05:59:02 2022-06-08 05:59:02 2022-06-08 05:59:02 501  20 runner
#> 64  2022-06-08 05:59:02 2022-06-08 05:59:02 2022-06-08 05:59:02 501  20 runner
#> 65  2022-06-08 05:59:03 2022-06-08 05:59:03 2022-06-08 05:59:03 501  20 runner
#> 66  2022-06-08 05:59:03 2022-06-08 05:59:03 2022-06-08 05:59:03 501  20 runner
#> 67  2022-06-08 05:59:04 2022-06-08 05:59:04 2022-06-08 05:59:04 501  20 runner
#> 68  2022-06-08 05:59:04 2022-06-08 05:59:04 2022-06-08 05:59:04 501  20 runner
#> 69  2022-06-08 05:59:05 2022-06-08 05:59:05 2022-06-08 05:59:05 501  20 runner
#> 70  2022-06-08 05:59:05 2022-06-08 05:59:05 2022-06-08 05:59:05 501  20 runner
#> 71  2022-06-08 05:59:06 2022-06-08 05:59:06 2022-06-08 05:59:06 501  20 runner
#> 72  2022-06-08 05:59:06 2022-06-08 05:59:06 2022-06-08 05:59:06 501  20 runner
#> 73  2022-06-08 05:59:07 2022-06-08 05:59:07 2022-06-08 05:59:07 501  20 runner
#> 74  2022-06-08 05:59:07 2022-06-08 05:59:07 2022-06-08 05:59:07 501  20 runner
#> 75  2022-06-08 05:59:08 2022-06-08 05:59:08 2022-06-08 05:59:08 501  20 runner
#> 76  2022-06-08 05:59:08 2022-06-08 05:59:08 2022-06-08 05:59:08 501  20 runner
#> 77  2022-06-08 05:59:08 2022-06-08 05:59:08 2022-06-08 05:59:08 501  20 runner
#> 78  2022-06-08 05:59:09 2022-06-08 05:59:09 2022-06-08 05:59:09 501  20 runner
#> 79  2022-06-08 05:59:09 2022-06-08 05:59:09 2022-06-08 05:59:09 501  20 runner
#> 80  2022-06-08 05:59:10 2022-06-08 05:59:10 2022-06-08 05:59:10 501  20 runner
#> 81  2022-06-08 05:59:10 2022-06-08 05:59:10 2022-06-08 05:59:10 501  20 runner
#> 82  2022-06-08 05:59:11 2022-06-08 05:59:11 2022-06-08 05:59:11 501  20 runner
#> 83  2022-06-08 05:59:11 2022-06-08 05:59:11 2022-06-08 05:59:11 501  20 runner
#> 84  2022-06-08 05:59:12 2022-06-08 05:59:12 2022-06-08 05:59:12 501  20 runner
#> 85  2022-06-08 05:59:12 2022-06-08 05:59:12 2022-06-08 05:59:12 501  20 runner
#> 86  2022-06-08 05:59:13 2022-06-08 05:59:13 2022-06-08 05:59:13 501  20 runner
#> 87  2022-06-08 05:59:13 2022-06-08 05:59:13 2022-06-08 05:59:13 501  20 runner
#> 88  2022-06-08 05:59:14 2022-06-08 05:59:14 2022-06-08 05:59:14 501  20 runner
#> 89  2022-06-08 05:59:15 2022-06-08 05:59:15 2022-06-08 05:59:15 501  20 runner
#> 90  2022-06-08 05:59:15 2022-06-08 05:59:15 2022-06-08 05:59:15 501  20 runner
#> 91  2022-06-08 05:59:16 2022-06-08 05:59:16 2022-06-08 05:59:16 501  20 runner
#> 92  2022-06-08 05:59:16 2022-06-08 05:59:16 2022-06-08 05:59:16 501  20 runner
#> 93  2022-06-08 05:59:17 2022-06-08 05:59:17 2022-06-08 05:59:17 501  20 runner
#> 94  2022-06-08 05:59:17 2022-06-08 05:59:17 2022-06-08 05:59:17 501  20 runner
#> 95  2022-06-08 05:59:18 2022-06-08 05:59:18 2022-06-08 05:59:18 501  20 runner
#> 96  2022-06-08 05:59:18 2022-06-08 05:59:18 2022-06-08 05:59:18 501  20 runner
#> 97  2022-06-08 05:59:19 2022-06-08 05:59:19 2022-06-08 05:59:19 501  20 runner
#> 98  2022-06-08 05:59:19 2022-06-08 05:59:19 2022-06-08 05:59:19 501  20 runner
#> 99  2022-06-08 05:59:20 2022-06-08 05:59:20 2022-06-08 05:59:20 501  20 runner
#> 100 2022-06-08 05:59:20 2022-06-08 05:59:20 2022-06-08 05:59:20 501  20 runner
#> 101 2022-06-08 05:59:21 2022-06-08 05:59:21 2022-06-08 05:59:21 501  20 runner
#> 102 2022-06-08 05:59:21 2022-06-08 05:59:21 2022-06-08 05:59:21 501  20 runner
#> 103 2022-06-08 05:59:22 2022-06-08 05:59:22 2022-06-08 05:59:22 501  20 runner
#> 104 2022-06-08 05:59:22 2022-06-08 05:59:22 2022-06-08 05:59:22 501  20 runner
#> 105 2022-06-08 05:59:23 2022-06-08 05:59:23 2022-06-08 05:59:23 501  20 runner
#> 106 2022-06-08 05:59:23 2022-06-08 05:59:23 2022-06-08 05:59:23 501  20 runner
#> 107 2022-06-08 05:59:24 2022-06-08 05:59:24 2022-06-08 05:59:24 501  20 runner
#> 108 2022-06-08 05:59:24 2022-06-08 05:59:24 2022-06-08 05:59:24 501  20 runner
#> 109 2022-06-08 05:59:25 2022-06-08 05:59:25 2022-06-08 05:59:25 501  20 runner
#> 110 2022-06-08 05:59:25 2022-06-08 05:59:25 2022-06-08 05:59:25 501  20 runner
#> 111 2022-06-08 05:59:26 2022-06-08 05:59:26 2022-06-08 05:59:26 501  20 runner
#> 112 2022-06-08 05:59:26 2022-06-08 05:59:26 2022-06-08 05:59:26 501  20 runner
#> 113 2022-06-08 05:59:27 2022-06-08 05:59:27 2022-06-08 05:59:27 501  20 runner
#> 114 2022-06-08 05:59:27 2022-06-08 05:59:27 2022-06-08 05:59:27 501  20 runner
#> 115 2022-06-08 05:59:28 2022-06-08 05:59:28 2022-06-08 05:59:28 501  20 runner
#> 116 2022-06-08 05:59:28 2022-06-08 05:59:28 2022-06-08 05:59:28 501  20 runner
#> 117 2022-06-08 05:59:29 2022-06-08 05:59:29 2022-06-08 05:59:29 501  20 runner
#> 118 2022-06-08 05:59:29 2022-06-08 05:59:29 2022-06-08 05:59:29 501  20 runner
#> 119 2022-06-08 05:59:30 2022-06-08 05:59:30 2022-06-08 05:59:30 501  20 runner
#> 120 2022-06-08 05:59:30 2022-06-08 05:59:30 2022-06-08 05:59:30 501  20 runner
#> 121 2022-06-08 05:59:31 2022-06-08 05:59:31 2022-06-08 05:59:31 501  20 runner
#> 122 2022-06-08 05:59:31 2022-06-08 05:59:31 2022-06-08 05:59:31 501  20 runner
#> 123 2022-06-08 05:59:32 2022-06-08 05:59:32 2022-06-08 05:59:32 501  20 runner
#> 124 2022-06-08 05:59:32 2022-06-08 05:59:32 2022-06-08 05:59:32 501  20 runner
#> 125 2022-06-08 05:59:32 2022-06-08 05:59:32 2022-06-08 05:59:32 501  20 runner
#> 126 2022-06-08 05:59:33 2022-06-08 05:59:33 2022-06-08 05:59:33 501  20 runner
#> 127 2022-06-08 05:59:33 2022-06-08 05:59:33 2022-06-08 05:59:33 501  20 runner
#> 128 2022-06-08 05:59:34 2022-06-08 05:59:34 2022-06-08 05:59:34 501  20 runner
#> 129 2022-06-08 05:59:34 2022-06-08 05:59:34 2022-06-08 05:59:34 501  20 runner
#> 130 2022-06-08 05:59:35 2022-06-08 05:59:35 2022-06-08 05:59:35 501  20 runner
#> 131 2022-06-08 05:59:35 2022-06-08 05:59:35 2022-06-08 05:59:35 501  20 runner
#> 132 2022-06-08 05:59:36 2022-06-08 05:59:36 2022-06-08 05:59:36 501  20 runner
#> 133 2022-06-08 05:59:36 2022-06-08 05:59:36 2022-06-08 05:59:36 501  20 runner
#> 134 2022-06-08 05:59:37 2022-06-08 05:59:37 2022-06-08 05:59:37 501  20 runner
#> 135 2022-06-08 05:59:37 2022-06-08 05:59:37 2022-06-08 05:59:37 501  20 runner
#> 136 2022-06-08 05:59:38 2022-06-08 05:59:38 2022-06-08 05:59:37 501  20 runner
#> 137 2022-06-08 05:59:38 2022-06-08 05:59:38 2022-06-08 05:59:38 501  20 runner
#> 138 2022-06-08 05:59:38 2022-06-08 05:59:38 2022-06-08 05:59:38 501  20 runner
#> 139 2022-06-08 05:59:39 2022-06-08 05:59:39 2022-06-08 05:59:39 501  20 runner
#> 140 2022-06-08 05:59:39 2022-06-08 05:59:39 2022-06-08 05:59:39 501  20 runner
#> 141 2022-06-08 05:59:40 2022-06-08 05:59:40 2022-06-08 05:59:40 501  20 runner
#> 142 2022-06-08 05:59:40 2022-06-08 05:59:40 2022-06-08 05:59:40 501  20 runner
#> 143 2022-06-08 05:59:41 2022-06-08 05:59:41 2022-06-08 05:59:41 501  20 runner
#> 144 2022-06-08 05:59:41 2022-06-08 05:59:41 2022-06-08 05:59:41 501  20 runner
#> 145 2022-06-08 05:59:42 2022-06-08 05:59:42 2022-06-08 05:59:42 501  20 runner
#> 146 2022-06-08 05:59:43 2022-06-08 05:59:43 2022-06-08 05:59:43 501  20 runner
#> 147 2022-06-08 05:59:43 2022-06-08 05:59:43 2022-06-08 05:59:43 501  20 runner
#> 148 2022-06-08 05:59:43 2022-06-08 05:59:43 2022-06-08 05:59:43 501  20 runner
#> 149 2022-06-08 05:59:44 2022-06-08 05:59:44 2022-06-08 05:59:44 501  20 runner
#> 150 2022-06-08 05:59:44 2022-06-08 05:59:44 2022-06-08 05:59:44 501  20 runner
#> 151 2022-06-08 05:59:45 2022-06-08 05:59:45 2022-06-08 05:59:45 501  20 runner
#> 152 2022-06-08 05:59:45 2022-06-08 05:59:45 2022-06-08 05:59:45 501  20 runner
#> 153 2022-06-08 05:59:46 2022-06-08 05:59:46 2022-06-08 05:59:46 501  20 runner
#> 154 2022-06-08 05:59:46 2022-06-08 05:59:46 2022-06-08 05:59:46 501  20 runner
#> 155 2022-06-08 05:58:32 2022-06-08 05:58:32 2022-06-08 05:58:32 501  20 runner
#> 156 2022-06-08 05:59:47 2022-06-08 05:59:47 2022-06-08 05:59:47 501  20 runner
#> 157 2022-06-08 05:59:47 2022-06-08 05:59:47 2022-06-08 05:59:47 501  20 runner
#> 158 2022-06-08 05:59:48 2022-06-08 05:59:48 2022-06-08 05:59:48 501  20 runner
#> 159 2022-06-08 05:59:48 2022-06-08 05:59:48 2022-06-08 05:59:48 501  20 runner
#> 160 2022-06-08 05:59:49 2022-06-08 05:59:49 2022-06-08 05:59:49 501  20 runner
#> 161 2022-06-08 05:59:49 2022-06-08 05:59:49 2022-06-08 05:59:49 501  20 runner
#> 162 2022-06-08 05:59:50 2022-06-08 05:59:50 2022-06-08 05:59:50 501  20 runner
#> 163 2022-06-08 05:59:50 2022-06-08 05:59:50 2022-06-08 05:59:50 501  20 runner
#> 164 2022-06-08 05:59:51 2022-06-08 05:59:51 2022-06-08 05:59:51 501  20 runner
#> 165 2022-06-08 05:59:51 2022-06-08 05:59:51 2022-06-08 05:59:51 501  20 runner
#> 166 2022-06-08 05:59:52 2022-06-08 05:59:52 2022-06-08 05:59:52 501  20 runner
#> 167 2022-06-08 05:59:52 2022-06-08 05:59:52 2022-06-08 05:59:52 501  20 runner
#> 168 2022-06-08 05:59:52 2022-06-08 05:59:52 2022-06-08 05:59:52 501  20 runner
#> 169 2022-06-08 05:59:53 2022-06-08 05:59:53 2022-06-08 05:59:53 501  20 runner
#> 170 2022-06-08 05:59:53 2022-06-08 05:59:53 2022-06-08 05:59:53 501  20 runner
#> 171 2022-06-08 05:59:54 2022-06-08 05:59:54 2022-06-08 05:59:54 501  20 runner
#> 172 2022-06-08 05:59:54 2022-06-08 05:59:54 2022-06-08 05:59:54 501  20 runner
#> 173 2022-06-08 05:59:55 2022-06-08 05:59:55 2022-06-08 05:59:55 501  20 runner
#> 174 2022-06-08 05:59:55 2022-06-08 05:59:55 2022-06-08 05:59:55 501  20 runner
#> 175 2022-06-08 05:59:56 2022-06-08 05:59:56 2022-06-08 05:59:56 501  20 runner
#>     grname
#> 1    staff
#> 2    staff
#> 3    staff
#> 4    staff
#> 5    staff
#> 6    staff
#> 7    staff
#> 8    staff
#> 9    staff
#> 10   staff
#> 11   staff
#> 12   staff
#> 13   staff
#> 14   staff
#> 15   staff
#> 16   staff
#> 17   staff
#> 18   staff
#> 19   staff
#> 20   staff
#> 21   staff
#> 22   staff
#> 23   staff
#> 24   staff
#> 25   staff
#> 26   staff
#> 27   staff
#> 28   staff
#> 29   staff
#> 30   staff
#> 31   staff
#> 32   staff
#> 33   staff
#> 34   staff
#> 35   staff
#> 36   staff
#> 37   staff
#> 38   staff
#> 39   staff
#> 40   staff
#> 41   staff
#> 42   staff
#> 43   staff
#> 44   staff
#> 45   staff
#> 46   staff
#> 47   staff
#> 48   staff
#> 49   staff
#> 50   staff
#> 51   staff
#> 52   staff
#> 53   staff
#> 54   staff
#> 55   staff
#> 56   staff
#> 57   staff
#> 58   staff
#> 59   staff
#> 60   staff
#> 61   staff
#> 62   staff
#> 63   staff
#> 64   staff
#> 65   staff
#> 66   staff
#> 67   staff
#> 68   staff
#> 69   staff
#> 70   staff
#> 71   staff
#> 72   staff
#> 73   staff
#> 74   staff
#> 75   staff
#> 76   staff
#> 77   staff
#> 78   staff
#> 79   staff
#> 80   staff
#> 81   staff
#> 82   staff
#> 83   staff
#> 84   staff
#> 85   staff
#> 86   staff
#> 87   staff
#> 88   staff
#> 89   staff
#> 90   staff
#> 91   staff
#> 92   staff
#> 93   staff
#> 94   staff
#> 95   staff
#> 96   staff
#> 97   staff
#> 98   staff
#> 99   staff
#> 100  staff
#> 101  staff
#> 102  staff
#> 103  staff
#> 104  staff
#> 105  staff
#> 106  staff
#> 107  staff
#> 108  staff
#> 109  staff
#> 110  staff
#> 111  staff
#> 112  staff
#> 113  staff
#> 114  staff
#> 115  staff
#> 116  staff
#> 117  staff
#> 118  staff
#> 119  staff
#> 120  staff
#> 121  staff
#> 122  staff
#> 123  staff
#> 124  staff
#> 125  staff
#> 126  staff
#> 127  staff
#> 128  staff
#> 129  staff
#> 130  staff
#> 131  staff
#> 132  staff
#> 133  staff
#> 134  staff
#> 135  staff
#> 136  staff
#> 137  staff
#> 138  staff
#> 139  staff
#> 140  staff
#> 141  staff
#> 142  staff
#> 143  staff
#> 144  staff
#> 145  staff
#> 146  staff
#> 147  staff
#> 148  staff
#> 149  staff
#> 150  staff
#> 151  staff
#> 152  staff
#> 153  staff
#> 154  staff
#> 155  staff
#> 156  staff
#> 157  staff
#> 158  staff
#> 159  staff
#> 160  staff
#> 161  staff
#> 162  staff
#> 163  staff
#> 164  staff
#> 165  staff
#> 166  staff
#> 167  staff
#> 168  staff
#> 169  staff
#> 170  staff
#> 171  staff
#> 172  staff
#> 173  staff
#> 174  staff
#> 175  staff

# ... even when being called with an empty character vector
FUN(character())
#> listing 
#> listing .
#> [1] file  isdir
#> <0 rows> (or 0-length row.names)
FUN(character(), full_info = TRUE)
#> listing 
#> listing .
#>  [1] file   isdir  size   mode   mtime  ctime  atime  uid    gid    uname 
#> [11] grname
#> <0 rows> (or 0-length row.names)

# Call the function recursively up to the given depth level
kwb.utils:::listToDepth(".", max_depth = 1, FUN = FUN)
#> listing .
#> listing .
#>                                        file isdir
#> 1                               DIN.A4.html FALSE
#> 2                              Rplot001.png FALSE
#> 3                              Rplot002.png FALSE
#> 4                             addClass.html FALSE
#> 5                       addRowWithName.html FALSE
#> 6                   addSuffixToColumns.html FALSE
#> 7                          allAreEqual.html FALSE
#> 8                      allAreIdentical.html FALSE
#> 9                          almostEqual.html FALSE
#> 10                        appendSuffix.html FALSE
#> 11                            arglist-1.png FALSE
#> 12                            arglist-2.png FALSE
#> 13                             arglist.html FALSE
#> 14                             argsCsv.html FALSE
#> 15                    arrayToDataFrame.html FALSE
#> 16                        asColumnList.html FALSE
#> 17                 asNoFactorDataFrame.html FALSE
#> 18                           asRowList.html FALSE
#> 19                    assertFinalSlash.html FALSE
#> 20                assertRowsAndColumns.html FALSE
#> 21                           assignAll.html FALSE
#> 22              assignArgumentDefaults.html FALSE
#> 23                      assignGlobally.html FALSE
#> 24                assignPackageObjects.html FALSE
#> 25                     atLeastOneRowIn.html FALSE
#> 26                           backspace.html FALSE
#> 27                     breakInSequence.html FALSE
#> 28                           callWith-1.png FALSE
#> 29                           callWith-2.png FALSE
#> 30                            callWith.html FALSE
#> 31            callWithStringsAsFactors.html FALSE
#> 32                           catAndRun.html FALSE
#> 33                          catChanges.html FALSE
#> 34                        catChangesIf.html FALSE
#> 35                               catIf.html FALSE
#> 36                            catLines.html FALSE
#> 37                        catNewLineIf.html FALSE
#> 38              checkForMissingColumns.html FALSE
#> 39                        clearConsole.html FALSE
#> 40                          clipMatrix.html FALSE
#> 41                         cmdLinePath.html FALSE
#> 42                           colMaxima.html FALSE
#> 43                           colMinima.html FALSE
#> 44                        colNaNumbers.html FALSE
#> 45             colStatisticOneFunction.html FALSE
#> 46                       colStatistics.html FALSE
#> 47                           collapsed.html FALSE
#> 48                    columnDescriptor.html FALSE
#> 49                        columnToDate.html FALSE
#> 50                columnwisePercentage.html FALSE
#> 51                combineAlternatingly.html FALSE
#> 52                      commaCollapsed.html FALSE
#> 53                         commonNames.html FALSE
#> 54                   compareDataFrames.html FALSE
#> 55                         compareSets.html FALSE
#> 56                   containsNulString.html FALSE
#> 57                      convertCsvFile.html FALSE
#> 58                      copyAttributes.html FALSE
#> 59              copyDirectoryStructure.html FALSE
#> 60                    copyListElements.html FALSE
#> 61                     countNaInColumn.html FALSE
#> 62                          countOrSum.html FALSE
#> 63                      createAccessor.html FALSE
#> 64              createDirAndReturnPath.html FALSE
#> 65                   createDirectories.html FALSE
#> 66                     createDirectory.html FALSE
#> 67           createFunctionExtdataFile.html FALSE
#> 68                       createIdAlong.html FALSE
#> 69                        createMatrix.html FALSE
#> 70                  createPasswordFile.html FALSE
#> 71                       createStorage.html FALSE
#> 72                  csvTextToDataFrame.html FALSE
#> 73                              decode.html FALSE
#> 74                           defaultIf.html FALSE
#> 75                         defaultIfNA.html FALSE
#> 76                       defaultIfNULL.html FALSE
#> 77                       defaultIfZero.html FALSE
#> 78                       defaultLevels.html FALSE
#> 79        defaultWindowsProgramFolders.html FALSE
#> 80                             desktop.html FALSE
#> 81                            diffrows.html FALSE
#> 82                       directoryName.html FALSE
#> 83                          dot-OStype.html FALSE
#> 84                             dot-log.html FALSE
#> 85                         dot-logline.html FALSE
#> 86                           dot-logok.html FALSE
#> 87                        dot-logstart.html FALSE
#> 88                             dropDim.html FALSE
#> 89              dropUnusedFactorLevels.html FALSE
#> 90                              encode.html FALSE
#> 91                       enlargeVector.html FALSE
#> 92                         excludeNULL.html FALSE
#> 93                          expandGrid.html FALSE
#> 94                         extdataFile.html FALSE
#> 95                        extendLimits.html FALSE
#> 96                    extractRowRanges.html FALSE
#> 97                    extractSubstring.html FALSE
#> 98                       fileExtension.html FALSE
#> 99                         findChanges.html FALSE
#> 100              findPartialDuplicates.html FALSE
#> 101                   finishAndShowPdf.html FALSE
#> 102                 finishAndShowPdfIf.html FALSE
#> 103                       firstElement.html FALSE
#> 104                   firstPosixColumn.html FALSE
#> 105                      frenchToAscii.html FALSE
#> 106                     frequencyTable.html FALSE
#> 107                        fullWinPath.html FALSE
#> 108                        fullySorted.html FALSE
#> 109                    generateKeyFile.html FALSE
#> 110                       getAttribute.html FALSE
#> 111       getByPositiveOrNegativeIndex.html FALSE
#> 112                  getElementLengths.html FALSE
#> 113                     getEvenNumbers.html FALSE
#> 114          getFunctionValueOrDefault.html FALSE
#> 115                        getGlobally.html FALSE
#> 116                getKeywordPositions.html FALSE
#> 117                        getListNode.html FALSE
#> 118      getNamesOfObjectsInRDataFiles.html FALSE
#> 119             getObjectFromRDataFile.html FALSE
#> 120                      getOddNumbers.html FALSE
#> 121                        getPassword.html FALSE
#> 122 getPathsAndValuesFromRecursiveList.html FALSE
#> 123                        getTagNames.html FALSE
#> 124                        get_homedir.html FALSE
#> 125                   guessSeparator.1.html FALSE
#> 126                   guessSeparator.2.html FALSE
#> 127                     guessSeparator.html FALSE
#> 128                      hasFinalSlash.html FALSE
#> 129                      hasZeroLength.html FALSE
#> 130                           headtail.html FALSE
#> 131                   hsAddMissingCols.html FALSE
#> 132                         hsChrToNum.html FALSE
#> 133                       hsCountInStr.html FALSE
#> 134                     hsDelEmptyCols.html FALSE
#> 135                 hsMatrixToListForm.html FALSE
#> 136                      hsMovingMean-1.png FALSE
#> 137                       hsMovingMean.html FALSE
#> 138              hsOpenWindowsExplorer.html FALSE
#> 139                          hsPrepPdf.html FALSE
#> 140                         hsQuoteChr.html FALSE
#> 141                    hsRenameColumns.html FALSE
#> 142                          hsResolve.html FALSE
#> 143                hsRestoreAttributes.html FALSE
#> 144                         hsSafeName.html FALSE
#> 145                            hsShell.html FALSE
#> 146                          hsShowPdf.html FALSE
#> 147                     hsStringToDate.html FALSE
#> 148                   hsStringToDouble.html FALSE
#> 149                   hsSubstSpecChars.html FALSE
#> 150                           hsSystem.html FALSE
#> 151                             hsTrim.html FALSE
#> 152                       hsValidValue.html FALSE
#> 153                            inRange.html FALSE
#> 154                             indent.html FALSE
#> 155                              index.html FALSE
#> 156                      insertColumns.html FALSE
#> 157                 intToNumeralSystem.html FALSE
#> 158                         is.unnamed.html FALSE
#> 159                            isASCII.html FALSE
#> 160                   isDotOrDoubleDot.html FALSE
#> 161                       isEvenNumber.html FALSE
#> 162                           isLoaded.html FALSE
#> 163                   isNaInAllColumns.html FALSE
#> 164                      isNaInAllRows.html FALSE
#> 165                        isNaOrEmpty.html FALSE
#> 166                      isNetworkPath.html FALSE
#> 167                      isNullOrEmpty.html FALSE
#> 168                        isOddNumber.html FALSE
#> 169                         isTryError.html FALSE
#> 170                        lastElement.html FALSE
#> 171                               left.html FALSE
#> 172                leftSubstringEquals.html FALSE
#> 173                       limitToRange.html FALSE
#> 174                  linearCombination.html FALSE
#> 175                        listObjects.html FALSE