List Elements Recursively up to Depth

List Elements Recursively up to Depth

listToDepth(
  path,
  max_depth = 0L,
  full_info = FALSE,
  FUN = listFiles,
  ...,
  depth = 0,
  prob_mutate = 0,
  template = NULL
)

Arguments

path

path to the element at which to start listing

max_depth

maximal depth level of which to list elements. A value of 0 means non-recursive listing, a value of NA represents fully recursive listing.

full_info

return only path and isdir information or the full information provided by FUN(full_info = TRUE)?

FUN

function called to get the listing of the element given in path. The function must accept a path as its first argument and it must define the argument full_info second. It may accept further arguments. It must always return a data frame. For full_info = FALSE the data frame must have columns file and isdir (is the "file" a directory?). For full_info = TRUE the function may return further columns. The function must provide an empty data frame with the expected columns when being called with character() as the first argument. The function is expected to set the attribute "failed" to the given path in case that the path could not be accessed (e.g. because of a broken internet connection if the listing is done remotely). See kwb.utils:::listFiles for an example implementation that somehow simulates the behaviour of the dir function. See kwb.dwd::list_url() for a more advanced usage of this function in order to recursively list the files on an FTP server (FTP = file transfer protocol).

...

further arguments passed to FUN

depth

start depth of recursion if max_depth > 0. This argument is for internal use and not intended to be set by the user!

prob_mutate

probability to alter the path so that it becomes useless. This is zero by default. Set the value only if you want to test how the function behaves if the listing of a path fails.

template

empty data frame (zero rows) and with columns that are identical to the columns in the data frame returned by FUN. If not provided the function will call FUN once with a zero-length path argument expecting to retrieve the template that is expected here.

Value

data frame containing at least the columns file and isdir. If full_info = TRUE the result data frame may contain further columns, as provided by the function given in FUN for full_info = TRUE.

Examples

# Example list function provided in this package (file listing)
FUN <- kwb.utils:::listFiles

# The list function must provide empty records when no path is given. The 
# returned data frame must provide the columns "file" and "isdir" ...
FUN()
#> listing .
#>                                        file isdir
#> 1                               DIN.A4.html FALSE
#> 2                              Rplot001.png FALSE
#> 3                              Rplot002.png FALSE
#> 4                             addClass.html FALSE
#> 5                       addRowWithName.html FALSE
#> 6                   addSuffixToColumns.html FALSE
#> 7                          allAreEqual.html FALSE
#> 8                      allAreIdentical.html FALSE
#> 9                          almostEqual.html FALSE
#> 10                        appendSuffix.html FALSE
#> 11                            arglist-1.png FALSE
#> 12                            arglist-2.png FALSE
#> 13                             arglist.html FALSE
#> 14                             argsCsv.html FALSE
#> 15                        asColumnList.html FALSE
#> 16                 asNoFactorDataFrame.html FALSE
#> 17                           asRowList.html FALSE
#> 18                    assertFinalSlash.html FALSE
#> 19                assertRowsAndColumns.html FALSE
#> 20                           assignAll.html FALSE
#> 21              assignArgumentDefaults.html FALSE
#> 22                      assignGlobally.html FALSE
#> 23                assignPackageObjects.html FALSE
#> 24                     atLeastOneRowIn.html FALSE
#> 25                           backspace.html FALSE
#> 26                     breakInSequence.html FALSE
#> 27                           callWith-1.png FALSE
#> 28                           callWith-2.png FALSE
#> 29                            callWith.html FALSE
#> 30            callWithStringsAsFactors.html FALSE
#> 31                           catAndRun.html FALSE
#> 32                               catIf.html FALSE
#> 33                            catLines.html FALSE
#> 34                        catNewLineIf.html FALSE
#> 35              checkForMissingColumns.html FALSE
#> 36                        clearConsole.html FALSE
#> 37                          clipMatrix.html FALSE
#> 38                         cmdLinePath.html FALSE
#> 39                           colMaxima.html FALSE
#> 40                           colMinima.html FALSE
#> 41                        colNaNumbers.html FALSE
#> 42             colStatisticOneFunction.html FALSE
#> 43                       colStatistics.html FALSE
#> 44                           collapsed.html FALSE
#> 45                    columnDescriptor.html FALSE
#> 46                        columnToDate.html FALSE
#> 47                columnwisePercentage.html FALSE
#> 48                combineAlternatingly.html FALSE
#> 49                      commaCollapsed.html FALSE
#> 50                   compareDataFrames.html FALSE
#> 51                         compareSets.html FALSE
#> 52                   containsNulString.html FALSE
#> 53                      convertCsvFile.html FALSE
#> 54              copyDirectoryStructure.html FALSE
#> 55                    copyListElements.html FALSE
#> 56                     countNaInColumn.html FALSE
#> 57                          countOrSum.html FALSE
#> 58                      createAccessor.html FALSE
#> 59              createDirAndReturnPath.html FALSE
#> 60                     createDirectory.html FALSE
#> 61                       createIdAlong.html FALSE
#> 62                        createMatrix.html FALSE
#> 63                  createPasswordFile.html FALSE
#> 64                       createStorage.html FALSE
#> 65                  csvTextToDataFrame.html FALSE
#> 66                              decode.html FALSE
#> 67                         defaultIfNA.html FALSE
#> 68                       defaultIfNULL.html FALSE
#> 69                       defaultIfZero.html FALSE
#> 70                       defaultLevels.html FALSE
#> 71        defaultWindowsProgramFolders.html FALSE
#> 72                             desktop.html FALSE
#> 73                            diffrows.html FALSE
#> 74                          dot-OStype.html FALSE
#> 75                       dot-defaultIf.html FALSE
#> 76                dot-guessSeparator.1.html FALSE
#> 77                dot-guessSeparator.2.html FALSE
#> 78                   dot-isNetworkPath.html FALSE
#> 79                             dot-log.html FALSE
#> 80                         dot-logline.html FALSE
#> 81                           dot-logok.html FALSE
#> 82                        dot-logstart.html FALSE
#> 83                             dropDim.html FALSE
#> 84              dropUnusedFactorLevels.html FALSE
#> 85                              encode.html FALSE
#> 86                       enlargeVector.html FALSE
#> 87                         excludeNULL.html FALSE
#> 88                          expandGrid.html FALSE
#> 89                        extendLimits.html FALSE
#> 90                    extractRowRanges.html FALSE
#> 91                    extractSubstring.html FALSE
#> 92                       fileExtension.html FALSE
#> 93                         findChanges.html FALSE
#> 94               findPartialDuplicates.html FALSE
#> 95                    finishAndShowPdf.html FALSE
#> 96                  finishAndShowPdfIf.html FALSE
#> 97                        firstElement.html FALSE
#> 98                    firstPosixColumn.html FALSE
#> 99                       frenchToAscii.html FALSE
#> 100                     frequencyTable.html FALSE
#> 101                        fullySorted.html FALSE
#> 102                    generateKeyFile.html FALSE
#> 103                       getAttribute.html FALSE
#> 104       getByPositiveOrNegativeIndex.html FALSE
#> 105                  getElementLengths.html FALSE
#> 106                     getEvenNumbers.html FALSE
#> 107                    getFunctionName.html FALSE
#> 108          getFunctionValueOrDefault.html FALSE
#> 109                        getGlobally.html FALSE
#> 110                getKeywordPositions.html FALSE
#> 111      getNamesOfObjectsInRDataFiles.html FALSE
#> 112             getObjectFromRDataFile.html FALSE
#> 113                      getOddNumbers.html FALSE
#> 114                        getPassword.html FALSE
#> 115 getPathsAndValuesFromRecursiveList.html FALSE
#> 116                        getTagNames.html FALSE
#> 117                        get_homedir.html FALSE
#> 118                     guessSeparator.html FALSE
#> 119                           headtail.html FALSE
#> 120                   hsAddMissingCols.html FALSE
#> 121                         hsChrToNum.html FALSE
#> 122                       hsCountInStr.html FALSE
#> 123                     hsDelEmptyCols.html FALSE
#> 124                 hsMatrixToListForm.html FALSE
#> 125                      hsMovingMean-1.png FALSE
#> 126                       hsMovingMean.html FALSE
#> 127              hsOpenWindowsExplorer.html FALSE
#> 128                          hsPrepPdf.html FALSE
#> 129                         hsQuoteChr.html FALSE
#> 130                    hsRenameColumns.html FALSE
#> 131                          hsResolve.html FALSE
#> 132                hsRestoreAttributes.html FALSE
#> 133                         hsSafeName.html FALSE
#> 134                            hsShell.html FALSE
#> 135                          hsShowPdf.html FALSE
#> 136                     hsStringToDate.html FALSE
#> 137                   hsStringToDouble.html FALSE
#> 138                   hsSubstSpecChars.html FALSE
#> 139                           hsSystem.html FALSE
#> 140                             hsTrim.html FALSE
#> 141                       hsValidValue.html FALSE
#> 142                            inRange.html FALSE
#> 143                              index.html FALSE
#> 144                      insertColumns.html FALSE
#> 145                 intToNumeralSystem.html FALSE
#> 146                         is.unnamed.html FALSE
#> 147                            isASCII.html FALSE
#> 148                   isDotOrDoubleDot.html FALSE
#> 149                       isEvenNumber.html FALSE
#> 150                           isLoaded.html FALSE
#> 151                   isNaInAllColumns.html FALSE
#> 152                      isNaInAllRows.html FALSE
#> 153                        isNaOrEmpty.html FALSE
#> 154                      isNullOrEmpty.html FALSE
#> 155                        isOddNumber.html FALSE
#> 156                        lastElement.html FALSE
#> 157                               left.html FALSE
#> 158                       limitToRange.html FALSE
#> 159                  linearCombination.html FALSE
#> 160                        listObjects.html FALSE
FUN(full_info = TRUE)
#> listing .
#>                                        file isdir   size mode
#> 1                               DIN.A4.html FALSE   8776  644
#> 2                              Rplot001.png FALSE   1011  644
#> 3                              Rplot002.png FALSE   5908  644
#> 4                             addClass.html FALSE   9660  644
#> 5                       addRowWithName.html FALSE   9366  644
#> 6                   addSuffixToColumns.html FALSE  11904  644
#> 7                          allAreEqual.html FALSE   9348  644
#> 8                      allAreIdentical.html FALSE   9166  644
#> 9                          almostEqual.html FALSE   9348  644
#> 10                        appendSuffix.html FALSE  10841  644
#> 11                            arglist-1.png FALSE  18498  644
#> 12                            arglist-2.png FALSE  18423  644
#> 13                             arglist.html FALSE  12663  644
#> 14                             argsCsv.html FALSE   9206  644
#> 15                        asColumnList.html FALSE  11099  644
#> 16                 asNoFactorDataFrame.html FALSE  11473  644
#> 17                           asRowList.html FALSE  10986  644
#> 18                    assertFinalSlash.html FALSE  12090  644
#> 19                assertRowsAndColumns.html FALSE  12525  644
#> 20                           assignAll.html FALSE   9233  644
#> 21              assignArgumentDefaults.html FALSE  10470  644
#> 22                      assignGlobally.html FALSE   8957  644
#> 23                assignPackageObjects.html FALSE   9393  644
#> 24                     atLeastOneRowIn.html FALSE   9003  644
#> 25                           backspace.html FALSE  10576  644
#> 26                     breakInSequence.html FALSE   9320  644
#> 27                           callWith-1.png FALSE  18498  644
#> 28                           callWith-2.png FALSE  18423  644
#> 29                            callWith.html FALSE  11917  644
#> 30            callWithStringsAsFactors.html FALSE  14411  644
#> 31                           catAndRun.html FALSE  12817  644
#> 32                               catIf.html FALSE   8991  644
#> 33                            catLines.html FALSE   9047  644
#> 34                        catNewLineIf.html FALSE   9233  644
#> 35              checkForMissingColumns.html FALSE  10021  644
#> 36                        clearConsole.html FALSE   8723  644
#> 37                          clipMatrix.html FALSE   9179  644
#> 38                         cmdLinePath.html FALSE   9028  644
#> 39                           colMaxima.html FALSE   9070  644
#> 40                           colMinima.html FALSE   9070  644
#> 41                        colNaNumbers.html FALSE   8995  644
#> 42             colStatisticOneFunction.html FALSE   9473  644
#> 43                       colStatistics.html FALSE  10023  644
#> 44                           collapsed.html FALSE   9246  644
#> 45                    columnDescriptor.html FALSE   9196  644
#> 46                        columnToDate.html FALSE  11448  644
#> 47                columnwisePercentage.html FALSE  14785  644
#> 48                combineAlternatingly.html FALSE  12102  644
#> 49                      commaCollapsed.html FALSE   8902  644
#> 50                   compareDataFrames.html FALSE  13457  644
#> 51                         compareSets.html FALSE  11764  644
#> 52                   containsNulString.html FALSE   9125  644
#> 53                      convertCsvFile.html FALSE  21763  644
#> 54              copyDirectoryStructure.html FALSE   9901  644
#> 55                    copyListElements.html FALSE  12017  644
#> 56                     countNaInColumn.html FALSE   9179  644
#> 57                          countOrSum.html FALSE  13819  644
#> 58                      createAccessor.html FALSE  10875  644
#> 59              createDirAndReturnPath.html FALSE   9643  644
#> 60                     createDirectory.html FALSE   9553  644
#> 61                       createIdAlong.html FALSE  11980  644
#> 62                        createMatrix.html FALSE  13792  644
#> 63                  createPasswordFile.html FALSE   9976  644
#> 64                       createStorage.html FALSE  14467  644
#> 65                  csvTextToDataFrame.html FALSE   9136  644
#> 66                              decode.html FALSE   9117  644
#> 67                         defaultIfNA.html FALSE  10475  644
#> 68                       defaultIfNULL.html FALSE  10496  644
#> 69                       defaultIfZero.html FALSE  11745  644
#> 70                       defaultLevels.html FALSE  10627  644
#> 71        defaultWindowsProgramFolders.html FALSE   8855  644
#> 72                             desktop.html FALSE   9180  644
#> 73                            diffrows.html FALSE  10906  644
#> 74                          dot-OStype.html FALSE   8982  644
#> 75                       dot-defaultIf.html FALSE   9599  644
#> 76                dot-guessSeparator.1.html FALSE   9295  644
#> 77                dot-guessSeparator.2.html FALSE   9171  644
#> 78                   dot-isNetworkPath.html FALSE   9405  644
#> 79                             dot-log.html FALSE   8994  644
#> 80                         dot-logline.html FALSE   9125  644
#> 81                           dot-logok.html FALSE   9068  644
#> 82                        dot-logstart.html FALSE   9256  644
#> 83                             dropDim.html FALSE  13849  644
#> 84              dropUnusedFactorLevels.html FALSE  13880  644
#> 85                              encode.html FALSE  11282  644
#> 86                       enlargeVector.html FALSE  11571  644
#> 87                         excludeNULL.html FALSE  11120  644
#> 88                          expandGrid.html FALSE  14634  644
#> 89                        extendLimits.html FALSE   9591  644
#> 90                    extractRowRanges.html FALSE  18965  644
#> 91                    extractSubstring.html FALSE  15160  644
#> 92                       fileExtension.html FALSE  10205  644
#> 93                         findChanges.html FALSE  13051  644
#> 94               findPartialDuplicates.html FALSE  13297  644
#> 95                    finishAndShowPdf.html FALSE   9693  644
#> 96                  finishAndShowPdfIf.html FALSE   9334  644
#> 97                        firstElement.html FALSE   8982  644
#> 98                    firstPosixColumn.html FALSE   8977  644
#> 99                       frenchToAscii.html FALSE   8992  644
#> 100                     frequencyTable.html FALSE  15710  644
#> 101                        fullySorted.html FALSE  14704  644
#> 102                    generateKeyFile.html FALSE   9217  644
#> 103                       getAttribute.html FALSE  13287  644
#> 104       getByPositiveOrNegativeIndex.html FALSE   9410  644
#> 105                  getElementLengths.html FALSE  10205  644
#> 106                     getEvenNumbers.html FALSE   8998  644
#> 107                    getFunctionName.html FALSE   8931  644
#> 108          getFunctionValueOrDefault.html FALSE   9500  644
#> 109                        getGlobally.html FALSE   9349  644
#> 110                getKeywordPositions.html FALSE   9683  644
#> 111      getNamesOfObjectsInRDataFiles.html FALSE   9141  644
#> 112             getObjectFromRDataFile.html FALSE   9605  644
#> 113                      getOddNumbers.html FALSE   8989  644
#> 114                        getPassword.html FALSE   9974  644
#> 115 getPathsAndValuesFromRecursiveList.html FALSE  12586  644
#> 116                        getTagNames.html FALSE   9946  644
#> 117                        get_homedir.html FALSE   9326  644
#> 118                     guessSeparator.html FALSE   9520  644
#> 119                           headtail.html FALSE  15123  644
#> 120                   hsAddMissingCols.html FALSE   9733  644
#> 121                         hsChrToNum.html FALSE   9800  644
#> 122                       hsCountInStr.html FALSE   9236  644
#> 123                     hsDelEmptyCols.html FALSE  10275  644
#> 124                 hsMatrixToListForm.html FALSE  10639  644
#> 125                      hsMovingMean-1.png FALSE 117570  644
#> 126                       hsMovingMean.html FALSE  12479  644
#> 127              hsOpenWindowsExplorer.html FALSE   9502  644
#> 128                          hsPrepPdf.html FALSE  10177  644
#> 129                         hsQuoteChr.html FALSE   9568  644
#> 130                    hsRenameColumns.html FALSE   9235  644
#> 131                          hsResolve.html FALSE   9710  644
#> 132                hsRestoreAttributes.html FALSE   9136  644
#> 133                         hsSafeName.html FALSE  11137  644
#> 134                            hsShell.html FALSE   9015  644
#> 135                          hsShowPdf.html FALSE  12097  644
#> 136                     hsStringToDate.html FALSE   9523  644
#> 137                   hsStringToDouble.html FALSE   9297  644
#> 138                   hsSubstSpecChars.html FALSE   9413  644
#> 139                           hsSystem.html FALSE   9026  644
#> 140                             hsTrim.html FALSE   9582  644
#> 141                       hsValidValue.html FALSE   9698  644
#> 142                            inRange.html FALSE   9481  644
#> 143                              index.html FALSE  60874  644
#> 144                      insertColumns.html FALSE  16255  644
#> 145                 intToNumeralSystem.html FALSE  14124  644
#> 146                         is.unnamed.html FALSE  11285  644
#> 147                            isASCII.html FALSE  10319  644
#> 148                   isDotOrDoubleDot.html FALSE   9789  644
#> 149                       isEvenNumber.html FALSE   8886  644
#> 150                           isLoaded.html FALSE  10779  644
#> 151                   isNaInAllColumns.html FALSE   9141  644
#> 152                      isNaInAllRows.html FALSE   9123  644
#> 153                        isNaOrEmpty.html FALSE   9205  644
#> 154                      isNullOrEmpty.html FALSE   9073  644
#> 155                        isOddNumber.html FALSE   8877  644
#> 156                        lastElement.html FALSE   9122  644
#> 157                               left.html FALSE   9492  644
#> 158                       limitToRange.html FALSE   9929  644
#> 159                  linearCombination.html FALSE  12314  644
#> 160                        listObjects.html FALSE  10311  644
#>                   mtime               ctime               atime uid gid  uname
#> 1   2022-05-28 08:30:06 2022-05-28 08:30:06 2022-05-28 08:30:06 501  20 runner
#> 2   2022-05-28 08:31:12 2022-05-28 08:31:12 2022-05-28 08:30:07 501  20 runner
#> 3   2022-05-28 08:30:15 2022-05-28 08:30:15 2022-05-28 08:30:09 501  20 runner
#> 4   2022-05-28 08:30:06 2022-05-28 08:30:06 2022-05-28 08:30:06 501  20 runner
#> 5   2022-05-28 08:30:06 2022-05-28 08:30:06 2022-05-28 08:30:06 501  20 runner
#> 6   2022-05-28 08:30:07 2022-05-28 08:30:07 2022-05-28 08:30:07 501  20 runner
#> 7   2022-05-28 08:30:08 2022-05-28 08:30:08 2022-05-28 08:30:08 501  20 runner
#> 8   2022-05-28 08:30:08 2022-05-28 08:30:08 2022-05-28 08:30:08 501  20 runner
#> 9   2022-05-28 08:30:08 2022-05-28 08:30:08 2022-05-28 08:30:08 501  20 runner
#> 10  2022-05-28 08:30:09 2022-05-28 08:30:09 2022-05-28 08:30:09 501  20 runner
#> 11  2022-05-28 08:30:09 2022-05-28 08:30:09 2022-05-28 08:30:09 501  20 runner
#> 12  2022-05-28 08:30:09 2022-05-28 08:30:09 2022-05-28 08:30:09 501  20 runner
#> 13  2022-05-28 08:30:10 2022-05-28 08:30:10 2022-05-28 08:30:10 501  20 runner
#> 14  2022-05-28 08:30:10 2022-05-28 08:30:10 2022-05-28 08:30:10 501  20 runner
#> 15  2022-05-28 08:30:10 2022-05-28 08:30:10 2022-05-28 08:30:10 501  20 runner
#> 16  2022-05-28 08:30:11 2022-05-28 08:30:11 2022-05-28 08:30:11 501  20 runner
#> 17  2022-05-28 08:30:11 2022-05-28 08:30:11 2022-05-28 08:30:11 501  20 runner
#> 18  2022-05-28 08:30:12 2022-05-28 08:30:12 2022-05-28 08:30:12 501  20 runner
#> 19  2022-05-28 08:30:12 2022-05-28 08:30:12 2022-05-28 08:30:12 501  20 runner
#> 20  2022-05-28 08:30:12 2022-05-28 08:30:12 2022-05-28 08:30:12 501  20 runner
#> 21  2022-05-28 08:30:13 2022-05-28 08:30:13 2022-05-28 08:30:13 501  20 runner
#> 22  2022-05-28 08:30:13 2022-05-28 08:30:13 2022-05-28 08:30:13 501  20 runner
#> 23  2022-05-28 08:30:14 2022-05-28 08:30:14 2022-05-28 08:30:14 501  20 runner
#> 24  2022-05-28 08:30:14 2022-05-28 08:30:14 2022-05-28 08:30:14 501  20 runner
#> 25  2022-05-28 08:30:15 2022-05-28 08:30:15 2022-05-28 08:30:15 501  20 runner
#> 26  2022-05-28 08:30:15 2022-05-28 08:30:15 2022-05-28 08:30:15 501  20 runner
#> 27  2022-05-28 08:30:15 2022-05-28 08:30:15 2022-05-28 08:30:15 501  20 runner
#> 28  2022-05-28 08:30:15 2022-05-28 08:30:15 2022-05-28 08:30:15 501  20 runner
#> 29  2022-05-28 08:30:16 2022-05-28 08:30:16 2022-05-28 08:30:16 501  20 runner
#> 30  2022-05-28 08:30:16 2022-05-28 08:30:16 2022-05-28 08:30:16 501  20 runner
#> 31  2022-05-28 08:30:17 2022-05-28 08:30:17 2022-05-28 08:30:17 501  20 runner
#> 32  2022-05-28 08:30:17 2022-05-28 08:30:17 2022-05-28 08:30:17 501  20 runner
#> 33  2022-05-28 08:30:17 2022-05-28 08:30:17 2022-05-28 08:30:17 501  20 runner
#> 34  2022-05-28 08:30:18 2022-05-28 08:30:18 2022-05-28 08:30:18 501  20 runner
#> 35  2022-05-28 08:30:18 2022-05-28 08:30:18 2022-05-28 08:30:18 501  20 runner
#> 36  2022-05-28 08:30:19 2022-05-28 08:30:19 2022-05-28 08:30:19 501  20 runner
#> 37  2022-05-28 08:30:19 2022-05-28 08:30:19 2022-05-28 08:30:19 501  20 runner
#> 38  2022-05-28 08:30:19 2022-05-28 08:30:19 2022-05-28 08:30:19 501  20 runner
#> 39  2022-05-28 08:30:20 2022-05-28 08:30:20 2022-05-28 08:30:20 501  20 runner
#> 40  2022-05-28 08:30:20 2022-05-28 08:30:20 2022-05-28 08:30:20 501  20 runner
#> 41  2022-05-28 08:30:20 2022-05-28 08:30:20 2022-05-28 08:30:20 501  20 runner
#> 42  2022-05-28 08:30:21 2022-05-28 08:30:21 2022-05-28 08:30:21 501  20 runner
#> 43  2022-05-28 08:30:21 2022-05-28 08:30:21 2022-05-28 08:30:21 501  20 runner
#> 44  2022-05-28 08:30:21 2022-05-28 08:30:21 2022-05-28 08:30:21 501  20 runner
#> 45  2022-05-28 08:30:22 2022-05-28 08:30:22 2022-05-28 08:30:22 501  20 runner
#> 46  2022-05-28 08:30:22 2022-05-28 08:30:22 2022-05-28 08:30:22 501  20 runner
#> 47  2022-05-28 08:30:23 2022-05-28 08:30:23 2022-05-28 08:30:23 501  20 runner
#> 48  2022-05-28 08:30:23 2022-05-28 08:30:23 2022-05-28 08:30:23 501  20 runner
#> 49  2022-05-28 08:30:23 2022-05-28 08:30:23 2022-05-28 08:30:23 501  20 runner
#> 50  2022-05-28 08:30:24 2022-05-28 08:30:24 2022-05-28 08:30:24 501  20 runner
#> 51  2022-05-28 08:30:24 2022-05-28 08:30:24 2022-05-28 08:30:24 501  20 runner
#> 52  2022-05-28 08:30:25 2022-05-28 08:30:25 2022-05-28 08:30:25 501  20 runner
#> 53  2022-05-28 08:30:25 2022-05-28 08:30:25 2022-05-28 08:30:25 501  20 runner
#> 54  2022-05-28 08:30:26 2022-05-28 08:30:26 2022-05-28 08:30:26 501  20 runner
#> 55  2022-05-28 08:30:26 2022-05-28 08:30:26 2022-05-28 08:30:26 501  20 runner
#> 56  2022-05-28 08:30:26 2022-05-28 08:30:26 2022-05-28 08:30:26 501  20 runner
#> 57  2022-05-28 08:30:27 2022-05-28 08:30:27 2022-05-28 08:30:27 501  20 runner
#> 58  2022-05-28 08:30:27 2022-05-28 08:30:27 2022-05-28 08:30:27 501  20 runner
#> 59  2022-05-28 08:30:28 2022-05-28 08:30:28 2022-05-28 08:30:28 501  20 runner
#> 60  2022-05-28 08:30:28 2022-05-28 08:30:28 2022-05-28 08:30:28 501  20 runner
#> 61  2022-05-28 08:30:28 2022-05-28 08:30:28 2022-05-28 08:30:28 501  20 runner
#> 62  2022-05-28 08:30:29 2022-05-28 08:30:29 2022-05-28 08:30:29 501  20 runner
#> 63  2022-05-28 08:30:29 2022-05-28 08:30:29 2022-05-28 08:30:29 501  20 runner
#> 64  2022-05-28 08:30:30 2022-05-28 08:30:30 2022-05-28 08:30:30 501  20 runner
#> 65  2022-05-28 08:30:30 2022-05-28 08:30:30 2022-05-28 08:30:30 501  20 runner
#> 66  2022-05-28 08:30:30 2022-05-28 08:30:30 2022-05-28 08:30:30 501  20 runner
#> 67  2022-05-28 08:30:31 2022-05-28 08:30:31 2022-05-28 08:30:31 501  20 runner
#> 68  2022-05-28 08:30:31 2022-05-28 08:30:31 2022-05-28 08:30:31 501  20 runner
#> 69  2022-05-28 08:30:32 2022-05-28 08:30:32 2022-05-28 08:30:32 501  20 runner
#> 70  2022-05-28 08:30:32 2022-05-28 08:30:32 2022-05-28 08:30:32 501  20 runner
#> 71  2022-05-28 08:30:32 2022-05-28 08:30:32 2022-05-28 08:30:32 501  20 runner
#> 72  2022-05-28 08:30:33 2022-05-28 08:30:33 2022-05-28 08:30:33 501  20 runner
#> 73  2022-05-28 08:30:33 2022-05-28 08:30:33 2022-05-28 08:30:33 501  20 runner
#> 74  2022-05-28 08:30:33 2022-05-28 08:30:33 2022-05-28 08:30:33 501  20 runner
#> 75  2022-05-28 08:30:34 2022-05-28 08:30:34 2022-05-28 08:30:34 501  20 runner
#> 76  2022-05-28 08:30:34 2022-05-28 08:30:34 2022-05-28 08:30:34 501  20 runner
#> 77  2022-05-28 08:30:34 2022-05-28 08:30:34 2022-05-28 08:30:34 501  20 runner
#> 78  2022-05-28 08:30:35 2022-05-28 08:30:35 2022-05-28 08:30:35 501  20 runner
#> 79  2022-05-28 08:30:35 2022-05-28 08:30:35 2022-05-28 08:30:35 501  20 runner
#> 80  2022-05-28 08:30:36 2022-05-28 08:30:36 2022-05-28 08:30:36 501  20 runner
#> 81  2022-05-28 08:30:36 2022-05-28 08:30:36 2022-05-28 08:30:36 501  20 runner
#> 82  2022-05-28 08:30:36 2022-05-28 08:30:36 2022-05-28 08:30:36 501  20 runner
#> 83  2022-05-28 08:30:37 2022-05-28 08:30:37 2022-05-28 08:30:37 501  20 runner
#> 84  2022-05-28 08:30:37 2022-05-28 08:30:37 2022-05-28 08:30:37 501  20 runner
#> 85  2022-05-28 08:30:38 2022-05-28 08:30:38 2022-05-28 08:30:38 501  20 runner
#> 86  2022-05-28 08:30:38 2022-05-28 08:30:38 2022-05-28 08:30:38 501  20 runner
#> 87  2022-05-28 08:30:38 2022-05-28 08:30:38 2022-05-28 08:30:38 501  20 runner
#> 88  2022-05-28 08:30:39 2022-05-28 08:30:39 2022-05-28 08:30:39 501  20 runner
#> 89  2022-05-28 08:30:39 2022-05-28 08:30:39 2022-05-28 08:30:39 501  20 runner
#> 90  2022-05-28 08:30:40 2022-05-28 08:30:40 2022-05-28 08:30:40 501  20 runner
#> 91  2022-05-28 08:30:40 2022-05-28 08:30:40 2022-05-28 08:30:40 501  20 runner
#> 92  2022-05-28 08:30:40 2022-05-28 08:30:40 2022-05-28 08:30:40 501  20 runner
#> 93  2022-05-28 08:30:41 2022-05-28 08:30:41 2022-05-28 08:30:41 501  20 runner
#> 94  2022-05-28 08:30:41 2022-05-28 08:30:41 2022-05-28 08:30:41 501  20 runner
#> 95  2022-05-28 08:30:42 2022-05-28 08:30:42 2022-05-28 08:30:42 501  20 runner
#> 96  2022-05-28 08:30:42 2022-05-28 08:30:42 2022-05-28 08:30:42 501  20 runner
#> 97  2022-05-28 08:30:42 2022-05-28 08:30:42 2022-05-28 08:30:42 501  20 runner
#> 98  2022-05-28 08:30:43 2022-05-28 08:30:43 2022-05-28 08:30:43 501  20 runner
#> 99  2022-05-28 08:30:43 2022-05-28 08:30:43 2022-05-28 08:30:43 501  20 runner
#> 100 2022-05-28 08:30:43 2022-05-28 08:30:43 2022-05-28 08:30:43 501  20 runner
#> 101 2022-05-28 08:30:44 2022-05-28 08:30:44 2022-05-28 08:30:44 501  20 runner
#> 102 2022-05-28 08:30:44 2022-05-28 08:30:44 2022-05-28 08:30:44 501  20 runner
#> 103 2022-05-28 08:30:45 2022-05-28 08:30:45 2022-05-28 08:30:45 501  20 runner
#> 104 2022-05-28 08:30:45 2022-05-28 08:30:45 2022-05-28 08:30:45 501  20 runner
#> 105 2022-05-28 08:30:45 2022-05-28 08:30:45 2022-05-28 08:30:45 501  20 runner
#> 106 2022-05-28 08:30:46 2022-05-28 08:30:46 2022-05-28 08:30:46 501  20 runner
#> 107 2022-05-28 08:30:46 2022-05-28 08:30:46 2022-05-28 08:30:46 501  20 runner
#> 108 2022-05-28 08:30:46 2022-05-28 08:30:46 2022-05-28 08:30:46 501  20 runner
#> 109 2022-05-28 08:30:47 2022-05-28 08:30:47 2022-05-28 08:30:47 501  20 runner
#> 110 2022-05-28 08:30:47 2022-05-28 08:30:47 2022-05-28 08:30:47 501  20 runner
#> 111 2022-05-28 08:30:47 2022-05-28 08:30:47 2022-05-28 08:30:47 501  20 runner
#> 112 2022-05-28 08:30:48 2022-05-28 08:30:48 2022-05-28 08:30:48 501  20 runner
#> 113 2022-05-28 08:30:48 2022-05-28 08:30:48 2022-05-28 08:30:48 501  20 runner
#> 114 2022-05-28 08:30:49 2022-05-28 08:30:49 2022-05-28 08:30:49 501  20 runner
#> 115 2022-05-28 08:30:49 2022-05-28 08:30:49 2022-05-28 08:30:49 501  20 runner
#> 116 2022-05-28 08:30:49 2022-05-28 08:30:49 2022-05-28 08:30:49 501  20 runner
#> 117 2022-05-28 08:30:50 2022-05-28 08:30:50 2022-05-28 08:30:50 501  20 runner
#> 118 2022-05-28 08:30:50 2022-05-28 08:30:50 2022-05-28 08:30:50 501  20 runner
#> 119 2022-05-28 08:30:50 2022-05-28 08:30:50 2022-05-28 08:30:50 501  20 runner
#> 120 2022-05-28 08:30:51 2022-05-28 08:30:51 2022-05-28 08:30:51 501  20 runner
#> 121 2022-05-28 08:30:51 2022-05-28 08:30:51 2022-05-28 08:30:51 501  20 runner
#> 122 2022-05-28 08:30:52 2022-05-28 08:30:52 2022-05-28 08:30:52 501  20 runner
#> 123 2022-05-28 08:30:52 2022-05-28 08:30:52 2022-05-28 08:30:52 501  20 runner
#> 124 2022-05-28 08:30:52 2022-05-28 08:30:52 2022-05-28 08:30:52 501  20 runner
#> 125 2022-05-28 08:30:53 2022-05-28 08:30:53 2022-05-28 08:30:52 501  20 runner
#> 126 2022-05-28 08:30:53 2022-05-28 08:30:53 2022-05-28 08:30:53 501  20 runner
#> 127 2022-05-28 08:30:53 2022-05-28 08:30:53 2022-05-28 08:30:53 501  20 runner
#> 128 2022-05-28 08:30:54 2022-05-28 08:30:54 2022-05-28 08:30:54 501  20 runner
#> 129 2022-05-28 08:30:54 2022-05-28 08:30:54 2022-05-28 08:30:54 501  20 runner
#> 130 2022-05-28 08:30:54 2022-05-28 08:30:54 2022-05-28 08:30:54 501  20 runner
#> 131 2022-05-28 08:30:55 2022-05-28 08:30:55 2022-05-28 08:30:55 501  20 runner
#> 132 2022-05-28 08:30:55 2022-05-28 08:30:55 2022-05-28 08:30:55 501  20 runner
#> 133 2022-05-28 08:30:55 2022-05-28 08:30:55 2022-05-28 08:30:55 501  20 runner
#> 134 2022-05-28 08:30:56 2022-05-28 08:30:56 2022-05-28 08:30:56 501  20 runner
#> 135 2022-05-28 08:30:56 2022-05-28 08:30:56 2022-05-28 08:30:56 501  20 runner
#> 136 2022-05-28 08:30:57 2022-05-28 08:30:57 2022-05-28 08:30:57 501  20 runner
#> 137 2022-05-28 08:30:57 2022-05-28 08:30:57 2022-05-28 08:30:57 501  20 runner
#> 138 2022-05-28 08:30:58 2022-05-28 08:30:58 2022-05-28 08:30:58 501  20 runner
#> 139 2022-05-28 08:30:58 2022-05-28 08:30:58 2022-05-28 08:30:58 501  20 runner
#> 140 2022-05-28 08:30:58 2022-05-28 08:30:58 2022-05-28 08:30:58 501  20 runner
#> 141 2022-05-28 08:30:59 2022-05-28 08:30:59 2022-05-28 08:30:59 501  20 runner
#> 142 2022-05-28 08:30:59 2022-05-28 08:30:59 2022-05-28 08:30:59 501  20 runner
#> 143 2022-05-28 08:30:05 2022-05-28 08:30:05 2022-05-28 08:30:05 501  20 runner
#> 144 2022-05-28 08:31:00 2022-05-28 08:31:00 2022-05-28 08:31:00 501  20 runner
#> 145 2022-05-28 08:31:00 2022-05-28 08:31:00 2022-05-28 08:31:00 501  20 runner
#> 146 2022-05-28 08:31:01 2022-05-28 08:31:01 2022-05-28 08:31:01 501  20 runner
#> 147 2022-05-28 08:31:01 2022-05-28 08:31:01 2022-05-28 08:31:01 501  20 runner
#> 148 2022-05-28 08:31:02 2022-05-28 08:31:02 2022-05-28 08:31:02 501  20 runner
#> 149 2022-05-28 08:31:02 2022-05-28 08:31:02 2022-05-28 08:31:02 501  20 runner
#> 150 2022-05-28 08:31:03 2022-05-28 08:31:03 2022-05-28 08:31:03 501  20 runner
#> 151 2022-05-28 08:31:04 2022-05-28 08:31:04 2022-05-28 08:31:04 501  20 runner
#> 152 2022-05-28 08:31:05 2022-05-28 08:31:05 2022-05-28 08:31:05 501  20 runner
#> 153 2022-05-28 08:31:06 2022-05-28 08:31:06 2022-05-28 08:31:06 501  20 runner
#> 154 2022-05-28 08:31:07 2022-05-28 08:31:07 2022-05-28 08:31:07 501  20 runner
#> 155 2022-05-28 08:31:08 2022-05-28 08:31:08 2022-05-28 08:31:08 501  20 runner
#> 156 2022-05-28 08:31:09 2022-05-28 08:31:09 2022-05-28 08:31:09 501  20 runner
#> 157 2022-05-28 08:31:10 2022-05-28 08:31:10 2022-05-28 08:31:10 501  20 runner
#> 158 2022-05-28 08:31:11 2022-05-28 08:31:11 2022-05-28 08:31:11 501  20 runner
#> 159 2022-05-28 08:31:12 2022-05-28 08:31:12 2022-05-28 08:31:12 501  20 runner
#> 160 2022-05-28 08:31:13 2022-05-28 08:31:13 2022-05-28 08:31:13 501  20 runner
#>     grname
#> 1    staff
#> 2    staff
#> 3    staff
#> 4    staff
#> 5    staff
#> 6    staff
#> 7    staff
#> 8    staff
#> 9    staff
#> 10   staff
#> 11   staff
#> 12   staff
#> 13   staff
#> 14   staff
#> 15   staff
#> 16   staff
#> 17   staff
#> 18   staff
#> 19   staff
#> 20   staff
#> 21   staff
#> 22   staff
#> 23   staff
#> 24   staff
#> 25   staff
#> 26   staff
#> 27   staff
#> 28   staff
#> 29   staff
#> 30   staff
#> 31   staff
#> 32   staff
#> 33   staff
#> 34   staff
#> 35   staff
#> 36   staff
#> 37   staff
#> 38   staff
#> 39   staff
#> 40   staff
#> 41   staff
#> 42   staff
#> 43   staff
#> 44   staff
#> 45   staff
#> 46   staff
#> 47   staff
#> 48   staff
#> 49   staff
#> 50   staff
#> 51   staff
#> 52   staff
#> 53   staff
#> 54   staff
#> 55   staff
#> 56   staff
#> 57   staff
#> 58   staff
#> 59   staff
#> 60   staff
#> 61   staff
#> 62   staff
#> 63   staff
#> 64   staff
#> 65   staff
#> 66   staff
#> 67   staff
#> 68   staff
#> 69   staff
#> 70   staff
#> 71   staff
#> 72   staff
#> 73   staff
#> 74   staff
#> 75   staff
#> 76   staff
#> 77   staff
#> 78   staff
#> 79   staff
#> 80   staff
#> 81   staff
#> 82   staff
#> 83   staff
#> 84   staff
#> 85   staff
#> 86   staff
#> 87   staff
#> 88   staff
#> 89   staff
#> 90   staff
#> 91   staff
#> 92   staff
#> 93   staff
#> 94   staff
#> 95   staff
#> 96   staff
#> 97   staff
#> 98   staff
#> 99   staff
#> 100  staff
#> 101  staff
#> 102  staff
#> 103  staff
#> 104  staff
#> 105  staff
#> 106  staff
#> 107  staff
#> 108  staff
#> 109  staff
#> 110  staff
#> 111  staff
#> 112  staff
#> 113  staff
#> 114  staff
#> 115  staff
#> 116  staff
#> 117  staff
#> 118  staff
#> 119  staff
#> 120  staff
#> 121  staff
#> 122  staff
#> 123  staff
#> 124  staff
#> 125  staff
#> 126  staff
#> 127  staff
#> 128  staff
#> 129  staff
#> 130  staff
#> 131  staff
#> 132  staff
#> 133  staff
#> 134  staff
#> 135  staff
#> 136  staff
#> 137  staff
#> 138  staff
#> 139  staff
#> 140  staff
#> 141  staff
#> 142  staff
#> 143  staff
#> 144  staff
#> 145  staff
#> 146  staff
#> 147  staff
#> 148  staff
#> 149  staff
#> 150  staff
#> 151  staff
#> 152  staff
#> 153  staff
#> 154  staff
#> 155  staff
#> 156  staff
#> 157  staff
#> 158  staff
#> 159  staff
#> 160  staff

# ... even when being called with an empty character vector
FUN(character())
#> listing 
#> listing .
#> [1] file  isdir
#> <0 rows> (or 0-length row.names)
FUN(character(), full_info = TRUE)
#> listing 
#> listing .
#>  [1] file   isdir  size   mode   mtime  ctime  atime  uid    gid    uname 
#> [11] grname
#> <0 rows> (or 0-length row.names)

# Call the function recursively up to the given depth level
kwb.utils:::listToDepth(".", max_depth = 1, FUN = FUN)
#> listing .
#> listing .
#>                                        file isdir
#> 1                               DIN.A4.html FALSE
#> 2                              Rplot001.png FALSE
#> 3                              Rplot002.png FALSE
#> 4                             addClass.html FALSE
#> 5                       addRowWithName.html FALSE
#> 6                   addSuffixToColumns.html FALSE
#> 7                          allAreEqual.html FALSE
#> 8                      allAreIdentical.html FALSE
#> 9                          almostEqual.html FALSE
#> 10                        appendSuffix.html FALSE
#> 11                            arglist-1.png FALSE
#> 12                            arglist-2.png FALSE
#> 13                             arglist.html FALSE
#> 14                             argsCsv.html FALSE
#> 15                        asColumnList.html FALSE
#> 16                 asNoFactorDataFrame.html FALSE
#> 17                           asRowList.html FALSE
#> 18                    assertFinalSlash.html FALSE
#> 19                assertRowsAndColumns.html FALSE
#> 20                           assignAll.html FALSE
#> 21              assignArgumentDefaults.html FALSE
#> 22                      assignGlobally.html FALSE
#> 23                assignPackageObjects.html FALSE
#> 24                     atLeastOneRowIn.html FALSE
#> 25                           backspace.html FALSE
#> 26                     breakInSequence.html FALSE
#> 27                           callWith-1.png FALSE
#> 28                           callWith-2.png FALSE
#> 29                            callWith.html FALSE
#> 30            callWithStringsAsFactors.html FALSE
#> 31                           catAndRun.html FALSE
#> 32                               catIf.html FALSE
#> 33                            catLines.html FALSE
#> 34                        catNewLineIf.html FALSE
#> 35              checkForMissingColumns.html FALSE
#> 36                        clearConsole.html FALSE
#> 37                          clipMatrix.html FALSE
#> 38                         cmdLinePath.html FALSE
#> 39                           colMaxima.html FALSE
#> 40                           colMinima.html FALSE
#> 41                        colNaNumbers.html FALSE
#> 42             colStatisticOneFunction.html FALSE
#> 43                       colStatistics.html FALSE
#> 44                           collapsed.html FALSE
#> 45                    columnDescriptor.html FALSE
#> 46                        columnToDate.html FALSE
#> 47                columnwisePercentage.html FALSE
#> 48                combineAlternatingly.html FALSE
#> 49                      commaCollapsed.html FALSE
#> 50                   compareDataFrames.html FALSE
#> 51                         compareSets.html FALSE
#> 52                   containsNulString.html FALSE
#> 53                      convertCsvFile.html FALSE
#> 54              copyDirectoryStructure.html FALSE
#> 55                    copyListElements.html FALSE
#> 56                     countNaInColumn.html FALSE
#> 57                          countOrSum.html FALSE
#> 58                      createAccessor.html FALSE
#> 59              createDirAndReturnPath.html FALSE
#> 60                     createDirectory.html FALSE
#> 61                       createIdAlong.html FALSE
#> 62                        createMatrix.html FALSE
#> 63                  createPasswordFile.html FALSE
#> 64                       createStorage.html FALSE
#> 65                  csvTextToDataFrame.html FALSE
#> 66                              decode.html FALSE
#> 67                         defaultIfNA.html FALSE
#> 68                       defaultIfNULL.html FALSE
#> 69                       defaultIfZero.html FALSE
#> 70                       defaultLevels.html FALSE
#> 71        defaultWindowsProgramFolders.html FALSE
#> 72                             desktop.html FALSE
#> 73                            diffrows.html FALSE
#> 74                          dot-OStype.html FALSE
#> 75                       dot-defaultIf.html FALSE
#> 76                dot-guessSeparator.1.html FALSE
#> 77                dot-guessSeparator.2.html FALSE
#> 78                   dot-isNetworkPath.html FALSE
#> 79                             dot-log.html FALSE
#> 80                         dot-logline.html FALSE
#> 81                           dot-logok.html FALSE
#> 82                        dot-logstart.html FALSE
#> 83                             dropDim.html FALSE
#> 84              dropUnusedFactorLevels.html FALSE
#> 85                              encode.html FALSE
#> 86                       enlargeVector.html FALSE
#> 87                         excludeNULL.html FALSE
#> 88                          expandGrid.html FALSE
#> 89                        extendLimits.html FALSE
#> 90                    extractRowRanges.html FALSE
#> 91                    extractSubstring.html FALSE
#> 92                       fileExtension.html FALSE
#> 93                         findChanges.html FALSE
#> 94               findPartialDuplicates.html FALSE
#> 95                    finishAndShowPdf.html FALSE
#> 96                  finishAndShowPdfIf.html FALSE
#> 97                        firstElement.html FALSE
#> 98                    firstPosixColumn.html FALSE
#> 99                       frenchToAscii.html FALSE
#> 100                     frequencyTable.html FALSE
#> 101                        fullySorted.html FALSE
#> 102                    generateKeyFile.html FALSE
#> 103                       getAttribute.html FALSE
#> 104       getByPositiveOrNegativeIndex.html FALSE
#> 105                  getElementLengths.html FALSE
#> 106                     getEvenNumbers.html FALSE
#> 107                    getFunctionName.html FALSE
#> 108          getFunctionValueOrDefault.html FALSE
#> 109                        getGlobally.html FALSE
#> 110                getKeywordPositions.html FALSE
#> 111      getNamesOfObjectsInRDataFiles.html FALSE
#> 112             getObjectFromRDataFile.html FALSE
#> 113                      getOddNumbers.html FALSE
#> 114                        getPassword.html FALSE
#> 115 getPathsAndValuesFromRecursiveList.html FALSE
#> 116                        getTagNames.html FALSE
#> 117                        get_homedir.html FALSE
#> 118                     guessSeparator.html FALSE
#> 119                           headtail.html FALSE
#> 120                   hsAddMissingCols.html FALSE
#> 121                         hsChrToNum.html FALSE
#> 122                       hsCountInStr.html FALSE
#> 123                     hsDelEmptyCols.html FALSE
#> 124                 hsMatrixToListForm.html FALSE
#> 125                      hsMovingMean-1.png FALSE
#> 126                       hsMovingMean.html FALSE
#> 127              hsOpenWindowsExplorer.html FALSE
#> 128                          hsPrepPdf.html FALSE
#> 129                         hsQuoteChr.html FALSE
#> 130                    hsRenameColumns.html FALSE
#> 131                          hsResolve.html FALSE
#> 132                hsRestoreAttributes.html FALSE
#> 133                         hsSafeName.html FALSE
#> 134                            hsShell.html FALSE
#> 135                          hsShowPdf.html FALSE
#> 136                     hsStringToDate.html FALSE
#> 137                   hsStringToDouble.html FALSE
#> 138                   hsSubstSpecChars.html FALSE
#> 139                           hsSystem.html FALSE
#> 140                             hsTrim.html FALSE
#> 141                       hsValidValue.html FALSE
#> 142                            inRange.html FALSE
#> 143                              index.html FALSE
#> 144                      insertColumns.html FALSE
#> 145                 intToNumeralSystem.html FALSE
#> 146                         is.unnamed.html FALSE
#> 147                            isASCII.html FALSE
#> 148                   isDotOrDoubleDot.html FALSE
#> 149                       isEvenNumber.html FALSE
#> 150                           isLoaded.html FALSE
#> 151                   isNaInAllColumns.html FALSE
#> 152                      isNaInAllRows.html FALSE
#> 153                        isNaOrEmpty.html FALSE
#> 154                      isNullOrEmpty.html FALSE
#> 155                        isOddNumber.html FALSE
#> 156                        lastElement.html FALSE
#> 157                               left.html FALSE
#> 158                       limitToRange.html FALSE
#> 159                  linearCombination.html FALSE
#> 160                        listObjects.html FALSE