List Elements Recursively up to Depth

List Elements Recursively up to Depth

Usage

listToDepth(
  path,
  max_depth = 0L,
  full_info = FALSE,
  FUN = listFiles,
  ...,
  depth = 0,
  prob_mutate = 0,
  template = NULL
)

Arguments

path

path to the element at which to start listing

max_depth

maximal depth level of which to list elements. A value of 0 means non-recursive listing, a value of NA represents fully recursive listing.

full_info

return only path and isdir information or the full information provided by FUN(full_info = TRUE)?

FUN

function called to get the listing of the element given in path. The function must accept a path as its first argument and it must define the argument full_info second. It may accept further arguments. It must always return a data frame. For full_info = FALSE the data frame must have columns file and isdir (is the "file" a directory?). For full_info = TRUE the function may return further columns. The function must provide an empty data frame with the expected columns when being called with character() as the first argument. The function is expected to set the attribute "failed" to the given path in case that the path could not be accessed (e.g. because of a broken internet connection if the listing is done remotely). See kwb.utils:::listFiles for an example implementation that somehow simulates the behaviour of the dir function. See kwb.dwd::list_url() for a more advanced usage of this function in order to recursively list the files on an FTP server (FTP = file transfer protocol).

...

further arguments passed to FUN

depth

start depth of recursion if max_depth > 0. This argument is for internal use and not intended to be set by the user!

prob_mutate

probability to alter the path so that it becomes useless. This is zero by default. Set the value only if you want to test how the function behaves if the listing of a path fails.

template

empty data frame (zero rows) and with columns that are identical to the columns in the data frame returned by FUN. If not provided the function will call FUN once with a zero-length path argument expecting to retrieve the template that is expected here.

Value

data frame containing at least the columns file and

isdir. If full_info = TRUE the result data frame may contain further columns, as provided by the function given in FUN for

full_info = TRUE.

Examples

# Example list function provided in this package (file listing)
FUN <- kwb.utils:::listFiles

# The list function must provide empty records when no path is given. The 
# returned data frame must provide the columns "file" and "isdir" ...
FUN()
#> listing .
#>                                        file isdir
#> 1                               DIN.A4.html FALSE
#> 2                              Rplot001.png FALSE
#> 3                              Rplot002.png FALSE
#> 4                             addClass.html FALSE
#> 5                       addRowWithName.html FALSE
#> 6                   addSuffixToColumns.html FALSE
#> 7                          allAreEqual.html FALSE
#> 8                      allAreIdentical.html FALSE
#> 9                          almostEqual.html FALSE
#> 10                        appendSuffix.html FALSE
#> 11                            arglist-1.png FALSE
#> 12                            arglist-2.png FALSE
#> 13                             arglist.html FALSE
#> 14                             argsCsv.html FALSE
#> 15                    arrayToDataFrame.html FALSE
#> 16                        asColumnList.html FALSE
#> 17                 asNoFactorDataFrame.html FALSE
#> 18                           asRowList.html FALSE
#> 19                    assertFinalSlash.html FALSE
#> 20                assertRowsAndColumns.html FALSE
#> 21                           assignAll.html FALSE
#> 22              assignArgumentDefaults.html FALSE
#> 23                      assignGlobally.html FALSE
#> 24                       assignObjects.html FALSE
#> 25                assignPackageObjects.html FALSE
#> 26                     atLeastOneRowIn.html FALSE
#> 27                           backspace.html FALSE
#> 28                     breakInSequence.html FALSE
#> 29                           callWith-1.png FALSE
#> 30                           callWith-2.png FALSE
#> 31                            callWith.html FALSE
#> 32                        callWithData.html FALSE
#> 33            callWithStringsAsFactors.html FALSE
#> 34                           catAndRun.html FALSE
#> 35                          catChanges.html FALSE
#> 36                        catChangesIf.html FALSE
#> 37                               catIf.html FALSE
#> 38                            catLines.html FALSE
#> 39                        catNewLineIf.html FALSE
#> 40                        catchWarning.html FALSE
#> 41              checkForMissingColumns.html FALSE
#> 42                        clearConsole.html FALSE
#> 43                          clipMatrix.html FALSE
#> 44                         cmdLinePath.html FALSE
#> 45                           colMaxima.html FALSE
#> 46                           colMinima.html FALSE
#> 47                        colNaNumbers.html FALSE
#> 48             colStatisticOneFunction.html FALSE
#> 49                       colStatistics.html FALSE
#> 50                           collapsed.html FALSE
#> 51                    columnDescriptor.html FALSE
#> 52                        columnToDate.html FALSE
#> 53                columnwisePercentage.html FALSE
#> 54                combineAlternatingly.html FALSE
#> 55                      commaCollapsed.html FALSE
#> 56                         commonNames.html FALSE
#> 57                   compareDataFrames.html FALSE
#> 58                         compareSets.html FALSE
#> 59                   containsNulString.html FALSE
#> 60                      convertCsvFile.html FALSE
#> 61                      copyAttributes.html FALSE
#> 62              copyDirectoryStructure.html FALSE
#> 63                    copyListElements.html FALSE
#> 64                     countNaInColumn.html FALSE
#> 65                          countOrSum.html FALSE
#> 66                      createAccessor.html FALSE
#> 67              createDirAndReturnPath.html FALSE
#> 68                   createDirectories.html FALSE
#> 69                     createDirectory.html FALSE
#> 70         createFunctionAssignObjects.html FALSE
#> 71           createFunctionExtdataFile.html FALSE
#> 72                       createIdAlong.html FALSE
#> 73                        createMatrix.html FALSE
#> 74                  createPasswordFile.html FALSE
#> 75                       createStorage.html FALSE
#> 76                  csvTextToDataFrame.html FALSE
#> 77                              decode.html FALSE
#> 78                           defaultIf.html FALSE
#> 79                         defaultIfNA.html FALSE
#> 80                       defaultIfNULL.html FALSE
#> 81                       defaultIfZero.html FALSE
#> 82                       defaultLevels.html FALSE
#> 83        defaultWindowsProgramFolders.html FALSE
#> 84                             desktop.html FALSE
#> 85                            diffrows.html FALSE
#> 86                       directoryName.html FALSE
#> 87                          dot-OStype.html FALSE
#> 88                             dot-log.html FALSE
#> 89                         dot-logline.html FALSE
#> 90                           dot-logok.html FALSE
#> 91                        dot-logstart.html FALSE
#> 92                             dropDim.html FALSE
#> 93              dropUnusedFactorLevels.html FALSE
#> 94                              encode.html FALSE
#> 95                       enlargeVector.html FALSE
#> 96                         excludeNULL.html FALSE
#> 97                          expandGrid.html FALSE
#> 98                         extdataFile.html FALSE
#> 99                        extendLimits.html FALSE
#> 100                   extractRowRanges.html FALSE
#> 101                   extractSubstring.html FALSE
#> 102                      fileExtension.html FALSE
#> 103                        findChanges.html FALSE
#> 104              findPartialDuplicates.html FALSE
#> 105                   finishAndShowPdf.html FALSE
#> 106                 finishAndShowPdfIf.html FALSE
#> 107                       firstElement.html FALSE
#> 108                   firstPosixColumn.html FALSE
#> 109                      frenchToAscii.html FALSE
#> 110                     frequencyTable.html FALSE
#> 111                        fullWinPath.html FALSE
#> 112                        fullySorted.html FALSE
#> 113                    generateKeyFile.html FALSE
#> 114                       getAttribute.html FALSE
#> 115       getByPositiveOrNegativeIndex.html FALSE
#> 116                  getElementLengths.html FALSE
#> 117                     getEvenNumbers.html FALSE
#> 118          getFunctionValueOrDefault.html FALSE
#> 119                        getGlobally.html FALSE
#> 120                getKeywordPositions.html FALSE
#> 121                        getListNode.html FALSE
#> 122      getNamesOfObjectsInRDataFiles.html FALSE
#> 123             getObjectFromRDataFile.html FALSE
#> 124                      getOddNumbers.html FALSE
#> 125                        getPassword.html FALSE
#> 126 getPathsAndValuesFromRecursiveList.html FALSE
#> 127                        getTagNames.html FALSE
#> 128                        get_homedir.html FALSE
#> 129                   guessSeparator.1.html FALSE
#> 130                   guessSeparator.2.html FALSE
#> 131                     guessSeparator.html FALSE
#> 132                      hasFinalSlash.html FALSE
#> 133                      hasZeroLength.html FALSE
#> 134                           headtail.html FALSE
#> 135                   hsAddMissingCols.html FALSE
#> 136                         hsChrToNum.html FALSE
#> 137                       hsCountInStr.html FALSE
#> 138                     hsDelEmptyCols.html FALSE
#> 139                 hsMatrixToListForm.html FALSE
#> 140                      hsMovingMean-1.png FALSE
#> 141                       hsMovingMean.html FALSE
#> 142              hsOpenWindowsExplorer.html FALSE
#> 143                          hsPrepPdf.html FALSE
#> 144                         hsQuoteChr.html FALSE
#> 145                    hsRenameColumns.html FALSE
#> 146                          hsResolve.html FALSE
#> 147                hsRestoreAttributes.html FALSE
#> 148                         hsSafeName.html FALSE
#> 149                            hsShell.html FALSE
#> 150                          hsShowPdf.html FALSE
#> 151                     hsStringToDate.html FALSE
#> 152                   hsStringToDouble.html FALSE
#> 153                   hsSubstSpecChars.html FALSE
#> 154                           hsSystem.html FALSE
#> 155                             hsTrim.html FALSE
#> 156                       hsValidValue.html FALSE
#> 157                            inRange.html FALSE
#> 158                             indent.html FALSE
#> 159                              index.html FALSE
#> 160                      insertColumns.html FALSE
#> 161                 intToNumeralSystem.html FALSE
#> 162                         is.unnamed.html FALSE
#> 163                            isASCII.html FALSE
#> 164                   isDotOrDoubleDot.html FALSE
#> 165                       isEvenNumber.html FALSE
#> 166                           isLoaded.html FALSE
#> 167                   isNaInAllColumns.html FALSE
#> 168                      isNaInAllRows.html FALSE
#> 169                        isNaOrEmpty.html FALSE
#> 170                      isNetworkPath.html FALSE
#> 171                      isNullOrEmpty.html FALSE
#> 172                        isOddNumber.html FALSE
#> 173                         isTryError.html FALSE
#> 174                        lastElement.html FALSE
#> 175                               left.html FALSE
#> 176                leftSubstringEquals.html FALSE
#> 177                       limitToRange.html FALSE
#> 178                  linearCombination.html FALSE
#> 179                        listObjects.html FALSE
FUN(full_info = TRUE)
#> listing .
#>                                        file isdir   size mode
#> 1                               DIN.A4.html FALSE   8375  644
#> 2                              Rplot001.png FALSE   1011  644
#> 3                              Rplot002.png FALSE   5908  644
#> 4                             addClass.html FALSE   9552  644
#> 5                       addRowWithName.html FALSE   9236  644
#> 6                   addSuffixToColumns.html FALSE  11997  644
#> 7                          allAreEqual.html FALSE   9191  644
#> 8                      allAreIdentical.html FALSE   9036  644
#> 9                          almostEqual.html FALSE   9207  644
#> 10                        appendSuffix.html FALSE  10889  644
#> 11                            arglist-1.png FALSE  18498  644
#> 12                            arglist-2.png FALSE  18423  644
#> 13                             arglist.html FALSE  13120  644
#> 14                             argsCsv.html FALSE   9072  644
#> 15                    arrayToDataFrame.html FALSE  13712  644
#> 16                        asColumnList.html FALSE  11103  644
#> 17                 asNoFactorDataFrame.html FALSE  11437  644
#> 18                           asRowList.html FALSE  10987  644
#> 19                    assertFinalSlash.html FALSE  12033  644
#> 20                assertRowsAndColumns.html FALSE  12575  644
#> 21                           assignAll.html FALSE   9027  644
#> 22              assignArgumentDefaults.html FALSE  10456  644
#> 23                      assignGlobally.html FALSE   8715  644
#> 24                       assignObjects.html FALSE   9786  644
#> 25                assignPackageObjects.html FALSE   9284  644
#> 26                     atLeastOneRowIn.html FALSE   8789  644
#> 27                           backspace.html FALSE  10557  644
#> 28                     breakInSequence.html FALSE   9168  644
#> 29                           callWith-1.png FALSE  18498  644
#> 30                           callWith-2.png FALSE  18423  644
#> 31                            callWith.html FALSE  12220  644
#> 32                        callWithData.html FALSE  12251  644
#> 33            callWithStringsAsFactors.html FALSE  14694  644
#> 34                           catAndRun.html FALSE  12980  644
#> 35                          catChanges.html FALSE   9642  644
#> 36                        catChangesIf.html FALSE   9089  644
#> 37                               catIf.html FALSE   8748  644
#> 38                            catLines.html FALSE   8851  644
#> 39                        catNewLineIf.html FALSE   9100  644
#> 40                        catchWarning.html FALSE  11303  644
#> 41              checkForMissingColumns.html FALSE   9994  644
#> 42                        clearConsole.html FALSE   8307  644
#> 43                          clipMatrix.html FALSE   9041  644
#> 44                         cmdLinePath.html FALSE   8824  644
#> 45                           colMaxima.html FALSE   8838  644
#> 46                           colMinima.html FALSE   8838  644
#> 47                        colNaNumbers.html FALSE   8774  644
#> 48             colStatisticOneFunction.html FALSE   9265  644
#> 49                       colStatistics.html FALSE   9881  644
#> 50                           collapsed.html FALSE   9031  644
#> 51                    columnDescriptor.html FALSE   8942  644
#> 52                        columnToDate.html FALSE  11481  644
#> 53                columnwisePercentage.html FALSE  14885  644
#> 54                combineAlternatingly.html FALSE  12152  644
#> 55                      commaCollapsed.html FALSE   8654  644
#> 56                         commonNames.html FALSE  10199  644
#> 57                   compareDataFrames.html FALSE  13614  644
#> 58                         compareSets.html FALSE  11923  644
#> 59                   containsNulString.html FALSE   8977  644
#> 60                      convertCsvFile.html FALSE  22650  644
#> 61                      copyAttributes.html FALSE  10680  644
#> 62              copyDirectoryStructure.html FALSE   9875  644
#> 63                    copyListElements.html FALSE  12016  644
#> 64                     countNaInColumn.html FALSE   9029  644
#> 65                          countOrSum.html FALSE  13958  644
#> 66                      createAccessor.html FALSE  10834  644
#> 67              createDirAndReturnPath.html FALSE   9509  644
#> 68                   createDirectories.html FALSE   9000  644
#> 69                     createDirectory.html FALSE   9402  644
#> 70         createFunctionAssignObjects.html FALSE  10193  644
#> 71           createFunctionExtdataFile.html FALSE  10213  644
#> 72                       createIdAlong.html FALSE  12014  644
#> 73                        createMatrix.html FALSE  14026  644
#> 74                  createPasswordFile.html FALSE   9841  644
#> 75                       createStorage.html FALSE  14976  644
#> 76                  csvTextToDataFrame.html FALSE   8887  644
#> 77                              decode.html FALSE   8951  644
#> 78                           defaultIf.html FALSE   9378  644
#> 79                         defaultIfNA.html FALSE  10556  644
#> 80                       defaultIfNULL.html FALSE  10579  644
#> 81                       defaultIfZero.html FALSE  11871  644
#> 82                       defaultLevels.html FALSE  10670  644
#> 83        defaultWindowsProgramFolders.html FALSE   8451  644
#> 84                             desktop.html FALSE   9016  644
#> 85                            diffrows.html FALSE  10953  644
#> 86                       directoryName.html FALSE  10273  644
#> 87                          dot-OStype.html FALSE   8700  644
#> 88                             dot-log.html FALSE   8749  644
#> 89                         dot-logline.html FALSE   8903  644
#> 90                           dot-logok.html FALSE   8839  644
#> 91                        dot-logstart.html FALSE   9034  644
#> 92                             dropDim.html FALSE  14070  644
#> 93              dropUnusedFactorLevels.html FALSE  14026  644
#> 94                              encode.html FALSE  11273  644
#> 95                       enlargeVector.html FALSE  11615  644
#> 96                         excludeNULL.html FALSE  11121  644
#> 97                          expandGrid.html FALSE  14794  644
#> 98                         extdataFile.html FALSE  10379  644
#> 99                        extendLimits.html FALSE   9359  644
#> 100                   extractRowRanges.html FALSE  19736  644
#> 101                   extractSubstring.html FALSE  15414  644
#> 102                      fileExtension.html FALSE  10154  644
#> 103                        findChanges.html FALSE  13342  644
#> 104              findPartialDuplicates.html FALSE  13423  644
#> 105                   finishAndShowPdf.html FALSE   9503  644
#> 106                 finishAndShowPdfIf.html FALSE   9112  644
#> 107                       firstElement.html FALSE   8839  644
#> 108                   firstPosixColumn.html FALSE   8734  644
#> 109                      frenchToAscii.html FALSE   8756  644
#> 110                     frequencyTable.html FALSE  15937  644
#> 111                        fullWinPath.html FALSE  10710  644
#> 112                        fullySorted.html FALSE  14627  644
#> 113                    generateKeyFile.html FALSE   9056  644
#> 114                       getAttribute.html FALSE  13409  644
#> 115       getByPositiveOrNegativeIndex.html FALSE   9273  644
#> 116                  getElementLengths.html FALSE  10170  644
#> 117                     getEvenNumbers.html FALSE   8769  644
#> 118          getFunctionValueOrDefault.html FALSE   9303  644
#> 119                        getGlobally.html FALSE   9108  644
#> 120                getKeywordPositions.html FALSE   9526  644
#> 121                        getListNode.html FALSE  12846  644
#> 122      getNamesOfObjectsInRDataFiles.html FALSE   8904  644
#> 123             getObjectFromRDataFile.html FALSE   9459  644
#> 124                      getOddNumbers.html FALSE   8759  644
#> 125                        getPassword.html FALSE  10065  644
#> 126 getPathsAndValuesFromRecursiveList.html FALSE  12659  644
#> 127                        getTagNames.html FALSE   9758  644
#> 128                        get_homedir.html FALSE   9172  644
#> 129                   guessSeparator.1.html FALSE   9053  644
#> 130                   guessSeparator.2.html FALSE   8927  644
#> 131                     guessSeparator.html FALSE   9286  644
#> 132                      hasFinalSlash.html FALSE   9784  644
#> 133                      hasZeroLength.html FALSE  11368  644
#> 134                           headtail.html FALSE  15213  644
#> 135                   hsAddMissingCols.html FALSE   9716  644
#> 136                         hsChrToNum.html FALSE   9684  644
#> 137                       hsCountInStr.html FALSE   9079  644
#> 138                     hsDelEmptyCols.html FALSE  10243  644
#> 139                 hsMatrixToListForm.html FALSE  10699  644
#> 140                      hsMovingMean-1.png FALSE 100974  644
#> 141                       hsMovingMean.html FALSE  12611  644
#> 142              hsOpenWindowsExplorer.html FALSE   9286  644
#> 143                          hsPrepPdf.html FALSE  10047  644
#> 144                         hsQuoteChr.html FALSE   9354  644
#> 145                    hsRenameColumns.html FALSE   9034  644
#> 146                          hsResolve.html FALSE   9588  644
#> 147                hsRestoreAttributes.html FALSE   8929  644
#> 148                         hsSafeName.html FALSE  11205  644
#> 149                            hsShell.html FALSE   8766  644
#> 150                          hsShowPdf.html FALSE  12277  644
#> 151                     hsStringToDate.html FALSE   9382  644
#> 152                   hsStringToDouble.html FALSE   9180  644
#> 153                   hsSubstSpecChars.html FALSE   9263  644
#> 154                           hsSystem.html FALSE   8778  644
#> 155                             hsTrim.html FALSE   9456  644
#> 156                       hsValidValue.html FALSE   9579  644
#> 157                            inRange.html FALSE   9344  644
#> 158                             indent.html FALSE   9006  644
#> 159                              index.html FALSE  65963  644
#> 160                      insertColumns.html FALSE  16515  644
#> 161                 intToNumeralSystem.html FALSE  14121  644
#> 162                         is.unnamed.html FALSE  11442  644
#> 163                            isASCII.html FALSE  10322  644
#> 164                   isDotOrDoubleDot.html FALSE   9760  644
#> 165                       isEvenNumber.html FALSE   8635  644
#> 166                           isLoaded.html FALSE  10977  644
#> 167                   isNaInAllColumns.html FALSE   8979  644
#> 168                      isNaInAllRows.html FALSE   8958  644
#> 169                        isNaOrEmpty.html FALSE   9078  644
#> 170                      isNetworkPath.html FALSE   9275  644
#> 171                      isNullOrEmpty.html FALSE   8894  644
#> 172                        isOddNumber.html FALSE   8625  644
#> 173                         isTryError.html FALSE   9930  644
#> 174                        lastElement.html FALSE   8987  644
#> 175                               left.html FALSE   9435  644
#> 176                leftSubstringEquals.html FALSE  10033  644
#> 177                       limitToRange.html FALSE   9858  644
#> 178                  linearCombination.html FALSE  12329  644
#> 179                        listObjects.html FALSE  10355  644
#>                   mtime               ctime               atime uid gid  uname
#> 1   2024-03-18 20:08:41 2024-03-18 20:08:41 2024-03-18 20:08:41 501  20 runner
#> 2   2024-03-18 20:09:54 2024-03-18 20:09:54 2024-03-18 20:08:42 501  20 runner
#> 3   2024-03-18 20:08:51 2024-03-18 20:08:51 2024-03-18 20:08:44 501  20 runner
#> 4   2024-03-18 20:08:41 2024-03-18 20:08:41 2024-03-18 20:08:41 501  20 runner
#> 5   2024-03-18 20:08:41 2024-03-18 20:08:41 2024-03-18 20:08:41 501  20 runner
#> 6   2024-03-18 20:08:42 2024-03-18 20:08:42 2024-03-18 20:08:42 501  20 runner
#> 7   2024-03-18 20:08:43 2024-03-18 20:08:43 2024-03-18 20:08:43 501  20 runner
#> 8   2024-03-18 20:08:43 2024-03-18 20:08:43 2024-03-18 20:08:43 501  20 runner
#> 9   2024-03-18 20:08:44 2024-03-18 20:08:44 2024-03-18 20:08:44 501  20 runner
#> 10  2024-03-18 20:08:44 2024-03-18 20:08:44 2024-03-18 20:08:44 501  20 runner
#> 11  2024-03-18 20:08:44 2024-03-18 20:08:44 2024-03-18 20:08:44 501  20 runner
#> 12  2024-03-18 20:08:44 2024-03-18 20:08:44 2024-03-18 20:08:44 501  20 runner
#> 13  2024-03-18 20:08:45 2024-03-18 20:08:45 2024-03-18 20:08:45 501  20 runner
#> 14  2024-03-18 20:08:45 2024-03-18 20:08:45 2024-03-18 20:08:45 501  20 runner
#> 15  2024-03-18 20:08:46 2024-03-18 20:08:46 2024-03-18 20:08:46 501  20 runner
#> 16  2024-03-18 20:08:46 2024-03-18 20:08:46 2024-03-18 20:08:46 501  20 runner
#> 17  2024-03-18 20:08:47 2024-03-18 20:08:47 2024-03-18 20:08:47 501  20 runner
#> 18  2024-03-18 20:08:47 2024-03-18 20:08:47 2024-03-18 20:08:47 501  20 runner
#> 19  2024-03-18 20:08:48 2024-03-18 20:08:48 2024-03-18 20:08:48 501  20 runner
#> 20  2024-03-18 20:08:48 2024-03-18 20:08:48 2024-03-18 20:08:48 501  20 runner
#> 21  2024-03-18 20:08:48 2024-03-18 20:08:48 2024-03-18 20:08:48 501  20 runner
#> 22  2024-03-18 20:08:49 2024-03-18 20:08:49 2024-03-18 20:08:49 501  20 runner
#> 23  2024-03-18 20:08:49 2024-03-18 20:08:49 2024-03-18 20:08:49 501  20 runner
#> 24  2024-03-18 20:08:50 2024-03-18 20:08:50 2024-03-18 20:08:50 501  20 runner
#> 25  2024-03-18 20:08:50 2024-03-18 20:08:50 2024-03-18 20:08:50 501  20 runner
#> 26  2024-03-18 20:08:50 2024-03-18 20:08:50 2024-03-18 20:08:50 501  20 runner
#> 27  2024-03-18 20:08:51 2024-03-18 20:08:51 2024-03-18 20:08:51 501  20 runner
#> 28  2024-03-18 20:08:51 2024-03-18 20:08:51 2024-03-18 20:08:51 501  20 runner
#> 29  2024-03-18 20:08:52 2024-03-18 20:08:52 2024-03-18 20:08:51 501  20 runner
#> 30  2024-03-18 20:08:52 2024-03-18 20:08:52 2024-03-18 20:08:52 501  20 runner
#> 31  2024-03-18 20:08:52 2024-03-18 20:08:52 2024-03-18 20:08:52 501  20 runner
#> 32  2024-03-18 20:08:53 2024-03-18 20:08:53 2024-03-18 20:08:53 501  20 runner
#> 33  2024-03-18 20:08:53 2024-03-18 20:08:53 2024-03-18 20:08:53 501  20 runner
#> 34  2024-03-18 20:08:54 2024-03-18 20:08:54 2024-03-18 20:08:54 501  20 runner
#> 35  2024-03-18 20:08:54 2024-03-18 20:08:54 2024-03-18 20:08:54 501  20 runner
#> 36  2024-03-18 20:08:54 2024-03-18 20:08:54 2024-03-18 20:08:54 501  20 runner
#> 37  2024-03-18 20:08:55 2024-03-18 20:08:55 2024-03-18 20:08:55 501  20 runner
#> 38  2024-03-18 20:08:55 2024-03-18 20:08:55 2024-03-18 20:08:55 501  20 runner
#> 39  2024-03-18 20:08:55 2024-03-18 20:08:55 2024-03-18 20:08:55 501  20 runner
#> 40  2024-03-18 20:08:56 2024-03-18 20:08:56 2024-03-18 20:08:56 501  20 runner
#> 41  2024-03-18 20:08:56 2024-03-18 20:08:56 2024-03-18 20:08:56 501  20 runner
#> 42  2024-03-18 20:08:57 2024-03-18 20:08:57 2024-03-18 20:08:57 501  20 runner
#> 43  2024-03-18 20:08:57 2024-03-18 20:08:57 2024-03-18 20:08:57 501  20 runner
#> 44  2024-03-18 20:08:57 2024-03-18 20:08:57 2024-03-18 20:08:57 501  20 runner
#> 45  2024-03-18 20:08:58 2024-03-18 20:08:58 2024-03-18 20:08:58 501  20 runner
#> 46  2024-03-18 20:08:58 2024-03-18 20:08:58 2024-03-18 20:08:58 501  20 runner
#> 47  2024-03-18 20:08:59 2024-03-18 20:08:59 2024-03-18 20:08:59 501  20 runner
#> 48  2024-03-18 20:08:59 2024-03-18 20:08:59 2024-03-18 20:08:59 501  20 runner
#> 49  2024-03-18 20:08:59 2024-03-18 20:08:59 2024-03-18 20:08:59 501  20 runner
#> 50  2024-03-18 20:09:00 2024-03-18 20:09:00 2024-03-18 20:09:00 501  20 runner
#> 51  2024-03-18 20:09:00 2024-03-18 20:09:00 2024-03-18 20:09:00 501  20 runner
#> 52  2024-03-18 20:09:01 2024-03-18 20:09:01 2024-03-18 20:09:01 501  20 runner
#> 53  2024-03-18 20:09:01 2024-03-18 20:09:01 2024-03-18 20:09:01 501  20 runner
#> 54  2024-03-18 20:09:02 2024-03-18 20:09:02 2024-03-18 20:09:02 501  20 runner
#> 55  2024-03-18 20:09:02 2024-03-18 20:09:02 2024-03-18 20:09:02 501  20 runner
#> 56  2024-03-18 20:09:02 2024-03-18 20:09:02 2024-03-18 20:09:02 501  20 runner
#> 57  2024-03-18 20:09:03 2024-03-18 20:09:03 2024-03-18 20:09:03 501  20 runner
#> 58  2024-03-18 20:09:03 2024-03-18 20:09:03 2024-03-18 20:09:03 501  20 runner
#> 59  2024-03-18 20:09:04 2024-03-18 20:09:04 2024-03-18 20:09:04 501  20 runner
#> 60  2024-03-18 20:09:04 2024-03-18 20:09:04 2024-03-18 20:09:04 501  20 runner
#> 61  2024-03-18 20:09:05 2024-03-18 20:09:05 2024-03-18 20:09:05 501  20 runner
#> 62  2024-03-18 20:09:05 2024-03-18 20:09:05 2024-03-18 20:09:05 501  20 runner
#> 63  2024-03-18 20:09:06 2024-03-18 20:09:06 2024-03-18 20:09:06 501  20 runner
#> 64  2024-03-18 20:09:06 2024-03-18 20:09:06 2024-03-18 20:09:06 501  20 runner
#> 65  2024-03-18 20:09:07 2024-03-18 20:09:07 2024-03-18 20:09:07 501  20 runner
#> 66  2024-03-18 20:09:07 2024-03-18 20:09:07 2024-03-18 20:09:07 501  20 runner
#> 67  2024-03-18 20:09:07 2024-03-18 20:09:07 2024-03-18 20:09:07 501  20 runner
#> 68  2024-03-18 20:09:08 2024-03-18 20:09:08 2024-03-18 20:09:08 501  20 runner
#> 69  2024-03-18 20:09:08 2024-03-18 20:09:08 2024-03-18 20:09:08 501  20 runner
#> 70  2024-03-18 20:09:09 2024-03-18 20:09:09 2024-03-18 20:09:09 501  20 runner
#> 71  2024-03-18 20:09:09 2024-03-18 20:09:09 2024-03-18 20:09:09 501  20 runner
#> 72  2024-03-18 20:09:09 2024-03-18 20:09:09 2024-03-18 20:09:09 501  20 runner
#> 73  2024-03-18 20:09:10 2024-03-18 20:09:10 2024-03-18 20:09:10 501  20 runner
#> 74  2024-03-18 20:09:10 2024-03-18 20:09:10 2024-03-18 20:09:10 501  20 runner
#> 75  2024-03-18 20:09:11 2024-03-18 20:09:11 2024-03-18 20:09:11 501  20 runner
#> 76  2024-03-18 20:09:11 2024-03-18 20:09:11 2024-03-18 20:09:11 501  20 runner
#> 77  2024-03-18 20:09:12 2024-03-18 20:09:12 2024-03-18 20:09:12 501  20 runner
#> 78  2024-03-18 20:09:12 2024-03-18 20:09:12 2024-03-18 20:09:12 501  20 runner
#> 79  2024-03-18 20:09:12 2024-03-18 20:09:12 2024-03-18 20:09:12 501  20 runner
#> 80  2024-03-18 20:09:13 2024-03-18 20:09:13 2024-03-18 20:09:13 501  20 runner
#> 81  2024-03-18 20:09:13 2024-03-18 20:09:13 2024-03-18 20:09:13 501  20 runner
#> 82  2024-03-18 20:09:14 2024-03-18 20:09:14 2024-03-18 20:09:14 501  20 runner
#> 83  2024-03-18 20:09:14 2024-03-18 20:09:14 2024-03-18 20:09:14 501  20 runner
#> 84  2024-03-18 20:09:14 2024-03-18 20:09:14 2024-03-18 20:09:14 501  20 runner
#> 85  2024-03-18 20:09:15 2024-03-18 20:09:15 2024-03-18 20:09:15 501  20 runner
#> 86  2024-03-18 20:09:15 2024-03-18 20:09:15 2024-03-18 20:09:15 501  20 runner
#> 87  2024-03-18 20:09:16 2024-03-18 20:09:16 2024-03-18 20:09:16 501  20 runner
#> 88  2024-03-18 20:09:16 2024-03-18 20:09:16 2024-03-18 20:09:16 501  20 runner
#> 89  2024-03-18 20:09:17 2024-03-18 20:09:17 2024-03-18 20:09:17 501  20 runner
#> 90  2024-03-18 20:09:17 2024-03-18 20:09:17 2024-03-18 20:09:17 501  20 runner
#> 91  2024-03-18 20:09:17 2024-03-18 20:09:17 2024-03-18 20:09:17 501  20 runner
#> 92  2024-03-18 20:09:18 2024-03-18 20:09:18 2024-03-18 20:09:18 501  20 runner
#> 93  2024-03-18 20:09:18 2024-03-18 20:09:18 2024-03-18 20:09:18 501  20 runner
#> 94  2024-03-18 20:09:19 2024-03-18 20:09:19 2024-03-18 20:09:19 501  20 runner
#> 95  2024-03-18 20:09:19 2024-03-18 20:09:19 2024-03-18 20:09:19 501  20 runner
#> 96  2024-03-18 20:09:20 2024-03-18 20:09:20 2024-03-18 20:09:20 501  20 runner
#> 97  2024-03-18 20:09:20 2024-03-18 20:09:20 2024-03-18 20:09:20 501  20 runner
#> 98  2024-03-18 20:09:21 2024-03-18 20:09:21 2024-03-18 20:09:21 501  20 runner
#> 99  2024-03-18 20:09:21 2024-03-18 20:09:21 2024-03-18 20:09:21 501  20 runner
#> 100 2024-03-18 20:09:22 2024-03-18 20:09:22 2024-03-18 20:09:22 501  20 runner
#> 101 2024-03-18 20:09:22 2024-03-18 20:09:22 2024-03-18 20:09:22 501  20 runner
#> 102 2024-03-18 20:09:23 2024-03-18 20:09:23 2024-03-18 20:09:23 501  20 runner
#> 103 2024-03-18 20:09:23 2024-03-18 20:09:23 2024-03-18 20:09:23 501  20 runner
#> 104 2024-03-18 20:09:24 2024-03-18 20:09:24 2024-03-18 20:09:24 501  20 runner
#> 105 2024-03-18 20:09:24 2024-03-18 20:09:24 2024-03-18 20:09:24 501  20 runner
#> 106 2024-03-18 20:09:25 2024-03-18 20:09:25 2024-03-18 20:09:25 501  20 runner
#> 107 2024-03-18 20:09:25 2024-03-18 20:09:25 2024-03-18 20:09:25 501  20 runner
#> 108 2024-03-18 20:09:25 2024-03-18 20:09:25 2024-03-18 20:09:25 501  20 runner
#> 109 2024-03-18 20:09:26 2024-03-18 20:09:26 2024-03-18 20:09:26 501  20 runner
#> 110 2024-03-18 20:09:26 2024-03-18 20:09:26 2024-03-18 20:09:26 501  20 runner
#> 111 2024-03-18 20:09:27 2024-03-18 20:09:27 2024-03-18 20:09:27 501  20 runner
#> 112 2024-03-18 20:09:27 2024-03-18 20:09:27 2024-03-18 20:09:27 501  20 runner
#> 113 2024-03-18 20:09:28 2024-03-18 20:09:28 2024-03-18 20:09:28 501  20 runner
#> 114 2024-03-18 20:09:28 2024-03-18 20:09:28 2024-03-18 20:09:28 501  20 runner
#> 115 2024-03-18 20:09:28 2024-03-18 20:09:28 2024-03-18 20:09:28 501  20 runner
#> 116 2024-03-18 20:09:29 2024-03-18 20:09:29 2024-03-18 20:09:29 501  20 runner
#> 117 2024-03-18 20:09:29 2024-03-18 20:09:29 2024-03-18 20:09:29 501  20 runner
#> 118 2024-03-18 20:09:29 2024-03-18 20:09:29 2024-03-18 20:09:29 501  20 runner
#> 119 2024-03-18 20:09:30 2024-03-18 20:09:30 2024-03-18 20:09:30 501  20 runner
#> 120 2024-03-18 20:09:30 2024-03-18 20:09:30 2024-03-18 20:09:30 501  20 runner
#> 121 2024-03-18 20:09:31 2024-03-18 20:09:31 2024-03-18 20:09:31 501  20 runner
#> 122 2024-03-18 20:09:31 2024-03-18 20:09:31 2024-03-18 20:09:31 501  20 runner
#> 123 2024-03-18 20:09:32 2024-03-18 20:09:32 2024-03-18 20:09:32 501  20 runner
#> 124 2024-03-18 20:09:32 2024-03-18 20:09:32 2024-03-18 20:09:32 501  20 runner
#> 125 2024-03-18 20:09:32 2024-03-18 20:09:32 2024-03-18 20:09:32 501  20 runner
#> 126 2024-03-18 20:09:33 2024-03-18 20:09:33 2024-03-18 20:09:33 501  20 runner
#> 127 2024-03-18 20:09:33 2024-03-18 20:09:33 2024-03-18 20:09:33 501  20 runner
#> 128 2024-03-18 20:09:34 2024-03-18 20:09:34 2024-03-18 20:09:34 501  20 runner
#> 129 2024-03-18 20:09:34 2024-03-18 20:09:34 2024-03-18 20:09:34 501  20 runner
#> 130 2024-03-18 20:09:34 2024-03-18 20:09:34 2024-03-18 20:09:34 501  20 runner
#> 131 2024-03-18 20:09:35 2024-03-18 20:09:35 2024-03-18 20:09:35 501  20 runner
#> 132 2024-03-18 20:09:35 2024-03-18 20:09:35 2024-03-18 20:09:35 501  20 runner
#> 133 2024-03-18 20:09:36 2024-03-18 20:09:36 2024-03-18 20:09:36 501  20 runner
#> 134 2024-03-18 20:09:36 2024-03-18 20:09:36 2024-03-18 20:09:36 501  20 runner
#> 135 2024-03-18 20:09:37 2024-03-18 20:09:37 2024-03-18 20:09:37 501  20 runner
#> 136 2024-03-18 20:09:37 2024-03-18 20:09:37 2024-03-18 20:09:37 501  20 runner
#> 137 2024-03-18 20:09:37 2024-03-18 20:09:37 2024-03-18 20:09:37 501  20 runner
#> 138 2024-03-18 20:09:38 2024-03-18 20:09:38 2024-03-18 20:09:38 501  20 runner
#> 139 2024-03-18 20:09:38 2024-03-18 20:09:38 2024-03-18 20:09:38 501  20 runner
#> 140 2024-03-18 20:09:39 2024-03-18 20:09:39 2024-03-18 20:09:38 501  20 runner
#> 141 2024-03-18 20:09:39 2024-03-18 20:09:39 2024-03-18 20:09:39 501  20 runner
#> 142 2024-03-18 20:09:39 2024-03-18 20:09:39 2024-03-18 20:09:39 501  20 runner
#> 143 2024-03-18 20:09:40 2024-03-18 20:09:40 2024-03-18 20:09:40 501  20 runner
#> 144 2024-03-18 20:09:40 2024-03-18 20:09:40 2024-03-18 20:09:40 501  20 runner
#> 145 2024-03-18 20:09:41 2024-03-18 20:09:41 2024-03-18 20:09:41 501  20 runner
#> 146 2024-03-18 20:09:41 2024-03-18 20:09:41 2024-03-18 20:09:41 501  20 runner
#> 147 2024-03-18 20:09:41 2024-03-18 20:09:41 2024-03-18 20:09:41 501  20 runner
#> 148 2024-03-18 20:09:42 2024-03-18 20:09:42 2024-03-18 20:09:42 501  20 runner
#> 149 2024-03-18 20:09:42 2024-03-18 20:09:42 2024-03-18 20:09:42 501  20 runner
#> 150 2024-03-18 20:09:43 2024-03-18 20:09:43 2024-03-18 20:09:43 501  20 runner
#> 151 2024-03-18 20:09:43 2024-03-18 20:09:43 2024-03-18 20:09:43 501  20 runner
#> 152 2024-03-18 20:09:44 2024-03-18 20:09:44 2024-03-18 20:09:44 501  20 runner
#> 153 2024-03-18 20:09:44 2024-03-18 20:09:44 2024-03-18 20:09:44 501  20 runner
#> 154 2024-03-18 20:09:44 2024-03-18 20:09:44 2024-03-18 20:09:44 501  20 runner
#> 155 2024-03-18 20:09:45 2024-03-18 20:09:45 2024-03-18 20:09:45 501  20 runner
#> 156 2024-03-18 20:09:45 2024-03-18 20:09:45 2024-03-18 20:09:45 501  20 runner
#> 157 2024-03-18 20:09:45 2024-03-18 20:09:45 2024-03-18 20:09:45 501  20 runner
#> 158 2024-03-18 20:09:46 2024-03-18 20:09:46 2024-03-18 20:09:46 501  20 runner
#> 159 2024-03-18 20:08:40 2024-03-18 20:08:40 2024-03-18 20:08:40 501  20 runner
#> 160 2024-03-18 20:09:46 2024-03-18 20:09:46 2024-03-18 20:09:46 501  20 runner
#> 161 2024-03-18 20:09:47 2024-03-18 20:09:47 2024-03-18 20:09:47 501  20 runner
#> 162 2024-03-18 20:09:47 2024-03-18 20:09:47 2024-03-18 20:09:47 501  20 runner
#> 163 2024-03-18 20:09:48 2024-03-18 20:09:48 2024-03-18 20:09:48 501  20 runner
#> 164 2024-03-18 20:09:48 2024-03-18 20:09:48 2024-03-18 20:09:48 501  20 runner
#> 165 2024-03-18 20:09:49 2024-03-18 20:09:49 2024-03-18 20:09:49 501  20 runner
#> 166 2024-03-18 20:09:49 2024-03-18 20:09:49 2024-03-18 20:09:49 501  20 runner
#> 167 2024-03-18 20:09:49 2024-03-18 20:09:49 2024-03-18 20:09:49 501  20 runner
#> 168 2024-03-18 20:09:50 2024-03-18 20:09:50 2024-03-18 20:09:50 501  20 runner
#> 169 2024-03-18 20:09:50 2024-03-18 20:09:50 2024-03-18 20:09:50 501  20 runner
#> 170 2024-03-18 20:09:51 2024-03-18 20:09:51 2024-03-18 20:09:51 501  20 runner
#> 171 2024-03-18 20:09:51 2024-03-18 20:09:51 2024-03-18 20:09:51 501  20 runner
#> 172 2024-03-18 20:09:51 2024-03-18 20:09:51 2024-03-18 20:09:51 501  20 runner
#> 173 2024-03-18 20:09:52 2024-03-18 20:09:52 2024-03-18 20:09:52 501  20 runner
#> 174 2024-03-18 20:09:52 2024-03-18 20:09:52 2024-03-18 20:09:52 501  20 runner
#> 175 2024-03-18 20:09:53 2024-03-18 20:09:53 2024-03-18 20:09:53 501  20 runner
#> 176 2024-03-18 20:09:53 2024-03-18 20:09:53 2024-03-18 20:09:53 501  20 runner
#> 177 2024-03-18 20:09:53 2024-03-18 20:09:53 2024-03-18 20:09:53 501  20 runner
#> 178 2024-03-18 20:09:54 2024-03-18 20:09:54 2024-03-18 20:09:54 501  20 runner
#> 179 2024-03-18 20:09:54 2024-03-18 20:09:54 2024-03-18 20:09:54 501  20 runner
#>     grname
#> 1    staff
#> 2    staff
#> 3    staff
#> 4    staff
#> 5    staff
#> 6    staff
#> 7    staff
#> 8    staff
#> 9    staff
#> 10   staff
#> 11   staff
#> 12   staff
#> 13   staff
#> 14   staff
#> 15   staff
#> 16   staff
#> 17   staff
#> 18   staff
#> 19   staff
#> 20   staff
#> 21   staff
#> 22   staff
#> 23   staff
#> 24   staff
#> 25   staff
#> 26   staff
#> 27   staff
#> 28   staff
#> 29   staff
#> 30   staff
#> 31   staff
#> 32   staff
#> 33   staff
#> 34   staff
#> 35   staff
#> 36   staff
#> 37   staff
#> 38   staff
#> 39   staff
#> 40   staff
#> 41   staff
#> 42   staff
#> 43   staff
#> 44   staff
#> 45   staff
#> 46   staff
#> 47   staff
#> 48   staff
#> 49   staff
#> 50   staff
#> 51   staff
#> 52   staff
#> 53   staff
#> 54   staff
#> 55   staff
#> 56   staff
#> 57   staff
#> 58   staff
#> 59   staff
#> 60   staff
#> 61   staff
#> 62   staff
#> 63   staff
#> 64   staff
#> 65   staff
#> 66   staff
#> 67   staff
#> 68   staff
#> 69   staff
#> 70   staff
#> 71   staff
#> 72   staff
#> 73   staff
#> 74   staff
#> 75   staff
#> 76   staff
#> 77   staff
#> 78   staff
#> 79   staff
#> 80   staff
#> 81   staff
#> 82   staff
#> 83   staff
#> 84   staff
#> 85   staff
#> 86   staff
#> 87   staff
#> 88   staff
#> 89   staff
#> 90   staff
#> 91   staff
#> 92   staff
#> 93   staff
#> 94   staff
#> 95   staff
#> 96   staff
#> 97   staff
#> 98   staff
#> 99   staff
#> 100  staff
#> 101  staff
#> 102  staff
#> 103  staff
#> 104  staff
#> 105  staff
#> 106  staff
#> 107  staff
#> 108  staff
#> 109  staff
#> 110  staff
#> 111  staff
#> 112  staff
#> 113  staff
#> 114  staff
#> 115  staff
#> 116  staff
#> 117  staff
#> 118  staff
#> 119  staff
#> 120  staff
#> 121  staff
#> 122  staff
#> 123  staff
#> 124  staff
#> 125  staff
#> 126  staff
#> 127  staff
#> 128  staff
#> 129  staff
#> 130  staff
#> 131  staff
#> 132  staff
#> 133  staff
#> 134  staff
#> 135  staff
#> 136  staff
#> 137  staff
#> 138  staff
#> 139  staff
#> 140  staff
#> 141  staff
#> 142  staff
#> 143  staff
#> 144  staff
#> 145  staff
#> 146  staff
#> 147  staff
#> 148  staff
#> 149  staff
#> 150  staff
#> 151  staff
#> 152  staff
#> 153  staff
#> 154  staff
#> 155  staff
#> 156  staff
#> 157  staff
#> 158  staff
#> 159  staff
#> 160  staff
#> 161  staff
#> 162  staff
#> 163  staff
#> 164  staff
#> 165  staff
#> 166  staff
#> 167  staff
#> 168  staff
#> 169  staff
#> 170  staff
#> 171  staff
#> 172  staff
#> 173  staff
#> 174  staff
#> 175  staff
#> 176  staff
#> 177  staff
#> 178  staff
#> 179  staff

# ... even when being called with an empty character vector
FUN(character())
#> listing 
#> [1] file  isdir
#> <0 rows> (or 0-length row.names)
FUN(character(), full_info = TRUE)
#> listing 
#>  [1] file   isdir  size   mode   mtime  ctime  atime  uid    gid    uname 
#> [11] grname
#> <0 rows> (or 0-length row.names)

# Call the function recursively up to the given depth level
kwb.utils:::listToDepth(".", max_depth = 1, FUN = FUN)
#> listing .
#> listing .
#>                                        file isdir
#> 1                               DIN.A4.html FALSE
#> 2                              Rplot001.png FALSE
#> 3                              Rplot002.png FALSE
#> 4                             addClass.html FALSE
#> 5                       addRowWithName.html FALSE
#> 6                   addSuffixToColumns.html FALSE
#> 7                          allAreEqual.html FALSE
#> 8                      allAreIdentical.html FALSE
#> 9                          almostEqual.html FALSE
#> 10                        appendSuffix.html FALSE
#> 11                            arglist-1.png FALSE
#> 12                            arglist-2.png FALSE
#> 13                             arglist.html FALSE
#> 14                             argsCsv.html FALSE
#> 15                    arrayToDataFrame.html FALSE
#> 16                        asColumnList.html FALSE
#> 17                 asNoFactorDataFrame.html FALSE
#> 18                           asRowList.html FALSE
#> 19                    assertFinalSlash.html FALSE
#> 20                assertRowsAndColumns.html FALSE
#> 21                           assignAll.html FALSE
#> 22              assignArgumentDefaults.html FALSE
#> 23                      assignGlobally.html FALSE
#> 24                       assignObjects.html FALSE
#> 25                assignPackageObjects.html FALSE
#> 26                     atLeastOneRowIn.html FALSE
#> 27                           backspace.html FALSE
#> 28                     breakInSequence.html FALSE
#> 29                           callWith-1.png FALSE
#> 30                           callWith-2.png FALSE
#> 31                            callWith.html FALSE
#> 32                        callWithData.html FALSE
#> 33            callWithStringsAsFactors.html FALSE
#> 34                           catAndRun.html FALSE
#> 35                          catChanges.html FALSE
#> 36                        catChangesIf.html FALSE
#> 37                               catIf.html FALSE
#> 38                            catLines.html FALSE
#> 39                        catNewLineIf.html FALSE
#> 40                        catchWarning.html FALSE
#> 41              checkForMissingColumns.html FALSE
#> 42                        clearConsole.html FALSE
#> 43                          clipMatrix.html FALSE
#> 44                         cmdLinePath.html FALSE
#> 45                           colMaxima.html FALSE
#> 46                           colMinima.html FALSE
#> 47                        colNaNumbers.html FALSE
#> 48             colStatisticOneFunction.html FALSE
#> 49                       colStatistics.html FALSE
#> 50                           collapsed.html FALSE
#> 51                    columnDescriptor.html FALSE
#> 52                        columnToDate.html FALSE
#> 53                columnwisePercentage.html FALSE
#> 54                combineAlternatingly.html FALSE
#> 55                      commaCollapsed.html FALSE
#> 56                         commonNames.html FALSE
#> 57                   compareDataFrames.html FALSE
#> 58                         compareSets.html FALSE
#> 59                   containsNulString.html FALSE
#> 60                      convertCsvFile.html FALSE
#> 61                      copyAttributes.html FALSE
#> 62              copyDirectoryStructure.html FALSE
#> 63                    copyListElements.html FALSE
#> 64                     countNaInColumn.html FALSE
#> 65                          countOrSum.html FALSE
#> 66                      createAccessor.html FALSE
#> 67              createDirAndReturnPath.html FALSE
#> 68                   createDirectories.html FALSE
#> 69                     createDirectory.html FALSE
#> 70         createFunctionAssignObjects.html FALSE
#> 71           createFunctionExtdataFile.html FALSE
#> 72                       createIdAlong.html FALSE
#> 73                        createMatrix.html FALSE
#> 74                  createPasswordFile.html FALSE
#> 75                       createStorage.html FALSE
#> 76                  csvTextToDataFrame.html FALSE
#> 77                              decode.html FALSE
#> 78                           defaultIf.html FALSE
#> 79                         defaultIfNA.html FALSE
#> 80                       defaultIfNULL.html FALSE
#> 81                       defaultIfZero.html FALSE
#> 82                       defaultLevels.html FALSE
#> 83        defaultWindowsProgramFolders.html FALSE
#> 84                             desktop.html FALSE
#> 85                            diffrows.html FALSE
#> 86                       directoryName.html FALSE
#> 87                          dot-OStype.html FALSE
#> 88                             dot-log.html FALSE
#> 89                         dot-logline.html FALSE
#> 90                           dot-logok.html FALSE
#> 91                        dot-logstart.html FALSE
#> 92                             dropDim.html FALSE
#> 93              dropUnusedFactorLevels.html FALSE
#> 94                              encode.html FALSE
#> 95                       enlargeVector.html FALSE
#> 96                         excludeNULL.html FALSE
#> 97                          expandGrid.html FALSE
#> 98                         extdataFile.html FALSE
#> 99                        extendLimits.html FALSE
#> 100                   extractRowRanges.html FALSE
#> 101                   extractSubstring.html FALSE
#> 102                      fileExtension.html FALSE
#> 103                        findChanges.html FALSE
#> 104              findPartialDuplicates.html FALSE
#> 105                   finishAndShowPdf.html FALSE
#> 106                 finishAndShowPdfIf.html FALSE
#> 107                       firstElement.html FALSE
#> 108                   firstPosixColumn.html FALSE
#> 109                      frenchToAscii.html FALSE
#> 110                     frequencyTable.html FALSE
#> 111                        fullWinPath.html FALSE
#> 112                        fullySorted.html FALSE
#> 113                    generateKeyFile.html FALSE
#> 114                       getAttribute.html FALSE
#> 115       getByPositiveOrNegativeIndex.html FALSE
#> 116                  getElementLengths.html FALSE
#> 117                     getEvenNumbers.html FALSE
#> 118          getFunctionValueOrDefault.html FALSE
#> 119                        getGlobally.html FALSE
#> 120                getKeywordPositions.html FALSE
#> 121                        getListNode.html FALSE
#> 122      getNamesOfObjectsInRDataFiles.html FALSE
#> 123             getObjectFromRDataFile.html FALSE
#> 124                      getOddNumbers.html FALSE
#> 125                        getPassword.html FALSE
#> 126 getPathsAndValuesFromRecursiveList.html FALSE
#> 127                        getTagNames.html FALSE
#> 128                        get_homedir.html FALSE
#> 129                   guessSeparator.1.html FALSE
#> 130                   guessSeparator.2.html FALSE
#> 131                     guessSeparator.html FALSE
#> 132                      hasFinalSlash.html FALSE
#> 133                      hasZeroLength.html FALSE
#> 134                           headtail.html FALSE
#> 135                   hsAddMissingCols.html FALSE
#> 136                         hsChrToNum.html FALSE
#> 137                       hsCountInStr.html FALSE
#> 138                     hsDelEmptyCols.html FALSE
#> 139                 hsMatrixToListForm.html FALSE
#> 140                      hsMovingMean-1.png FALSE
#> 141                       hsMovingMean.html FALSE
#> 142              hsOpenWindowsExplorer.html FALSE
#> 143                          hsPrepPdf.html FALSE
#> 144                         hsQuoteChr.html FALSE
#> 145                    hsRenameColumns.html FALSE
#> 146                          hsResolve.html FALSE
#> 147                hsRestoreAttributes.html FALSE
#> 148                         hsSafeName.html FALSE
#> 149                            hsShell.html FALSE
#> 150                          hsShowPdf.html FALSE
#> 151                     hsStringToDate.html FALSE
#> 152                   hsStringToDouble.html FALSE
#> 153                   hsSubstSpecChars.html FALSE
#> 154                           hsSystem.html FALSE
#> 155                             hsTrim.html FALSE
#> 156                       hsValidValue.html FALSE
#> 157                            inRange.html FALSE
#> 158                             indent.html FALSE
#> 159                              index.html FALSE
#> 160                      insertColumns.html FALSE
#> 161                 intToNumeralSystem.html FALSE
#> 162                         is.unnamed.html FALSE
#> 163                            isASCII.html FALSE
#> 164                   isDotOrDoubleDot.html FALSE
#> 165                       isEvenNumber.html FALSE
#> 166                           isLoaded.html FALSE
#> 167                   isNaInAllColumns.html FALSE
#> 168                      isNaInAllRows.html FALSE
#> 169                        isNaOrEmpty.html FALSE
#> 170                      isNetworkPath.html FALSE
#> 171                      isNullOrEmpty.html FALSE
#> 172                        isOddNumber.html FALSE
#> 173                         isTryError.html FALSE
#> 174                        lastElement.html FALSE
#> 175                               left.html FALSE
#> 176                leftSubstringEquals.html FALSE
#> 177                       limitToRange.html FALSE
#> 178                  linearCombination.html FALSE
#> 179                        listObjects.html FALSE