List Elements Recursively up to Depth

List Elements Recursively up to Depth

Usage

listToDepth(
  path,
  max_depth = 0L,
  full_info = FALSE,
  FUN = listFiles,
  ...,
  depth = 0,
  prob_mutate = 0,
  template = NULL
)

Arguments

path

path to the element at which to start listing

max_depth

maximal depth level of which to list elements. A value of 0 means non-recursive listing, a value of NA represents fully recursive listing.

full_info

return only path and isdir information or the full information provided by FUN(full_info = TRUE)?

FUN

function called to get the listing of the element given in path. The function must accept a path as its first argument and it must define the argument full_info second. It may accept further arguments. It must always return a data frame. For full_info = FALSE the data frame must have columns file and isdir (is the "file" a directory?). For full_info = TRUE the function may return further columns. The function must provide an empty data frame with the expected columns when being called with character() as the first argument. The function is expected to set the attribute "failed" to the given path in case that the path could not be accessed (e.g. because of a broken internet connection if the listing is done remotely). See kwb.utils:::listFiles for an example implementation that somehow simulates the behaviour of the dir function. See kwb.dwd::list_url() for a more advanced usage of this function in order to recursively list the files on an FTP server (FTP = file transfer protocol).

...

further arguments passed to FUN

depth

start depth of recursion if max_depth > 0. This argument is for internal use and not intended to be set by the user!

prob_mutate

probability to alter the path so that it becomes useless. This is zero by default. Set the value only if you want to test how the function behaves if the listing of a path fails.

template

empty data frame (zero rows) and with columns that are identical to the columns in the data frame returned by FUN. If not provided the function will call FUN once with a zero-length path argument expecting to retrieve the template that is expected here.

Value

data frame containing at least the columns file and isdir. If full_info = TRUE the result data frame may contain further columns, as provided by the function given in FUN for full_info = TRUE.

Examples

# Example list function provided in this package (file listing)
FUN <- kwb.utils:::listFiles

# The list function must provide empty records when no path is given. The 
# returned data frame must provide the columns "file" and "isdir" ...
FUN()
#> listing .
#>                                        file isdir
#> 1                               DIN.A4.html FALSE
#> 2                              Rplot001.png FALSE
#> 3                              Rplot002.png FALSE
#> 4                             addClass.html FALSE
#> 5                       addRowWithName.html FALSE
#> 6                   addSuffixToColumns.html FALSE
#> 7                          allAreEqual.html FALSE
#> 8                      allAreIdentical.html FALSE
#> 9                          almostEqual.html FALSE
#> 10                        appendSuffix.html FALSE
#> 11                            arglist-1.png FALSE
#> 12                            arglist-2.png FALSE
#> 13                             arglist.html FALSE
#> 14                             argsCsv.html FALSE
#> 15                    arrayToDataFrame.html FALSE
#> 16                        asColumnList.html FALSE
#> 17                 asNoFactorDataFrame.html FALSE
#> 18                           asRowList.html FALSE
#> 19                    assertFinalSlash.html FALSE
#> 20                assertRowsAndColumns.html FALSE
#> 21                           assignAll.html FALSE
#> 22              assignArgumentDefaults.html FALSE
#> 23                      assignGlobally.html FALSE
#> 24                assignPackageObjects.html FALSE
#> 25                     atLeastOneRowIn.html FALSE
#> 26                           backspace.html FALSE
#> 27                     breakInSequence.html FALSE
#> 28                           callWith-1.png FALSE
#> 29                           callWith-2.png FALSE
#> 30                            callWith.html FALSE
#> 31            callWithStringsAsFactors.html FALSE
#> 32                           catAndRun.html FALSE
#> 33                          catChanges.html FALSE
#> 34                        catChangesIf.html FALSE
#> 35                               catIf.html FALSE
#> 36                            catLines.html FALSE
#> 37                        catNewLineIf.html FALSE
#> 38              checkForMissingColumns.html FALSE
#> 39                        clearConsole.html FALSE
#> 40                          clipMatrix.html FALSE
#> 41                         cmdLinePath.html FALSE
#> 42                           colMaxima.html FALSE
#> 43                           colMinima.html FALSE
#> 44                        colNaNumbers.html FALSE
#> 45             colStatisticOneFunction.html FALSE
#> 46                       colStatistics.html FALSE
#> 47                           collapsed.html FALSE
#> 48                    columnDescriptor.html FALSE
#> 49                        columnToDate.html FALSE
#> 50                columnwisePercentage.html FALSE
#> 51                combineAlternatingly.html FALSE
#> 52                      commaCollapsed.html FALSE
#> 53                         commonNames.html FALSE
#> 54                   compareDataFrames.html FALSE
#> 55                         compareSets.html FALSE
#> 56                   containsNulString.html FALSE
#> 57                      convertCsvFile.html FALSE
#> 58                      copyAttributes.html FALSE
#> 59              copyDirectoryStructure.html FALSE
#> 60                    copyListElements.html FALSE
#> 61                     countNaInColumn.html FALSE
#> 62                          countOrSum.html FALSE
#> 63                      createAccessor.html FALSE
#> 64              createDirAndReturnPath.html FALSE
#> 65                   createDirectories.html FALSE
#> 66                     createDirectory.html FALSE
#> 67                       createIdAlong.html FALSE
#> 68                        createMatrix.html FALSE
#> 69                  createPasswordFile.html FALSE
#> 70                       createStorage.html FALSE
#> 71                  csvTextToDataFrame.html FALSE
#> 72                              decode.html FALSE
#> 73                         defaultIfNA.html FALSE
#> 74                       defaultIfNULL.html FALSE
#> 75                       defaultIfZero.html FALSE
#> 76                       defaultLevels.html FALSE
#> 77        defaultWindowsProgramFolders.html FALSE
#> 78                             desktop.html FALSE
#> 79                            diffrows.html FALSE
#> 80                          dot-OStype.html FALSE
#> 81                       dot-defaultIf.html FALSE
#> 82                dot-guessSeparator.1.html FALSE
#> 83                dot-guessSeparator.2.html FALSE
#> 84                   dot-isNetworkPath.html FALSE
#> 85                             dot-log.html FALSE
#> 86                         dot-logline.html FALSE
#> 87                           dot-logok.html FALSE
#> 88                        dot-logstart.html FALSE
#> 89                             dropDim.html FALSE
#> 90              dropUnusedFactorLevels.html FALSE
#> 91                              encode.html FALSE
#> 92                       enlargeVector.html FALSE
#> 93                         excludeNULL.html FALSE
#> 94                          expandGrid.html FALSE
#> 95                        extendLimits.html FALSE
#> 96                    extractRowRanges.html FALSE
#> 97                    extractSubstring.html FALSE
#> 98                       fileExtension.html FALSE
#> 99                         findChanges.html FALSE
#> 100              findPartialDuplicates.html FALSE
#> 101                   finishAndShowPdf.html FALSE
#> 102                 finishAndShowPdfIf.html FALSE
#> 103                       firstElement.html FALSE
#> 104                   firstPosixColumn.html FALSE
#> 105                      frenchToAscii.html FALSE
#> 106                     frequencyTable.html FALSE
#> 107                        fullySorted.html FALSE
#> 108                    generateKeyFile.html FALSE
#> 109                       getAttribute.html FALSE
#> 110       getByPositiveOrNegativeIndex.html FALSE
#> 111                  getElementLengths.html FALSE
#> 112                     getEvenNumbers.html FALSE
#> 113          getFunctionValueOrDefault.html FALSE
#> 114                        getGlobally.html FALSE
#> 115                getKeywordPositions.html FALSE
#> 116      getNamesOfObjectsInRDataFiles.html FALSE
#> 117             getObjectFromRDataFile.html FALSE
#> 118                      getOddNumbers.html FALSE
#> 119                        getPassword.html FALSE
#> 120 getPathsAndValuesFromRecursiveList.html FALSE
#> 121                        getTagNames.html FALSE
#> 122                        get_homedir.html FALSE
#> 123                     guessSeparator.html FALSE
#> 124                      hasFinalSlash.html FALSE
#> 125                           headtail.html FALSE
#> 126                   hsAddMissingCols.html FALSE
#> 127                         hsChrToNum.html FALSE
#> 128                       hsCountInStr.html FALSE
#> 129                     hsDelEmptyCols.html FALSE
#> 130                 hsMatrixToListForm.html FALSE
#> 131                      hsMovingMean-1.png FALSE
#> 132                       hsMovingMean.html FALSE
#> 133              hsOpenWindowsExplorer.html FALSE
#> 134                          hsPrepPdf.html FALSE
#> 135                         hsQuoteChr.html FALSE
#> 136                    hsRenameColumns.html FALSE
#> 137                          hsResolve.html FALSE
#> 138                hsRestoreAttributes.html FALSE
#> 139                         hsSafeName.html FALSE
#> 140                            hsShell.html FALSE
#> 141                          hsShowPdf.html FALSE
#> 142                     hsStringToDate.html FALSE
#> 143                   hsStringToDouble.html FALSE
#> 144                   hsSubstSpecChars.html FALSE
#> 145                           hsSystem.html FALSE
#> 146                             hsTrim.html FALSE
#> 147                       hsValidValue.html FALSE
#> 148                            inRange.html FALSE
#> 149                             indent.html FALSE
#> 150                              index.html FALSE
#> 151                      insertColumns.html FALSE
#> 152                 intToNumeralSystem.html FALSE
#> 153                         is.unnamed.html FALSE
#> 154                            isASCII.html FALSE
#> 155                   isDotOrDoubleDot.html FALSE
#> 156                       isEvenNumber.html FALSE
#> 157                           isLoaded.html FALSE
#> 158                   isNaInAllColumns.html FALSE
#> 159                      isNaInAllRows.html FALSE
#> 160                        isNaOrEmpty.html FALSE
#> 161                      isNullOrEmpty.html FALSE
#> 162                        isOddNumber.html FALSE
#> 163                         isTryError.html FALSE
#> 164                        lastElement.html FALSE
#> 165                               left.html FALSE
#> 166                leftSubstringEquals.html FALSE
#> 167                       limitToRange.html FALSE
#> 168                  linearCombination.html FALSE
#> 169                        listObjects.html FALSE
FUN(full_info = TRUE)
#> listing .
#>                                        file isdir   size mode
#> 1                               DIN.A4.html FALSE   8269  644
#> 2                              Rplot001.png FALSE   1011  644
#> 3                              Rplot002.png FALSE   5908  644
#> 4                             addClass.html FALSE   9407  644
#> 5                       addRowWithName.html FALSE   9123  644
#> 6                   addSuffixToColumns.html FALSE  11715  644
#> 7                          allAreEqual.html FALSE   9078  644
#> 8                      allAreIdentical.html FALSE   8911  644
#> 9                          almostEqual.html FALSE   9094  644
#> 10                        appendSuffix.html FALSE  10670  644
#> 11                            arglist-1.png FALSE  18498  644
#> 12                            arglist-2.png FALSE  18423  644
#> 13                             arglist.html FALSE  12892  644
#> 14                             argsCsv.html FALSE   8954  644
#> 15                    arrayToDataFrame.html FALSE  13474  644
#> 16                        asColumnList.html FALSE  10916  644
#> 17                 asNoFactorDataFrame.html FALSE  11226  644
#> 18                           asRowList.html FALSE  10800  644
#> 19                    assertFinalSlash.html FALSE  11833  644
#> 20                assertRowsAndColumns.html FALSE  12319  644
#> 21                           assignAll.html FALSE   8916  644
#> 22              assignArgumentDefaults.html FALSE  10232  644
#> 23                      assignGlobally.html FALSE   8606  644
#> 24                assignPackageObjects.html FALSE   9151  644
#> 25                     atLeastOneRowIn.html FALSE   8682  644
#> 26                           backspace.html FALSE  10383  644
#> 27                     breakInSequence.html FALSE   9057  644
#> 28                           callWith-1.png FALSE  18498  644
#> 29                           callWith-2.png FALSE  18423  644
#> 30                            callWith.html FALSE  11992  644
#> 31            callWithStringsAsFactors.html FALSE  14267  644
#> 32                           catAndRun.html FALSE  12652  644
#> 33                          catChanges.html FALSE   9520  644
#> 34                        catChangesIf.html FALSE   8978  644
#> 35                               catIf.html FALSE   8639  644
#> 36                            catLines.html FALSE   8744  644
#> 37                        catNewLineIf.html FALSE   8991  644
#> 38              checkForMissingColumns.html FALSE   9814  644
#> 39                        clearConsole.html FALSE   8201  644
#> 40                          clipMatrix.html FALSE   8932  644
#> 41                         cmdLinePath.html FALSE   8717  644
#> 42                           colMaxima.html FALSE   8729  644
#> 43                           colMinima.html FALSE   8729  644
#> 44                        colNaNumbers.html FALSE   8667  644
#> 45             colStatisticOneFunction.html FALSE   9154  644
#> 46                       colStatistics.html FALSE   9703  644
#> 47                           collapsed.html FALSE   8922  644
#> 48                    columnDescriptor.html FALSE   8833  644
#> 49                        columnToDate.html FALSE  11269  644
#> 50                columnwisePercentage.html FALSE  14579  644
#> 51                combineAlternatingly.html FALSE  11920  644
#> 52                      commaCollapsed.html FALSE   8547  644
#> 53                         commonNames.html FALSE  10060  644
#> 54                   compareDataFrames.html FALSE  13276  644
#> 55                         compareSets.html FALSE  11691  644
#> 56                   containsNulString.html FALSE   8852  644
#> 57                      convertCsvFile.html FALSE  21993  644
#> 58                      copyAttributes.html FALSE  10530  644
#> 59              copyDirectoryStructure.html FALSE   9680  644
#> 60                    copyListElements.html FALSE  11841  644
#> 61                     countNaInColumn.html FALSE   8916  644
#> 62                          countOrSum.html FALSE  13648  644
#> 63                      createAccessor.html FALSE  10649  644
#> 64              createDirAndReturnPath.html FALSE   9396  644
#> 65                   createDirectories.html FALSE   8889  644
#> 66                     createDirectory.html FALSE   9289  644
#> 67                       createIdAlong.html FALSE  11836  644
#> 68                        createMatrix.html FALSE  13675  644
#> 69                  createPasswordFile.html FALSE   9728  644
#> 70                       createStorage.html FALSE  14485  644
#> 71                  csvTextToDataFrame.html FALSE   8778  644
#> 72                              decode.html FALSE   8842  644
#> 73                         defaultIfNA.html FALSE  10404  644
#> 74                       defaultIfNULL.html FALSE  10427  644
#> 75                       defaultIfZero.html FALSE  11680  644
#> 76                       defaultLevels.html FALSE  10520  644
#> 77        defaultWindowsProgramFolders.html FALSE   8345  644
#> 78                             desktop.html FALSE   8907  644
#> 79                            diffrows.html FALSE  10722  644
#> 80                          dot-OStype.html FALSE   8594  644
#> 81                       dot-defaultIf.html FALSE   9272  644
#> 82                dot-guessSeparator.1.html FALSE   8949  644
#> 83                dot-guessSeparator.2.html FALSE   8825  644
#> 84                   dot-isNetworkPath.html FALSE   9150  644
#> 85                             dot-log.html FALSE   8642  644
#> 86                         dot-logline.html FALSE   8796  644
#> 87                           dot-logok.html FALSE   8732  644
#> 88                        dot-logstart.html FALSE   8925  644
#> 89                             dropDim.html FALSE  13671  644
#> 90              dropUnusedFactorLevels.html FALSE  13712  644
#> 91                              encode.html FALSE  11095  644
#> 92                       enlargeVector.html FALSE  11420  644
#> 93                         excludeNULL.html FALSE  10941  644
#> 94                          expandGrid.html FALSE  14477  644
#> 95                        extendLimits.html FALSE   9246  644
#> 96                    extractRowRanges.html FALSE  18779  644
#> 97                    extractSubstring.html FALSE  14940  644
#> 98                       fileExtension.html FALSE   9943  644
#> 99                         findChanges.html FALSE  13166  644
#> 100              findPartialDuplicates.html FALSE  13178  644
#> 101                   finishAndShowPdf.html FALSE   9392  644
#> 102                 finishAndShowPdfIf.html FALSE   9001  644
#> 103                       firstElement.html FALSE   8730  644
#> 104                   firstPosixColumn.html FALSE   8627  644
#> 105                      frenchToAscii.html FALSE   8648  644
#> 106                     frequencyTable.html FALSE  15582  644
#> 107                        fullySorted.html FALSE  14449  644
#> 108                    generateKeyFile.html FALSE   8949  644
#> 109                       getAttribute.html FALSE  13103  644
#> 110       getByPositiveOrNegativeIndex.html FALSE   9160  644
#> 111                  getElementLengths.html FALSE  10022  644
#> 112                     getEvenNumbers.html FALSE   8662  644
#> 113          getFunctionValueOrDefault.html FALSE   9190  644
#> 114                        getGlobally.html FALSE   8997  644
#> 115                getKeywordPositions.html FALSE   9413  644
#> 116      getNamesOfObjectsInRDataFiles.html FALSE   8797  644
#> 117             getObjectFromRDataFile.html FALSE   9346  644
#> 118                      getOddNumbers.html FALSE   8652  644
#> 119                        getPassword.html FALSE   9954  644
#> 120 getPathsAndValuesFromRecursiveList.html FALSE  12418  644
#> 121                        getTagNames.html FALSE   9645  644
#> 122                        get_homedir.html FALSE   9063  644
#> 123                     guessSeparator.html FALSE   9175  644
#> 124                      hasFinalSlash.html FALSE   9647  644
#> 125                           headtail.html FALSE  15022  644
#> 126                   hsAddMissingCols.html FALSE   9601  644
#> 127                         hsChrToNum.html FALSE   9557  644
#> 128                       hsCountInStr.html FALSE   8966  644
#> 129                     hsDelEmptyCols.html FALSE  10130  644
#> 130                 hsMatrixToListForm.html FALSE  10489  644
#> 131                      hsMovingMean-1.png FALSE 109898  644
#> 132                       hsMovingMean.html FALSE  12313  644
#> 133              hsOpenWindowsExplorer.html FALSE   9143  644
#> 134                          hsPrepPdf.html FALSE   9839  644
#> 135                         hsQuoteChr.html FALSE   9243  644
#> 136                    hsRenameColumns.html FALSE   8925  644
#> 137                          hsResolve.html FALSE   9473  644
#> 138                hsRestoreAttributes.html FALSE   8820  644
#> 139                         hsSafeName.html FALSE  11003  644
#> 140                            hsShell.html FALSE   8657  644
#> 141                          hsShowPdf.html FALSE  11962  644
#> 142                     hsStringToDate.html FALSE   9271  644
#> 143                   hsStringToDouble.html FALSE   9067  644
#> 144                   hsSubstSpecChars.html FALSE   9154  644
#> 145                           hsSystem.html FALSE   8669  644
#> 146                             hsTrim.html FALSE   9343  644
#> 147                       hsValidValue.html FALSE   9466  644
#> 148                            inRange.html FALSE   9229  644
#> 149                             indent.html FALSE   8895  644
#> 150                              index.html FALSE  63591  644
#> 151                      insertColumns.html FALSE  16122  644
#> 152                 intToNumeralSystem.html FALSE  13943  644
#> 153                         is.unnamed.html FALSE  11268  644
#> 154                            isASCII.html FALSE  10148  644
#> 155                   isDotOrDoubleDot.html FALSE   9625  644
#> 156                       isEvenNumber.html FALSE   8528  644
#> 157                           isLoaded.html FALSE  10766  644
#> 158                   isNaInAllColumns.html FALSE   8856  644
#> 159                      isNaInAllRows.html FALSE   8835  644
#> 160                        isNaOrEmpty.html FALSE   8969  644
#> 161                      isNullOrEmpty.html FALSE   8785  644
#> 162                        isOddNumber.html FALSE   8518  644
#> 163                         isTryError.html FALSE   9795  644
#> 164                        lastElement.html FALSE   8869  644
#> 165                               left.html FALSE   9298  644
#> 166                leftSubstringEquals.html FALSE   9885  644
#> 167                       limitToRange.html FALSE   9721  644
#> 168                  linearCombination.html FALSE  12064  644
#> 169                        listObjects.html FALSE  10053  644
#>                   mtime               ctime               atime uid gid  uname
#> 1   2022-06-04 19:56:45 2022-06-04 19:56:45 2022-06-04 19:56:45 501  20 runner
#> 2   2022-06-04 19:58:06 2022-06-04 19:58:06 2022-06-04 19:56:46 501  20 runner
#> 3   2022-06-04 19:56:57 2022-06-04 19:56:57 2022-06-04 19:56:49 501  20 runner
#> 4   2022-06-04 19:56:45 2022-06-04 19:56:45 2022-06-04 19:56:45 501  20 runner
#> 5   2022-06-04 19:56:46 2022-06-04 19:56:46 2022-06-04 19:56:46 501  20 runner
#> 6   2022-06-04 19:56:47 2022-06-04 19:56:47 2022-06-04 19:56:47 501  20 runner
#> 7   2022-06-04 19:56:47 2022-06-04 19:56:47 2022-06-04 19:56:47 501  20 runner
#> 8   2022-06-04 19:56:48 2022-06-04 19:56:48 2022-06-04 19:56:48 501  20 runner
#> 9   2022-06-04 19:56:48 2022-06-04 19:56:48 2022-06-04 19:56:48 501  20 runner
#> 10  2022-06-04 19:56:49 2022-06-04 19:56:49 2022-06-04 19:56:49 501  20 runner
#> 11  2022-06-04 19:56:49 2022-06-04 19:56:49 2022-06-04 19:56:49 501  20 runner
#> 12  2022-06-04 19:56:49 2022-06-04 19:56:49 2022-06-04 19:56:49 501  20 runner
#> 13  2022-06-04 19:56:49 2022-06-04 19:56:49 2022-06-04 19:56:49 501  20 runner
#> 14  2022-06-04 19:56:50 2022-06-04 19:56:50 2022-06-04 19:56:50 501  20 runner
#> 15  2022-06-04 19:56:50 2022-06-04 19:56:50 2022-06-04 19:56:50 501  20 runner
#> 16  2022-06-04 19:56:51 2022-06-04 19:56:51 2022-06-04 19:56:51 501  20 runner
#> 17  2022-06-04 19:56:52 2022-06-04 19:56:52 2022-06-04 19:56:52 501  20 runner
#> 18  2022-06-04 19:56:52 2022-06-04 19:56:52 2022-06-04 19:56:52 501  20 runner
#> 19  2022-06-04 19:56:53 2022-06-04 19:56:53 2022-06-04 19:56:53 501  20 runner
#> 20  2022-06-04 19:56:53 2022-06-04 19:56:53 2022-06-04 19:56:53 501  20 runner
#> 21  2022-06-04 19:56:53 2022-06-04 19:56:53 2022-06-04 19:56:53 501  20 runner
#> 22  2022-06-04 19:56:54 2022-06-04 19:56:54 2022-06-04 19:56:54 501  20 runner
#> 23  2022-06-04 19:56:54 2022-06-04 19:56:54 2022-06-04 19:56:54 501  20 runner
#> 24  2022-06-04 19:56:55 2022-06-04 19:56:55 2022-06-04 19:56:55 501  20 runner
#> 25  2022-06-04 19:56:56 2022-06-04 19:56:56 2022-06-04 19:56:56 501  20 runner
#> 26  2022-06-04 19:56:56 2022-06-04 19:56:56 2022-06-04 19:56:56 501  20 runner
#> 27  2022-06-04 19:56:57 2022-06-04 19:56:57 2022-06-04 19:56:57 501  20 runner
#> 28  2022-06-04 19:56:57 2022-06-04 19:56:57 2022-06-04 19:56:57 501  20 runner
#> 29  2022-06-04 19:56:57 2022-06-04 19:56:57 2022-06-04 19:56:57 501  20 runner
#> 30  2022-06-04 19:56:57 2022-06-04 19:56:57 2022-06-04 19:56:57 501  20 runner
#> 31  2022-06-04 19:56:58 2022-06-04 19:56:58 2022-06-04 19:56:58 501  20 runner
#> 32  2022-06-04 19:56:59 2022-06-04 19:56:59 2022-06-04 19:56:59 501  20 runner
#> 33  2022-06-04 19:56:59 2022-06-04 19:56:59 2022-06-04 19:56:59 501  20 runner
#> 34  2022-06-04 19:56:59 2022-06-04 19:56:59 2022-06-04 19:56:59 501  20 runner
#> 35  2022-06-04 19:57:00 2022-06-04 19:57:00 2022-06-04 19:57:00 501  20 runner
#> 36  2022-06-04 19:57:00 2022-06-04 19:57:00 2022-06-04 19:57:00 501  20 runner
#> 37  2022-06-04 19:57:01 2022-06-04 19:57:01 2022-06-04 19:57:01 501  20 runner
#> 38  2022-06-04 19:57:01 2022-06-04 19:57:01 2022-06-04 19:57:01 501  20 runner
#> 39  2022-06-04 19:57:02 2022-06-04 19:57:02 2022-06-04 19:57:02 501  20 runner
#> 40  2022-06-04 19:57:03 2022-06-04 19:57:03 2022-06-04 19:57:03 501  20 runner
#> 41  2022-06-04 19:57:03 2022-06-04 19:57:03 2022-06-04 19:57:03 501  20 runner
#> 42  2022-06-04 19:57:04 2022-06-04 19:57:04 2022-06-04 19:57:04 501  20 runner
#> 43  2022-06-04 19:57:04 2022-06-04 19:57:04 2022-06-04 19:57:04 501  20 runner
#> 44  2022-06-04 19:57:05 2022-06-04 19:57:05 2022-06-04 19:57:05 501  20 runner
#> 45  2022-06-04 19:57:05 2022-06-04 19:57:05 2022-06-04 19:57:05 501  20 runner
#> 46  2022-06-04 19:57:05 2022-06-04 19:57:05 2022-06-04 19:57:05 501  20 runner
#> 47  2022-06-04 19:57:06 2022-06-04 19:57:06 2022-06-04 19:57:06 501  20 runner
#> 48  2022-06-04 19:57:06 2022-06-04 19:57:06 2022-06-04 19:57:06 501  20 runner
#> 49  2022-06-04 19:57:07 2022-06-04 19:57:07 2022-06-04 19:57:07 501  20 runner
#> 50  2022-06-04 19:57:08 2022-06-04 19:57:08 2022-06-04 19:57:08 501  20 runner
#> 51  2022-06-04 19:57:08 2022-06-04 19:57:08 2022-06-04 19:57:08 501  20 runner
#> 52  2022-06-04 19:57:09 2022-06-04 19:57:09 2022-06-04 19:57:09 501  20 runner
#> 53  2022-06-04 19:57:09 2022-06-04 19:57:09 2022-06-04 19:57:09 501  20 runner
#> 54  2022-06-04 19:57:10 2022-06-04 19:57:10 2022-06-04 19:57:10 501  20 runner
#> 55  2022-06-04 19:57:10 2022-06-04 19:57:10 2022-06-04 19:57:10 501  20 runner
#> 56  2022-06-04 19:57:11 2022-06-04 19:57:11 2022-06-04 19:57:11 501  20 runner
#> 57  2022-06-04 19:57:11 2022-06-04 19:57:11 2022-06-04 19:57:11 501  20 runner
#> 58  2022-06-04 19:57:12 2022-06-04 19:57:12 2022-06-04 19:57:12 501  20 runner
#> 59  2022-06-04 19:57:12 2022-06-04 19:57:12 2022-06-04 19:57:12 501  20 runner
#> 60  2022-06-04 19:57:13 2022-06-04 19:57:13 2022-06-04 19:57:13 501  20 runner
#> 61  2022-06-04 19:57:13 2022-06-04 19:57:13 2022-06-04 19:57:13 501  20 runner
#> 62  2022-06-04 19:57:14 2022-06-04 19:57:14 2022-06-04 19:57:14 501  20 runner
#> 63  2022-06-04 19:57:15 2022-06-04 19:57:15 2022-06-04 19:57:15 501  20 runner
#> 64  2022-06-04 19:57:15 2022-06-04 19:57:15 2022-06-04 19:57:15 501  20 runner
#> 65  2022-06-04 19:57:16 2022-06-04 19:57:16 2022-06-04 19:57:16 501  20 runner
#> 66  2022-06-04 19:57:16 2022-06-04 19:57:16 2022-06-04 19:57:16 501  20 runner
#> 67  2022-06-04 19:57:17 2022-06-04 19:57:17 2022-06-04 19:57:17 501  20 runner
#> 68  2022-06-04 19:57:17 2022-06-04 19:57:17 2022-06-04 19:57:17 501  20 runner
#> 69  2022-06-04 19:57:18 2022-06-04 19:57:18 2022-06-04 19:57:18 501  20 runner
#> 70  2022-06-04 19:57:18 2022-06-04 19:57:18 2022-06-04 19:57:18 501  20 runner
#> 71  2022-06-04 19:57:19 2022-06-04 19:57:19 2022-06-04 19:57:19 501  20 runner
#> 72  2022-06-04 19:57:19 2022-06-04 19:57:19 2022-06-04 19:57:19 501  20 runner
#> 73  2022-06-04 19:57:20 2022-06-04 19:57:20 2022-06-04 19:57:20 501  20 runner
#> 74  2022-06-04 19:57:20 2022-06-04 19:57:20 2022-06-04 19:57:20 501  20 runner
#> 75  2022-06-04 19:57:21 2022-06-04 19:57:21 2022-06-04 19:57:21 501  20 runner
#> 76  2022-06-04 19:57:21 2022-06-04 19:57:21 2022-06-04 19:57:21 501  20 runner
#> 77  2022-06-04 19:57:22 2022-06-04 19:57:22 2022-06-04 19:57:22 501  20 runner
#> 78  2022-06-04 19:57:22 2022-06-04 19:57:22 2022-06-04 19:57:22 501  20 runner
#> 79  2022-06-04 19:57:23 2022-06-04 19:57:23 2022-06-04 19:57:23 501  20 runner
#> 80  2022-06-04 19:57:23 2022-06-04 19:57:23 2022-06-04 19:57:23 501  20 runner
#> 81  2022-06-04 19:57:24 2022-06-04 19:57:24 2022-06-04 19:57:24 501  20 runner
#> 82  2022-06-04 19:57:24 2022-06-04 19:57:24 2022-06-04 19:57:24 501  20 runner
#> 83  2022-06-04 19:57:24 2022-06-04 19:57:24 2022-06-04 19:57:24 501  20 runner
#> 84  2022-06-04 19:57:25 2022-06-04 19:57:25 2022-06-04 19:57:25 501  20 runner
#> 85  2022-06-04 19:57:25 2022-06-04 19:57:25 2022-06-04 19:57:25 501  20 runner
#> 86  2022-06-04 19:57:26 2022-06-04 19:57:26 2022-06-04 19:57:26 501  20 runner
#> 87  2022-06-04 19:57:26 2022-06-04 19:57:26 2022-06-04 19:57:26 501  20 runner
#> 88  2022-06-04 19:57:27 2022-06-04 19:57:27 2022-06-04 19:57:27 501  20 runner
#> 89  2022-06-04 19:57:27 2022-06-04 19:57:27 2022-06-04 19:57:27 501  20 runner
#> 90  2022-06-04 19:57:28 2022-06-04 19:57:28 2022-06-04 19:57:28 501  20 runner
#> 91  2022-06-04 19:57:28 2022-06-04 19:57:28 2022-06-04 19:57:28 501  20 runner
#> 92  2022-06-04 19:57:29 2022-06-04 19:57:29 2022-06-04 19:57:29 501  20 runner
#> 93  2022-06-04 19:57:29 2022-06-04 19:57:29 2022-06-04 19:57:29 501  20 runner
#> 94  2022-06-04 19:57:30 2022-06-04 19:57:30 2022-06-04 19:57:30 501  20 runner
#> 95  2022-06-04 19:57:30 2022-06-04 19:57:30 2022-06-04 19:57:30 501  20 runner
#> 96  2022-06-04 19:57:31 2022-06-04 19:57:31 2022-06-04 19:57:31 501  20 runner
#> 97  2022-06-04 19:57:31 2022-06-04 19:57:31 2022-06-04 19:57:31 501  20 runner
#> 98  2022-06-04 19:57:32 2022-06-04 19:57:32 2022-06-04 19:57:32 501  20 runner
#> 99  2022-06-04 19:57:32 2022-06-04 19:57:32 2022-06-04 19:57:32 501  20 runner
#> 100 2022-06-04 19:57:33 2022-06-04 19:57:33 2022-06-04 19:57:33 501  20 runner
#> 101 2022-06-04 19:57:33 2022-06-04 19:57:33 2022-06-04 19:57:33 501  20 runner
#> 102 2022-06-04 19:57:34 2022-06-04 19:57:34 2022-06-04 19:57:34 501  20 runner
#> 103 2022-06-04 19:57:34 2022-06-04 19:57:34 2022-06-04 19:57:34 501  20 runner
#> 104 2022-06-04 19:57:35 2022-06-04 19:57:35 2022-06-04 19:57:35 501  20 runner
#> 105 2022-06-04 19:57:35 2022-06-04 19:57:35 2022-06-04 19:57:35 501  20 runner
#> 106 2022-06-04 19:57:35 2022-06-04 19:57:35 2022-06-04 19:57:35 501  20 runner
#> 107 2022-06-04 19:57:36 2022-06-04 19:57:36 2022-06-04 19:57:36 501  20 runner
#> 108 2022-06-04 19:57:36 2022-06-04 19:57:36 2022-06-04 19:57:36 501  20 runner
#> 109 2022-06-04 19:57:37 2022-06-04 19:57:37 2022-06-04 19:57:37 501  20 runner
#> 110 2022-06-04 19:57:37 2022-06-04 19:57:37 2022-06-04 19:57:37 501  20 runner
#> 111 2022-06-04 19:57:38 2022-06-04 19:57:38 2022-06-04 19:57:38 501  20 runner
#> 112 2022-06-04 19:57:38 2022-06-04 19:57:38 2022-06-04 19:57:38 501  20 runner
#> 113 2022-06-04 19:57:39 2022-06-04 19:57:39 2022-06-04 19:57:39 501  20 runner
#> 114 2022-06-04 19:57:39 2022-06-04 19:57:39 2022-06-04 19:57:39 501  20 runner
#> 115 2022-06-04 19:57:40 2022-06-04 19:57:40 2022-06-04 19:57:40 501  20 runner
#> 116 2022-06-04 19:57:40 2022-06-04 19:57:40 2022-06-04 19:57:40 501  20 runner
#> 117 2022-06-04 19:57:41 2022-06-04 19:57:41 2022-06-04 19:57:41 501  20 runner
#> 118 2022-06-04 19:57:41 2022-06-04 19:57:41 2022-06-04 19:57:41 501  20 runner
#> 119 2022-06-04 19:57:42 2022-06-04 19:57:42 2022-06-04 19:57:42 501  20 runner
#> 120 2022-06-04 19:57:42 2022-06-04 19:57:42 2022-06-04 19:57:42 501  20 runner
#> 121 2022-06-04 19:57:43 2022-06-04 19:57:43 2022-06-04 19:57:43 501  20 runner
#> 122 2022-06-04 19:57:43 2022-06-04 19:57:43 2022-06-04 19:57:43 501  20 runner
#> 123 2022-06-04 19:57:44 2022-06-04 19:57:44 2022-06-04 19:57:44 501  20 runner
#> 124 2022-06-04 19:57:44 2022-06-04 19:57:44 2022-06-04 19:57:44 501  20 runner
#> 125 2022-06-04 19:57:44 2022-06-04 19:57:44 2022-06-04 19:57:44 501  20 runner
#> 126 2022-06-04 19:57:45 2022-06-04 19:57:45 2022-06-04 19:57:45 501  20 runner
#> 127 2022-06-04 19:57:45 2022-06-04 19:57:45 2022-06-04 19:57:45 501  20 runner
#> 128 2022-06-04 19:57:46 2022-06-04 19:57:46 2022-06-04 19:57:46 501  20 runner
#> 129 2022-06-04 19:57:46 2022-06-04 19:57:46 2022-06-04 19:57:46 501  20 runner
#> 130 2022-06-04 19:57:47 2022-06-04 19:57:47 2022-06-04 19:57:47 501  20 runner
#> 131 2022-06-04 19:57:48 2022-06-04 19:57:48 2022-06-04 19:57:48 501  20 runner
#> 132 2022-06-04 19:57:49 2022-06-04 19:57:49 2022-06-04 19:57:49 501  20 runner
#> 133 2022-06-04 19:57:49 2022-06-04 19:57:49 2022-06-04 19:57:49 501  20 runner
#> 134 2022-06-04 19:57:50 2022-06-04 19:57:50 2022-06-04 19:57:50 501  20 runner
#> 135 2022-06-04 19:57:50 2022-06-04 19:57:50 2022-06-04 19:57:50 501  20 runner
#> 136 2022-06-04 19:57:51 2022-06-04 19:57:51 2022-06-04 19:57:51 501  20 runner
#> 137 2022-06-04 19:57:51 2022-06-04 19:57:51 2022-06-04 19:57:51 501  20 runner
#> 138 2022-06-04 19:57:52 2022-06-04 19:57:52 2022-06-04 19:57:52 501  20 runner
#> 139 2022-06-04 19:57:52 2022-06-04 19:57:52 2022-06-04 19:57:52 501  20 runner
#> 140 2022-06-04 19:57:53 2022-06-04 19:57:53 2022-06-04 19:57:53 501  20 runner
#> 141 2022-06-04 19:57:53 2022-06-04 19:57:53 2022-06-04 19:57:53 501  20 runner
#> 142 2022-06-04 19:57:54 2022-06-04 19:57:54 2022-06-04 19:57:54 501  20 runner
#> 143 2022-06-04 19:57:54 2022-06-04 19:57:54 2022-06-04 19:57:54 501  20 runner
#> 144 2022-06-04 19:57:54 2022-06-04 19:57:54 2022-06-04 19:57:54 501  20 runner
#> 145 2022-06-04 19:57:55 2022-06-04 19:57:55 2022-06-04 19:57:55 501  20 runner
#> 146 2022-06-04 19:57:55 2022-06-04 19:57:55 2022-06-04 19:57:55 501  20 runner
#> 147 2022-06-04 19:57:56 2022-06-04 19:57:56 2022-06-04 19:57:56 501  20 runner
#> 148 2022-06-04 19:57:56 2022-06-04 19:57:56 2022-06-04 19:57:56 501  20 runner
#> 149 2022-06-04 19:57:57 2022-06-04 19:57:57 2022-06-04 19:57:57 501  20 runner
#> 150 2022-06-04 19:56:44 2022-06-04 19:56:44 2022-06-04 19:56:44 501  20 runner
#> 151 2022-06-04 19:57:57 2022-06-04 19:57:57 2022-06-04 19:57:57 501  20 runner
#> 152 2022-06-04 19:57:58 2022-06-04 19:57:58 2022-06-04 19:57:58 501  20 runner
#> 153 2022-06-04 19:57:58 2022-06-04 19:57:58 2022-06-04 19:57:58 501  20 runner
#> 154 2022-06-04 19:57:59 2022-06-04 19:57:59 2022-06-04 19:57:59 501  20 runner
#> 155 2022-06-04 19:57:59 2022-06-04 19:57:59 2022-06-04 19:57:59 501  20 runner
#> 156 2022-06-04 19:58:00 2022-06-04 19:58:00 2022-06-04 19:58:00 501  20 runner
#> 157 2022-06-04 19:58:01 2022-06-04 19:58:01 2022-06-04 19:58:01 501  20 runner
#> 158 2022-06-04 19:58:01 2022-06-04 19:58:01 2022-06-04 19:58:01 501  20 runner
#> 159 2022-06-04 19:58:01 2022-06-04 19:58:01 2022-06-04 19:58:01 501  20 runner
#> 160 2022-06-04 19:58:02 2022-06-04 19:58:02 2022-06-04 19:58:02 501  20 runner
#> 161 2022-06-04 19:58:02 2022-06-04 19:58:02 2022-06-04 19:58:02 501  20 runner
#> 162 2022-06-04 19:58:03 2022-06-04 19:58:03 2022-06-04 19:58:03 501  20 runner
#> 163 2022-06-04 19:58:03 2022-06-04 19:58:03 2022-06-04 19:58:03 501  20 runner
#> 164 2022-06-04 19:58:04 2022-06-04 19:58:04 2022-06-04 19:58:04 501  20 runner
#> 165 2022-06-04 19:58:04 2022-06-04 19:58:04 2022-06-04 19:58:04 501  20 runner
#> 166 2022-06-04 19:58:05 2022-06-04 19:58:05 2022-06-04 19:58:05 501  20 runner
#> 167 2022-06-04 19:58:05 2022-06-04 19:58:05 2022-06-04 19:58:05 501  20 runner
#> 168 2022-06-04 19:58:06 2022-06-04 19:58:06 2022-06-04 19:58:06 501  20 runner
#> 169 2022-06-04 19:58:06 2022-06-04 19:58:06 2022-06-04 19:58:06 501  20 runner
#>     grname
#> 1    staff
#> 2    staff
#> 3    staff
#> 4    staff
#> 5    staff
#> 6    staff
#> 7    staff
#> 8    staff
#> 9    staff
#> 10   staff
#> 11   staff
#> 12   staff
#> 13   staff
#> 14   staff
#> 15   staff
#> 16   staff
#> 17   staff
#> 18   staff
#> 19   staff
#> 20   staff
#> 21   staff
#> 22   staff
#> 23   staff
#> 24   staff
#> 25   staff
#> 26   staff
#> 27   staff
#> 28   staff
#> 29   staff
#> 30   staff
#> 31   staff
#> 32   staff
#> 33   staff
#> 34   staff
#> 35   staff
#> 36   staff
#> 37   staff
#> 38   staff
#> 39   staff
#> 40   staff
#> 41   staff
#> 42   staff
#> 43   staff
#> 44   staff
#> 45   staff
#> 46   staff
#> 47   staff
#> 48   staff
#> 49   staff
#> 50   staff
#> 51   staff
#> 52   staff
#> 53   staff
#> 54   staff
#> 55   staff
#> 56   staff
#> 57   staff
#> 58   staff
#> 59   staff
#> 60   staff
#> 61   staff
#> 62   staff
#> 63   staff
#> 64   staff
#> 65   staff
#> 66   staff
#> 67   staff
#> 68   staff
#> 69   staff
#> 70   staff
#> 71   staff
#> 72   staff
#> 73   staff
#> 74   staff
#> 75   staff
#> 76   staff
#> 77   staff
#> 78   staff
#> 79   staff
#> 80   staff
#> 81   staff
#> 82   staff
#> 83   staff
#> 84   staff
#> 85   staff
#> 86   staff
#> 87   staff
#> 88   staff
#> 89   staff
#> 90   staff
#> 91   staff
#> 92   staff
#> 93   staff
#> 94   staff
#> 95   staff
#> 96   staff
#> 97   staff
#> 98   staff
#> 99   staff
#> 100  staff
#> 101  staff
#> 102  staff
#> 103  staff
#> 104  staff
#> 105  staff
#> 106  staff
#> 107  staff
#> 108  staff
#> 109  staff
#> 110  staff
#> 111  staff
#> 112  staff
#> 113  staff
#> 114  staff
#> 115  staff
#> 116  staff
#> 117  staff
#> 118  staff
#> 119  staff
#> 120  staff
#> 121  staff
#> 122  staff
#> 123  staff
#> 124  staff
#> 125  staff
#> 126  staff
#> 127  staff
#> 128  staff
#> 129  staff
#> 130  staff
#> 131  staff
#> 132  staff
#> 133  staff
#> 134  staff
#> 135  staff
#> 136  staff
#> 137  staff
#> 138  staff
#> 139  staff
#> 140  staff
#> 141  staff
#> 142  staff
#> 143  staff
#> 144  staff
#> 145  staff
#> 146  staff
#> 147  staff
#> 148  staff
#> 149  staff
#> 150  staff
#> 151  staff
#> 152  staff
#> 153  staff
#> 154  staff
#> 155  staff
#> 156  staff
#> 157  staff
#> 158  staff
#> 159  staff
#> 160  staff
#> 161  staff
#> 162  staff
#> 163  staff
#> 164  staff
#> 165  staff
#> 166  staff
#> 167  staff
#> 168  staff
#> 169  staff

# ... even when being called with an empty character vector
FUN(character())
#> listing 
#> listing .
#> [1] file  isdir
#> <0 rows> (or 0-length row.names)
FUN(character(), full_info = TRUE)
#> listing 
#> listing .
#>  [1] file   isdir  size   mode   mtime  ctime  atime  uid    gid    uname 
#> [11] grname
#> <0 rows> (or 0-length row.names)

# Call the function recursively up to the given depth level
kwb.utils:::listToDepth(".", max_depth = 1, FUN = FUN)
#> listing .
#> listing .
#>                                        file isdir
#> 1                               DIN.A4.html FALSE
#> 2                              Rplot001.png FALSE
#> 3                              Rplot002.png FALSE
#> 4                             addClass.html FALSE
#> 5                       addRowWithName.html FALSE
#> 6                   addSuffixToColumns.html FALSE
#> 7                          allAreEqual.html FALSE
#> 8                      allAreIdentical.html FALSE
#> 9                          almostEqual.html FALSE
#> 10                        appendSuffix.html FALSE
#> 11                            arglist-1.png FALSE
#> 12                            arglist-2.png FALSE
#> 13                             arglist.html FALSE
#> 14                             argsCsv.html FALSE
#> 15                    arrayToDataFrame.html FALSE
#> 16                        asColumnList.html FALSE
#> 17                 asNoFactorDataFrame.html FALSE
#> 18                           asRowList.html FALSE
#> 19                    assertFinalSlash.html FALSE
#> 20                assertRowsAndColumns.html FALSE
#> 21                           assignAll.html FALSE
#> 22              assignArgumentDefaults.html FALSE
#> 23                      assignGlobally.html FALSE
#> 24                assignPackageObjects.html FALSE
#> 25                     atLeastOneRowIn.html FALSE
#> 26                           backspace.html FALSE
#> 27                     breakInSequence.html FALSE
#> 28                           callWith-1.png FALSE
#> 29                           callWith-2.png FALSE
#> 30                            callWith.html FALSE
#> 31            callWithStringsAsFactors.html FALSE
#> 32                           catAndRun.html FALSE
#> 33                          catChanges.html FALSE
#> 34                        catChangesIf.html FALSE
#> 35                               catIf.html FALSE
#> 36                            catLines.html FALSE
#> 37                        catNewLineIf.html FALSE
#> 38              checkForMissingColumns.html FALSE
#> 39                        clearConsole.html FALSE
#> 40                          clipMatrix.html FALSE
#> 41                         cmdLinePath.html FALSE
#> 42                           colMaxima.html FALSE
#> 43                           colMinima.html FALSE
#> 44                        colNaNumbers.html FALSE
#> 45             colStatisticOneFunction.html FALSE
#> 46                       colStatistics.html FALSE
#> 47                           collapsed.html FALSE
#> 48                    columnDescriptor.html FALSE
#> 49                        columnToDate.html FALSE
#> 50                columnwisePercentage.html FALSE
#> 51                combineAlternatingly.html FALSE
#> 52                      commaCollapsed.html FALSE
#> 53                         commonNames.html FALSE
#> 54                   compareDataFrames.html FALSE
#> 55                         compareSets.html FALSE
#> 56                   containsNulString.html FALSE
#> 57                      convertCsvFile.html FALSE
#> 58                      copyAttributes.html FALSE
#> 59              copyDirectoryStructure.html FALSE
#> 60                    copyListElements.html FALSE
#> 61                     countNaInColumn.html FALSE
#> 62                          countOrSum.html FALSE
#> 63                      createAccessor.html FALSE
#> 64              createDirAndReturnPath.html FALSE
#> 65                   createDirectories.html FALSE
#> 66                     createDirectory.html FALSE
#> 67                       createIdAlong.html FALSE
#> 68                        createMatrix.html FALSE
#> 69                  createPasswordFile.html FALSE
#> 70                       createStorage.html FALSE
#> 71                  csvTextToDataFrame.html FALSE
#> 72                              decode.html FALSE
#> 73                         defaultIfNA.html FALSE
#> 74                       defaultIfNULL.html FALSE
#> 75                       defaultIfZero.html FALSE
#> 76                       defaultLevels.html FALSE
#> 77        defaultWindowsProgramFolders.html FALSE
#> 78                             desktop.html FALSE
#> 79                            diffrows.html FALSE
#> 80                          dot-OStype.html FALSE
#> 81                       dot-defaultIf.html FALSE
#> 82                dot-guessSeparator.1.html FALSE
#> 83                dot-guessSeparator.2.html FALSE
#> 84                   dot-isNetworkPath.html FALSE
#> 85                             dot-log.html FALSE
#> 86                         dot-logline.html FALSE
#> 87                           dot-logok.html FALSE
#> 88                        dot-logstart.html FALSE
#> 89                             dropDim.html FALSE
#> 90              dropUnusedFactorLevels.html FALSE
#> 91                              encode.html FALSE
#> 92                       enlargeVector.html FALSE
#> 93                         excludeNULL.html FALSE
#> 94                          expandGrid.html FALSE
#> 95                        extendLimits.html FALSE
#> 96                    extractRowRanges.html FALSE
#> 97                    extractSubstring.html FALSE
#> 98                       fileExtension.html FALSE
#> 99                         findChanges.html FALSE
#> 100              findPartialDuplicates.html FALSE
#> 101                   finishAndShowPdf.html FALSE
#> 102                 finishAndShowPdfIf.html FALSE
#> 103                       firstElement.html FALSE
#> 104                   firstPosixColumn.html FALSE
#> 105                      frenchToAscii.html FALSE
#> 106                     frequencyTable.html FALSE
#> 107                        fullySorted.html FALSE
#> 108                    generateKeyFile.html FALSE
#> 109                       getAttribute.html FALSE
#> 110       getByPositiveOrNegativeIndex.html FALSE
#> 111                  getElementLengths.html FALSE
#> 112                     getEvenNumbers.html FALSE
#> 113          getFunctionValueOrDefault.html FALSE
#> 114                        getGlobally.html FALSE
#> 115                getKeywordPositions.html FALSE
#> 116      getNamesOfObjectsInRDataFiles.html FALSE
#> 117             getObjectFromRDataFile.html FALSE
#> 118                      getOddNumbers.html FALSE
#> 119                        getPassword.html FALSE
#> 120 getPathsAndValuesFromRecursiveList.html FALSE
#> 121                        getTagNames.html FALSE
#> 122                        get_homedir.html FALSE
#> 123                     guessSeparator.html FALSE
#> 124                      hasFinalSlash.html FALSE
#> 125                           headtail.html FALSE
#> 126                   hsAddMissingCols.html FALSE
#> 127                         hsChrToNum.html FALSE
#> 128                       hsCountInStr.html FALSE
#> 129                     hsDelEmptyCols.html FALSE
#> 130                 hsMatrixToListForm.html FALSE
#> 131                      hsMovingMean-1.png FALSE
#> 132                       hsMovingMean.html FALSE
#> 133              hsOpenWindowsExplorer.html FALSE
#> 134                          hsPrepPdf.html FALSE
#> 135                         hsQuoteChr.html FALSE
#> 136                    hsRenameColumns.html FALSE
#> 137                          hsResolve.html FALSE
#> 138                hsRestoreAttributes.html FALSE
#> 139                         hsSafeName.html FALSE
#> 140                            hsShell.html FALSE
#> 141                          hsShowPdf.html FALSE
#> 142                     hsStringToDate.html FALSE
#> 143                   hsStringToDouble.html FALSE
#> 144                   hsSubstSpecChars.html FALSE
#> 145                           hsSystem.html FALSE
#> 146                             hsTrim.html FALSE
#> 147                       hsValidValue.html FALSE
#> 148                            inRange.html FALSE
#> 149                             indent.html FALSE
#> 150                              index.html FALSE
#> 151                      insertColumns.html FALSE
#> 152                 intToNumeralSystem.html FALSE
#> 153                         is.unnamed.html FALSE
#> 154                            isASCII.html FALSE
#> 155                   isDotOrDoubleDot.html FALSE
#> 156                       isEvenNumber.html FALSE
#> 157                           isLoaded.html FALSE
#> 158                   isNaInAllColumns.html FALSE
#> 159                      isNaInAllRows.html FALSE
#> 160                        isNaOrEmpty.html FALSE
#> 161                      isNullOrEmpty.html FALSE
#> 162                        isOddNumber.html FALSE
#> 163                         isTryError.html FALSE
#> 164                        lastElement.html FALSE
#> 165                               left.html FALSE
#> 166                leftSubstringEquals.html FALSE
#> 167                       limitToRange.html FALSE
#> 168                  linearCombination.html FALSE
#> 169                        listObjects.html FALSE