Order a Data Frame by One or more Columns

Order a Data Frame by One or more Columns

orderBy(df, by = NULL, ...)

Arguments

df
data frame

by
vector of column names specifying the columns by which to order

...
further arguments passed to order, such as decreasing

Value

df being sorted and with newly renumbered rows

Examples

orderBy(iris, by = "Sepal.Length")
#>     Sepal.Length Sepal.Width Petal.Length Petal.Width    Species
#> 1            4.3         3.0          1.1         0.1     setosa
#> 2            4.4         2.9          1.4         0.2     setosa
#> 3            4.4         3.0          1.3         0.2     setosa
#> 4            4.4         3.2          1.3         0.2     setosa
#> 5            4.5         2.3          1.3         0.3     setosa
#> 6            4.6         3.1          1.5         0.2     setosa
#> 7            4.6         3.4          1.4         0.3     setosa
#> 8            4.6         3.6          1.0         0.2     setosa
#> 9            4.6         3.2          1.4         0.2     setosa
#> 10           4.7         3.2          1.3         0.2     setosa
#> 11           4.7         3.2          1.6         0.2     setosa
#> 12           4.8         3.4          1.6         0.2     setosa
#> 13           4.8         3.0          1.4         0.1     setosa
#> 14           4.8         3.4          1.9         0.2     setosa
#> 15           4.8         3.1          1.6         0.2     setosa
#> 16           4.8         3.0          1.4         0.3     setosa
#> 17           4.9         3.0          1.4         0.2     setosa
#> 18           4.9         3.1          1.5         0.1     setosa
#> 19           4.9         3.1          1.5         0.2     setosa
#> 20           4.9         3.6          1.4         0.1     setosa
#> 21           4.9         2.4          3.3         1.0 versicolor
#> 22           4.9         2.5          4.5         1.7  virginica
#> 23           5.0         3.6          1.4         0.2     setosa
#> 24           5.0         3.4          1.5         0.2     setosa
#> 25           5.0         3.0          1.6         0.2     setosa
#> 26           5.0         3.4          1.6         0.4     setosa
#> 27           5.0         3.2          1.2         0.2     setosa
#> 28           5.0         3.5          1.3         0.3     setosa
#> 29           5.0         3.5          1.6         0.6     setosa
#> 30           5.0         3.3          1.4         0.2     setosa
#> 31           5.0         2.0          3.5         1.0 versicolor
#> 32           5.0         2.3          3.3         1.0 versicolor
#> 33           5.1         3.5          1.4         0.2     setosa
#> 34           5.1         3.5          1.4         0.3     setosa
#> 35           5.1         3.8          1.5         0.3     setosa
#> 36           5.1         3.7          1.5         0.4     setosa
#> 37           5.1         3.3          1.7         0.5     setosa
#> 38           5.1         3.4          1.5         0.2     setosa
#> 39           5.1         3.8          1.9         0.4     setosa
#> 40           5.1         3.8          1.6         0.2     setosa
#> 41           5.1         2.5          3.0         1.1 versicolor
#> 42           5.2         3.5          1.5         0.2     setosa
#> 43           5.2         3.4          1.4         0.2     setosa
#> 44           5.2         4.1          1.5         0.1     setosa
#> 45           5.2         2.7          3.9         1.4 versicolor
#> 46           5.3         3.7          1.5         0.2     setosa
#> 47           5.4         3.9          1.7         0.4     setosa
#> 48           5.4         3.7          1.5         0.2     setosa
#> 49           5.4         3.9          1.3         0.4     setosa
#> 50           5.4         3.4          1.7         0.2     setosa
#> 51           5.4         3.4          1.5         0.4     setosa
#> 52           5.4         3.0          4.5         1.5 versicolor
#> 53           5.5         4.2          1.4         0.2     setosa
#> 54           5.5         3.5          1.3         0.2     setosa
#> 55           5.5         2.3          4.0         1.3 versicolor
#> 56           5.5         2.4          3.8         1.1 versicolor
#> 57           5.5         2.4          3.7         1.0 versicolor
#> 58           5.5         2.5          4.0         1.3 versicolor
#> 59           5.5         2.6          4.4         1.2 versicolor
#> 60           5.6         2.9          3.6         1.3 versicolor
#> 61           5.6         3.0          4.5         1.5 versicolor
#> 62           5.6         2.5          3.9         1.1 versicolor
#> 63           5.6         3.0          4.1         1.3 versicolor
#> 64           5.6         2.7          4.2         1.3 versicolor
#> 65           5.6         2.8          4.9         2.0  virginica
#> 66           5.7         4.4          1.5         0.4     setosa
#> 67           5.7         3.8          1.7         0.3     setosa
#> 68           5.7         2.8          4.5         1.3 versicolor
#> 69           5.7         2.6          3.5         1.0 versicolor
#> 70           5.7         3.0          4.2         1.2 versicolor
#> 71           5.7         2.9          4.2         1.3 versicolor
#> 72           5.7         2.8          4.1         1.3 versicolor
#> 73           5.7         2.5          5.0         2.0  virginica
#> 74           5.8         4.0          1.2         0.2     setosa
#> 75           5.8         2.7          4.1         1.0 versicolor
#> 76           5.8         2.7          3.9         1.2 versicolor
#> 77           5.8         2.6          4.0         1.2 versicolor
#> 78           5.8         2.7          5.1         1.9  virginica
#> 79           5.8         2.8          5.1         2.4  virginica
#> 80           5.8         2.7          5.1         1.9  virginica
#> 81           5.9         3.0          4.2         1.5 versicolor
#> 82           5.9         3.2          4.8         1.8 versicolor
#> 83           5.9         3.0          5.1         1.8  virginica
#> 84           6.0         2.2          4.0         1.0 versicolor
#> 85           6.0         2.9          4.5         1.5 versicolor
#> 86           6.0         2.7          5.1         1.6 versicolor
#> 87           6.0         3.4          4.5         1.6 versicolor
#> 88           6.0         2.2          5.0         1.5  virginica
#> 89           6.0         3.0          4.8         1.8  virginica
#> 90           6.1         2.9          4.7         1.4 versicolor
#> 91           6.1         2.8          4.0         1.3 versicolor
#> 92           6.1         2.8          4.7         1.2 versicolor
#> 93           6.1         3.0          4.6         1.4 versicolor
#> 94           6.1         3.0          4.9         1.8  virginica
#> 95           6.1         2.6          5.6         1.4  virginica
#> 96           6.2         2.2          4.5         1.5 versicolor
#> 97           6.2         2.9          4.3         1.3 versicolor
#> 98           6.2         2.8          4.8         1.8  virginica
#> 99           6.2         3.4          5.4         2.3  virginica
#> 100          6.3         3.3          4.7         1.6 versicolor
#> 101          6.3         2.5          4.9         1.5 versicolor
#> 102          6.3         2.3          4.4         1.3 versicolor
#> 103          6.3         3.3          6.0         2.5  virginica
#> 104          6.3         2.9          5.6         1.8  virginica
#> 105          6.3         2.7          4.9         1.8  virginica
#> 106          6.3         2.8          5.1         1.5  virginica
#> 107          6.3         3.4          5.6         2.4  virginica
#> 108          6.3         2.5          5.0         1.9  virginica
#> 109          6.4         3.2          4.5         1.5 versicolor
#> 110          6.4         2.9          4.3         1.3 versicolor
#> 111          6.4         2.7          5.3         1.9  virginica
#> 112          6.4         3.2          5.3         2.3  virginica
#> 113          6.4         2.8          5.6         2.1  virginica
#> 114          6.4         2.8          5.6         2.2  virginica
#> 115          6.4         3.1          5.5         1.8  virginica
#> 116          6.5         2.8          4.6         1.5 versicolor
#> 117          6.5         3.0          5.8         2.2  virginica
#> 118          6.5         3.2          5.1         2.0  virginica
#> 119          6.5         3.0          5.5         1.8  virginica
#> 120          6.5         3.0          5.2         2.0  virginica
#> 121          6.6         2.9          4.6         1.3 versicolor
#> 122          6.6         3.0          4.4         1.4 versicolor
#> 123          6.7         3.1          4.4         1.4 versicolor
#> 124          6.7         3.0          5.0         1.7 versicolor
#> 125          6.7         3.1          4.7         1.5 versicolor
#> 126          6.7         2.5          5.8         1.8  virginica
#> 127          6.7         3.3          5.7         2.1  virginica
#> 128          6.7         3.1          5.6         2.4  virginica
#> 129          6.7         3.3          5.7         2.5  virginica
#> 130          6.7         3.0          5.2         2.3  virginica
#> 131          6.8         2.8          4.8         1.4 versicolor
#> 132          6.8         3.0          5.5         2.1  virginica
#> 133          6.8         3.2          5.9         2.3  virginica
#> 134          6.9         3.1          4.9         1.5 versicolor
#> 135          6.9         3.2          5.7         2.3  virginica
#> 136          6.9         3.1          5.4         2.1  virginica
#> 137          6.9         3.1          5.1         2.3  virginica
#> 138          7.0         3.2          4.7         1.4 versicolor
#> 139          7.1         3.0          5.9         2.1  virginica
#> 140          7.2         3.6          6.1         2.5  virginica
#> 141          7.2         3.2          6.0         1.8  virginica
#> 142          7.2         3.0          5.8         1.6  virginica
#> 143          7.3         2.9          6.3         1.8  virginica
#> 144          7.4         2.8          6.1         1.9  virginica
#> 145          7.6         3.0          6.6         2.1  virginica
#> 146          7.7         3.8          6.7         2.2  virginica
#> 147          7.7         2.6          6.9         2.3  virginica
#> 148          7.7         2.8          6.7         2.0  virginica
#> 149          7.7         3.0          6.1         2.3  virginica
#> 150          7.9         3.8          6.4         2.0  virginica
orderBy(iris, by = "Species", decreasing = TRUE)
#>     Sepal.Length Sepal.Width Petal.Length Petal.Width    Species
#> 1            6.3         3.3          6.0         2.5  virginica
#> 2            5.8         2.7          5.1         1.9  virginica
#> 3            7.1         3.0          5.9         2.1  virginica
#> 4            6.3         2.9          5.6         1.8  virginica
#> 5            6.5         3.0          5.8         2.2  virginica
#> 6            7.6         3.0          6.6         2.1  virginica
#> 7            4.9         2.5          4.5         1.7  virginica
#> 8            7.3         2.9          6.3         1.8  virginica
#> 9            6.7         2.5          5.8         1.8  virginica
#> 10           7.2         3.6          6.1         2.5  virginica
#> 11           6.5         3.2          5.1         2.0  virginica
#> 12           6.4         2.7          5.3         1.9  virginica
#> 13           6.8         3.0          5.5         2.1  virginica
#> 14           5.7         2.5          5.0         2.0  virginica
#> 15           5.8         2.8          5.1         2.4  virginica
#> 16           6.4         3.2          5.3         2.3  virginica
#> 17           6.5         3.0          5.5         1.8  virginica
#> 18           7.7         3.8          6.7         2.2  virginica
#> 19           7.7         2.6          6.9         2.3  virginica
#> 20           6.0         2.2          5.0         1.5  virginica
#> 21           6.9         3.2          5.7         2.3  virginica
#> 22           5.6         2.8          4.9         2.0  virginica
#> 23           7.7         2.8          6.7         2.0  virginica
#> 24           6.3         2.7          4.9         1.8  virginica
#> 25           6.7         3.3          5.7         2.1  virginica
#> 26           7.2         3.2          6.0         1.8  virginica
#> 27           6.2         2.8          4.8         1.8  virginica
#> 28           6.1         3.0          4.9         1.8  virginica
#> 29           6.4         2.8          5.6         2.1  virginica
#> 30           7.2         3.0          5.8         1.6  virginica
#> 31           7.4         2.8          6.1         1.9  virginica
#> 32           7.9         3.8          6.4         2.0  virginica
#> 33           6.4         2.8          5.6         2.2  virginica
#> 34           6.3         2.8          5.1         1.5  virginica
#> 35           6.1         2.6          5.6         1.4  virginica
#> 36           7.7         3.0          6.1         2.3  virginica
#> 37           6.3         3.4          5.6         2.4  virginica
#> 38           6.4         3.1          5.5         1.8  virginica
#> 39           6.0         3.0          4.8         1.8  virginica
#> 40           6.9         3.1          5.4         2.1  virginica
#> 41           6.7         3.1          5.6         2.4  virginica
#> 42           6.9         3.1          5.1         2.3  virginica
#> 43           5.8         2.7          5.1         1.9  virginica
#> 44           6.8         3.2          5.9         2.3  virginica
#> 45           6.7         3.3          5.7         2.5  virginica
#> 46           6.7         3.0          5.2         2.3  virginica
#> 47           6.3         2.5          5.0         1.9  virginica
#> 48           6.5         3.0          5.2         2.0  virginica
#> 49           6.2         3.4          5.4         2.3  virginica
#> 50           5.9         3.0          5.1         1.8  virginica
#> 51           7.0         3.2          4.7         1.4 versicolor
#> 52           6.4         3.2          4.5         1.5 versicolor
#> 53           6.9         3.1          4.9         1.5 versicolor
#> 54           5.5         2.3          4.0         1.3 versicolor
#> 55           6.5         2.8          4.6         1.5 versicolor
#> 56           5.7         2.8          4.5         1.3 versicolor
#> 57           6.3         3.3          4.7         1.6 versicolor
#> 58           4.9         2.4          3.3         1.0 versicolor
#> 59           6.6         2.9          4.6         1.3 versicolor
#> 60           5.2         2.7          3.9         1.4 versicolor
#> 61           5.0         2.0          3.5         1.0 versicolor
#> 62           5.9         3.0          4.2         1.5 versicolor
#> 63           6.0         2.2          4.0         1.0 versicolor
#> 64           6.1         2.9          4.7         1.4 versicolor
#> 65           5.6         2.9          3.6         1.3 versicolor
#> 66           6.7         3.1          4.4         1.4 versicolor
#> 67           5.6         3.0          4.5         1.5 versicolor
#> 68           5.8         2.7          4.1         1.0 versicolor
#> 69           6.2         2.2          4.5         1.5 versicolor
#> 70           5.6         2.5          3.9         1.1 versicolor
#> 71           5.9         3.2          4.8         1.8 versicolor
#> 72           6.1         2.8          4.0         1.3 versicolor
#> 73           6.3         2.5          4.9         1.5 versicolor
#> 74           6.1         2.8          4.7         1.2 versicolor
#> 75           6.4         2.9          4.3         1.3 versicolor
#> 76           6.6         3.0          4.4         1.4 versicolor
#> 77           6.8         2.8          4.8         1.4 versicolor
#> 78           6.7         3.0          5.0         1.7 versicolor
#> 79           6.0         2.9          4.5         1.5 versicolor
#> 80           5.7         2.6          3.5         1.0 versicolor
#> 81           5.5         2.4          3.8         1.1 versicolor
#> 82           5.5         2.4          3.7         1.0 versicolor
#> 83           5.8         2.7          3.9         1.2 versicolor
#> 84           6.0         2.7          5.1         1.6 versicolor
#> 85           5.4         3.0          4.5         1.5 versicolor
#> 86           6.0         3.4          4.5         1.6 versicolor
#> 87           6.7         3.1          4.7         1.5 versicolor
#> 88           6.3         2.3          4.4         1.3 versicolor
#> 89           5.6         3.0          4.1         1.3 versicolor
#> 90           5.5         2.5          4.0         1.3 versicolor
#> 91           5.5         2.6          4.4         1.2 versicolor
#> 92           6.1         3.0          4.6         1.4 versicolor
#> 93           5.8         2.6          4.0         1.2 versicolor
#> 94           5.0         2.3          3.3         1.0 versicolor
#> 95           5.6         2.7          4.2         1.3 versicolor
#> 96           5.7         3.0          4.2         1.2 versicolor
#> 97           5.7         2.9          4.2         1.3 versicolor
#> 98           6.2         2.9          4.3         1.3 versicolor
#> 99           5.1         2.5          3.0         1.1 versicolor
#> 100          5.7         2.8          4.1         1.3 versicolor
#> 101          5.1         3.5          1.4         0.2     setosa
#> 102          4.9         3.0          1.4         0.2     setosa
#> 103          4.7         3.2          1.3         0.2     setosa
#> 104          4.6         3.1          1.5         0.2     setosa
#> 105          5.0         3.6          1.4         0.2     setosa
#> 106          5.4         3.9          1.7         0.4     setosa
#> 107          4.6         3.4          1.4         0.3     setosa
#> 108          5.0         3.4          1.5         0.2     setosa
#> 109          4.4         2.9          1.4         0.2     setosa
#> 110          4.9         3.1          1.5         0.1     setosa
#> 111          5.4         3.7          1.5         0.2     setosa
#> 112          4.8         3.4          1.6         0.2     setosa
#> 113          4.8         3.0          1.4         0.1     setosa
#> 114          4.3         3.0          1.1         0.1     setosa
#> 115          5.8         4.0          1.2         0.2     setosa
#> 116          5.7         4.4          1.5         0.4     setosa
#> 117          5.4         3.9          1.3         0.4     setosa
#> 118          5.1         3.5          1.4         0.3     setosa
#> 119          5.7         3.8          1.7         0.3     setosa
#> 120          5.1         3.8          1.5         0.3     setosa
#> 121          5.4         3.4          1.7         0.2     setosa
#> 122          5.1         3.7          1.5         0.4     setosa
#> 123          4.6         3.6          1.0         0.2     setosa
#> 124          5.1         3.3          1.7         0.5     setosa
#> 125          4.8         3.4          1.9         0.2     setosa
#> 126          5.0         3.0          1.6         0.2     setosa
#> 127          5.0         3.4          1.6         0.4     setosa
#> 128          5.2         3.5          1.5         0.2     setosa
#> 129          5.2         3.4          1.4         0.2     setosa
#> 130          4.7         3.2          1.6         0.2     setosa
#> 131          4.8         3.1          1.6         0.2     setosa
#> 132          5.4         3.4          1.5         0.4     setosa
#> 133          5.2         4.1          1.5         0.1     setosa
#> 134          5.5         4.2          1.4         0.2     setosa
#> 135          4.9         3.1          1.5         0.2     setosa
#> 136          5.0         3.2          1.2         0.2     setosa
#> 137          5.5         3.5          1.3         0.2     setosa
#> 138          4.9         3.6          1.4         0.1     setosa
#> 139          4.4         3.0          1.3         0.2     setosa
#> 140          5.1         3.4          1.5         0.2     setosa
#> 141          5.0         3.5          1.3         0.3     setosa
#> 142          4.5         2.3          1.3         0.3     setosa
#> 143          4.4         3.2          1.3         0.2     setosa
#> 144          5.0         3.5          1.6         0.6     setosa
#> 145          5.1         3.8          1.9         0.4     setosa
#> 146          4.8         3.0          1.4         0.3     setosa
#> 147          5.1         3.8          1.6         0.2     setosa
#> 148          4.6         3.2          1.4         0.2     setosa
#> 149          5.3         3.7          1.5         0.2     setosa
#> 150          5.0         3.3          1.4         0.2     setosa
orderBy(
  iris, 
  by = c("Species", "Petal.Width", "Petal.Length"), 
  decreasing = TRUE
)
#>     Sepal.Length Sepal.Width Petal.Length Petal.Width    Species
#> 1            7.2         3.6          6.1         2.5  virginica
#> 2            6.3         3.3          6.0         2.5  virginica
#> 3            6.7         3.3          5.7         2.5  virginica
#> 4            6.3         3.4          5.6         2.4  virginica
#> 5            6.7         3.1          5.6         2.4  virginica
#> 6            5.8         2.8          5.1         2.4  virginica
#> 7            7.7         2.6          6.9         2.3  virginica
#> 8            7.7         3.0          6.1         2.3  virginica
#> 9            6.8         3.2          5.9         2.3  virginica
#> 10           6.9         3.2          5.7         2.3  virginica
#> 11           6.2         3.4          5.4         2.3  virginica
#> 12           6.4         3.2          5.3         2.3  virginica
#> 13           6.7         3.0          5.2         2.3  virginica
#> 14           6.9         3.1          5.1         2.3  virginica
#> 15           7.7         3.8          6.7         2.2  virginica
#> 16           6.5         3.0          5.8         2.2  virginica
#> 17           6.4         2.8          5.6         2.2  virginica
#> 18           7.6         3.0          6.6         2.1  virginica
#> 19           7.1         3.0          5.9         2.1  virginica
#> 20           6.7         3.3          5.7         2.1  virginica
#> 21           6.4         2.8          5.6         2.1  virginica
#> 22           6.8         3.0          5.5         2.1  virginica
#> 23           6.9         3.1          5.4         2.1  virginica
#> 24           7.7         2.8          6.7         2.0  virginica
#> 25           7.9         3.8          6.4         2.0  virginica
#> 26           6.5         3.0          5.2         2.0  virginica
#> 27           6.5         3.2          5.1         2.0  virginica
#> 28           5.7         2.5          5.0         2.0  virginica
#> 29           5.6         2.8          4.9         2.0  virginica
#> 30           7.4         2.8          6.1         1.9  virginica
#> 31           6.4         2.7          5.3         1.9  virginica
#> 32           5.8         2.7          5.1         1.9  virginica
#> 33           5.8         2.7          5.1         1.9  virginica
#> 34           6.3         2.5          5.0         1.9  virginica
#> 35           7.3         2.9          6.3         1.8  virginica
#> 36           7.2         3.2          6.0         1.8  virginica
#> 37           6.7         2.5          5.8         1.8  virginica
#> 38           6.3         2.9          5.6         1.8  virginica
#> 39           6.5         3.0          5.5         1.8  virginica
#> 40           6.4         3.1          5.5         1.8  virginica
#> 41           5.9         3.0          5.1         1.8  virginica
#> 42           6.3         2.7          4.9         1.8  virginica
#> 43           6.1         3.0          4.9         1.8  virginica
#> 44           6.2         2.8          4.8         1.8  virginica
#> 45           6.0         3.0          4.8         1.8  virginica
#> 46           4.9         2.5          4.5         1.7  virginica
#> 47           7.2         3.0          5.8         1.6  virginica
#> 48           6.3         2.8          5.1         1.5  virginica
#> 49           6.0         2.2          5.0         1.5  virginica
#> 50           6.1         2.6          5.6         1.4  virginica
#> 51           5.9         3.2          4.8         1.8 versicolor
#> 52           6.7         3.0          5.0         1.7 versicolor
#> 53           6.0         2.7          5.1         1.6 versicolor
#> 54           6.3         3.3          4.7         1.6 versicolor
#> 55           6.0         3.4          4.5         1.6 versicolor
#> 56           6.9         3.1          4.9         1.5 versicolor
#> 57           6.3         2.5          4.9         1.5 versicolor
#> 58           6.7         3.1          4.7         1.5 versicolor
#> 59           6.5         2.8          4.6         1.5 versicolor
#> 60           6.4         3.2          4.5         1.5 versicolor
#> 61           5.6         3.0          4.5         1.5 versicolor
#> 62           6.2         2.2          4.5         1.5 versicolor
#> 63           6.0         2.9          4.5         1.5 versicolor
#> 64           5.4         3.0          4.5         1.5 versicolor
#> 65           5.9         3.0          4.2         1.5 versicolor
#> 66           6.8         2.8          4.8         1.4 versicolor
#> 67           7.0         3.2          4.7         1.4 versicolor
#> 68           6.1         2.9          4.7         1.4 versicolor
#> 69           6.1         3.0          4.6         1.4 versicolor
#> 70           6.7         3.1          4.4         1.4 versicolor
#> 71           6.6         3.0          4.4         1.4 versicolor
#> 72           5.2         2.7          3.9         1.4 versicolor
#> 73           6.6         2.9          4.6         1.3 versicolor
#> 74           5.7         2.8          4.5         1.3 versicolor
#> 75           6.3         2.3          4.4         1.3 versicolor
#> 76           6.4         2.9          4.3         1.3 versicolor
#> 77           6.2         2.9          4.3         1.3 versicolor
#> 78           5.6         2.7          4.2         1.3 versicolor
#> 79           5.7         2.9          4.2         1.3 versicolor
#> 80           5.6         3.0          4.1         1.3 versicolor
#> 81           5.7         2.8          4.1         1.3 versicolor
#> 82           5.5         2.3          4.0         1.3 versicolor
#> 83           6.1         2.8          4.0         1.3 versicolor
#> 84           5.5         2.5          4.0         1.3 versicolor
#> 85           5.6         2.9          3.6         1.3 versicolor
#> 86           6.1         2.8          4.7         1.2 versicolor
#> 87           5.5         2.6          4.4         1.2 versicolor
#> 88           5.7         3.0          4.2         1.2 versicolor
#> 89           5.8         2.6          4.0         1.2 versicolor
#> 90           5.8         2.7          3.9         1.2 versicolor
#> 91           5.6         2.5          3.9         1.1 versicolor
#> 92           5.5         2.4          3.8         1.1 versicolor
#> 93           5.1         2.5          3.0         1.1 versicolor
#> 94           5.8         2.7          4.1         1.0 versicolor
#> 95           6.0         2.2          4.0         1.0 versicolor
#> 96           5.5         2.4          3.7         1.0 versicolor
#> 97           5.0         2.0          3.5         1.0 versicolor
#> 98           5.7         2.6          3.5         1.0 versicolor
#> 99           4.9         2.4          3.3         1.0 versicolor
#> 100          5.0         2.3          3.3         1.0 versicolor
#> 101          5.0         3.5          1.6         0.6     setosa
#> 102          5.1         3.3          1.7         0.5     setosa
#> 103          5.1         3.8          1.9         0.4     setosa
#> 104          5.4         3.9          1.7         0.4     setosa
#> 105          5.0         3.4          1.6         0.4     setosa
#> 106          5.7         4.4          1.5         0.4     setosa
#> 107          5.1         3.7          1.5         0.4     setosa
#> 108          5.4         3.4          1.5         0.4     setosa
#> 109          5.4         3.9          1.3         0.4     setosa
#> 110          5.7         3.8          1.7         0.3     setosa
#> 111          5.1         3.8          1.5         0.3     setosa
#> 112          4.6         3.4          1.4         0.3     setosa
#> 113          5.1         3.5          1.4         0.3     setosa
#> 114          4.8         3.0          1.4         0.3     setosa
#> 115          5.0         3.5          1.3         0.3     setosa
#> 116          4.5         2.3          1.3         0.3     setosa
#> 117          4.8         3.4          1.9         0.2     setosa
#> 118          5.4         3.4          1.7         0.2     setosa
#> 119          4.8         3.4          1.6         0.2     setosa
#> 120          5.0         3.0          1.6         0.2     setosa
#> 121          4.7         3.2          1.6         0.2     setosa
#> 122          4.8         3.1          1.6         0.2     setosa
#> 123          5.1         3.8          1.6         0.2     setosa
#> 124          4.6         3.1          1.5         0.2     setosa
#> 125          5.0         3.4          1.5         0.2     setosa
#> 126          5.4         3.7          1.5         0.2     setosa
#> 127          5.2         3.5          1.5         0.2     setosa
#> 128          4.9         3.1          1.5         0.2     setosa
#> 129          5.1         3.4          1.5         0.2     setosa
#> 130          5.3         3.7          1.5         0.2     setosa
#> 131          5.1         3.5          1.4         0.2     setosa
#> 132          4.9         3.0          1.4         0.2     setosa
#> 133          5.0         3.6          1.4         0.2     setosa
#> 134          4.4         2.9          1.4         0.2     setosa
#> 135          5.2         3.4          1.4         0.2     setosa
#> 136          5.5         4.2          1.4         0.2     setosa
#> 137          4.6         3.2          1.4         0.2     setosa
#> 138          5.0         3.3          1.4         0.2     setosa
#> 139          4.7         3.2          1.3         0.2     setosa
#> 140          5.5         3.5          1.3         0.2     setosa
#> 141          4.4         3.0          1.3         0.2     setosa
#> 142          4.4         3.2          1.3         0.2     setosa
#> 143          5.8         4.0          1.2         0.2     setosa
#> 144          5.0         3.2          1.2         0.2     setosa
#> 145          4.6         3.6          1.0         0.2     setosa
#> 146          4.9         3.1          1.5         0.1     setosa
#> 147          5.2         4.1          1.5         0.1     setosa
#> 148          4.8         3.0          1.4         0.1     setosa
#> 149          4.9         3.6          1.4         0.1     setosa
#> 150          4.3         3.0          1.1         0.1     setosa