Create a Matrix with Random Integer Values

Create a matrix of given dimension and fill it with random integer values

randomMatrix(dim = c(sample(10, 1), sample(10, 1)), values = seq_len(100))

Arguments

dim
integer vector of length two containing the number of rows and columns, respectively, that the output matrix shall contain

values
set of values to be used within the matrix

Examples


# By default, the matrix has a random number of rows between 1 and 10 and
# a random number of columns between 1 and 10 and random values of 1:100
randomMatrix()

#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7]
#> [1,]    6   52   25    3   82   42   67
#> [2,]  100   85   55   47   41   67   52
#> [3,]   15   72   84   81   22   51   84


# You may specify the dimensions (here: 5 rows, 3 columns)...
randomMatrix(dim = c(5, 3))

#>      [,1] [,2] [,3]
#> [1,]   71   51   86
#> [2,]   88   36   57
#> [3,]    2   78   26
#> [4,]   89   59   92
#> [5,]  100   64   87


# ... and the set of values to be used within the matrix
randomMatrix(dim = c(5, 3), values = c(0, 0.5, 1, NA))

#>      [,1] [,2] [,3]
#> [1,]  0.0    0    1
#> [2,]  0.5    1    1
#> [3,]   NA   NA    0
#> [4,]  0.0   NA    0
#> [5,]  0.0    1   NA